### 2025-10-24 ccdd86f
Update 11_assisted_plotting.py
