### 2025-10-24 ea9c036
Update plotbot.py
