+++ paper/paper.bib
@@ -91,32 +91,31 @@
 
 @software{spacy2020,
   author       = {Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane},
-  title        = {spaCy: Industrial-strength Natural Language Processing in Python},
+  title        = {{spaCy}: Industrial-strength Natural Language Processing in {P}ython},
   year         = {2020},
+  publisher    = {Explosion},
   url          = {https://spacy.io},
-  note         = {Version 3.x}
+  version      = {3.x}
 }
 
 @software{streamlit2023,
-  author       = {{Streamlit Inc.}},
-  title        = {Streamlit: Rapid development framework for data apps},
+  author       = {{Snowflake Inc.}},
+  title        = {{Streamlit}},
   year         = {2023},
-  url          = {https://streamlit.io},
-  note         = {Version 1.x}
+  url          = {https://streamlit.io}
 }
 
-@misc{polars2023,
+@software{polars2023,
   author       = {Vink, Ritchie},
-  title        = {Polars: A fast DataFrame library},
+  title        = {{Polars}},
   year         = {2023},
   url          = {https://pola.rs},
-  note         = {Version 1.x}
+  doi          = {10.5281/zenodo.7697217}
 }
 
-@software{voyant2016,
-  author       = {Sinclair, Geoffrey and Rockwell, Geoffrey},
-  title        = {Voyant Tools: A web-based tool for text analysis},
+@article{voyant2016,
+  author       = {Sinclair, Stéfan and Rockwell, Geoffrey},
+  title        = {Voyant Tools},
   year         = {2016},
-  url          = {https://voyant-tools.org},
-  note         = {Version 1.x}
+  url          = {https://voyant-tools.org}
 }
