### 2026-02-02 9b972bc
Update paper.md
