### 2026-03-30 #9274 @logological
@browndw I'm looking forward to coordinating the review of your manuscript, which I expect to be able to start in the next week or so.  In the meantime, I would be grateful if you could help identify potential reviewers.  If you have any suggestions, please mention them in a comment here (without tagging them with an `@`). You might want to search [our directory of volunteers](https://reviewers.joss.theoj.org/lookup) to see if any seem suitable for your submission.
