### 2026-03-30 #9274 @browndw
@logological Thanks so much! Looking forward to it, too. One potential reviewer would be kbenoit. Most of his text processing programming is in R (and C++) rather than Python. However, much of the logic is similar, and I've certainly been inspired by his work. I'll comb through the directory more in the next couple of days. Thanks again.
