### 2026-04-15 #10418 @editorialbot
[REVIEW]: DocuScope Corpus Analysis & Concordancer: A Streamlit Application for Rhetorical and Linguistic Text Analysis
**Submitting author:** @browndw (<a href="http://orcid.org/0000-0001-7745-6354">David Brown</a>)
**Repository:** https://github.com/browndw/docuscope-ca-online
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.0
