### 2026-04-29 browndw/docuscope-ca-online#3 @PetrKorab
Provide details on backend: https://docuscope-ca.eberly.cmu.edu
Most users of this project will likely use the [web app](https://docuscope-ca.eberly.cmu.edu/)., as it saves time for lecturers.

To make the project more practical for teaching, the backend of the web app should be documented and included in the project documentation, as well as in the README of the GitHub repository. Please add a section to the documentation that explains:

- How many concurrent users the app can support
- The limits on data usage and traffic per user
- What measures are in place to prevent overload or shutdown under high traffic

This information should also be mentioned in the README of the project [repository](https://github.com/browndw/docuscope-ca-online).
