### 2026-04-29 #10418 @PetrKorab
I find this project very useful. But many users will just use the web app, especially those without a deeper technical background. I created an issue to address the comments on the backend of the app. https://github.com/browndw/docuscope-ca-online/issues/3.
