+++ load_tests/scenarios/keyness-internal-single.yml
@@ -1,0 +1,22 @@
+config:
+  target: "http://127.0.0.1:8501"
+  phases:
+    - duration: 5
+      arrivalRate: 1
+      maxVusers: 1
+      name: single_user_validation
+  engines:
+    playwright: {}
+  processor: "../processors/streamlit-metrics.mjs"
+  defaults:
+    headers:
+      Cache-Control: no-cache
+  plugins:
+    metrics-by-endpoint: {}
+
+scenarios:
+  - name: internal-keyness-single
+    engine: playwright
+    beforeScenario: beforeScenario
+    afterScenario: afterScenario
+    testFunction: internalKeynessScenario
