### 2026-05-02 #10418 @browndw
@PetrKorab Thanks so much. I added load testing with Arsenal/Playwright and reported the results in the `REAMDE.md` along with some other information about the architecture. Just let me know if anything is unclear or isn't what you had in mind.
