### 2026-05-02 browndw/docuscope-ca-online#3 @browndw
@PetrKorab Thanks for this suggestion! Added a `load_tests` directory with Arsenal/Playwright load testing scenarios. The results with other information about the architecture has been added to the `README.md`. Just let me know if you think any additional testing would be useful or if anything is unclear.
