### 2026-06-10 03ce6b8
Revised text for clarity and detail on tools
