### 2026-06-30 0807945
Added option for tabular data
