+++ tests/unit/processing/test_corpus_processing.py
@@ -5,6 +5,7 @@ Tests corpus processing workflows, memory management, and data handling.
 """
 
 import polars as pl
+import io
 import sys
 from unittest.mock import patch, MagicMock
 
@@ -19,9 +20,141 @@ except ImportError:
 from webapp.utilities.processing.corpus_processing import (
     finalize_corpus_load,
     finalize_corpus_load_optimized,
+    handle_uploaded_tabular,
     process_new,
 )
 from webapp.utilities.state import SessionKeys
+
+
+class MockUploadedFile:
+    """Minimal uploaded file stand-in for Streamlit upload handlers."""
+
+    def __init__(self, name: str, data: bytes):
+        self.name = name
+        self._data = data
+
+    def getvalue(self):
+        return self._data
+
+
+def tabular_upload(df: pl.DataFrame, name: str) -> MockUploadedFile:
+    """Build a mock uploaded file from a Polars DataFrame."""
+    extension = name.rsplit(".", 1)[-1]
+    if extension == "parquet":
+        buffer = io.BytesIO()
+        df.write_parquet(buffer)
+        data = buffer.getvalue()
+    elif extension == "tsv":
+        data = df.write_csv(separator="\t").encode("utf-8")
+    else:
+        data = df.write_csv().encode("utf-8")
+    return MockUploadedFile(name, data)
+
+
+class TestHandleUploadedTabular:
+    """Test tabular raw-corpus upload handling."""
+
+    def setup_method(self):
+        self.df = pl.DataFrame({
+            "doc_id": ["doc 2", "doc1"],
+            "text": [" Second document ", "First document"],
+            "group": ["B", "A"]
+        })
+
+    @patch('streamlit.success')
+    def test_handle_uploaded_tabular_parquet(self, mock_success):
+        uploaded = tabular_upload(self.df, "corpus.parquet")
+
+        result_df, ready, exceptions = handle_uploaded_tabular(
+            uploaded, check_size=True, max_size=10_000
+        )
+
+        assert ready is True
+        assert exceptions == []
+        assert result_df.columns == ["doc_id", "text"]
+        assert result_df.get_column("doc_id").to_list() == ["doc1", "doc2"]
+        assert result_df.get_column("text").to_list() == [
+            "First document", "Second document"
+        ]
+        mock_success.assert_called_once()
+
+    @patch('streamlit.success')
+    def test_handle_uploaded_tabular_csv(self, mock_success):
+        uploaded = tabular_upload(self.df, "corpus.csv")
+
+        result_df, ready, exceptions = handle_uploaded_tabular(
+            uploaded, check_size=False, max_size=0
+        )
+
+        assert ready is True
+        assert exceptions == []
+        assert result_df.height == 2
+        mock_success.assert_called_once()
+
+    @patch('streamlit.success')
+    def test_handle_uploaded_tabular_tsv(self, mock_success):
+        uploaded = tabular_upload(self.df, "corpus.tsv")
+
+        result_df, ready, exceptions = handle_uploaded_tabular(
+            uploaded, check_size=False, max_size=0
+        )
+
+        assert ready is True
+        assert exceptions == []
+        assert result_df.height == 2
+        mock_success.assert_called_once()
+
+    @patch('streamlit.error')
+    def test_handle_uploaded_tabular_requires_doc_id_and_text(self, mock_error):
+        uploaded = tabular_upload(
+            pl.DataFrame({"doc_id": ["doc1"], "body": ["Text"]}),
+            "corpus.csv"
+        )
+
+        result_df, ready, exceptions = handle_uploaded_tabular(
+            uploaded, check_size=False, max_size=0
+        )
+
+        assert result_df is None
+        assert ready is False
+        assert exceptions == []
[…]
