+++ webapp/utilities/processing/corpus_processing.py
@@ -6,6 +6,7 @@ newly uploaded text, and corpus finalization.
 """
 
 import os
+import io
 import unidecode
 import polars as pl
 import streamlit as st
@@ -16,7 +17,7 @@ from webapp.utilities.processing.corpus_loading import load_corpus_internal, loa
 from webapp.utilities.session import (
     init_metadata_target, init_metadata_reference
 )
-from webapp.utilities.analysis.data_validation import check_corpus_external, check_corpus_new  # noqa: E501
+from webapp.utilities.analysis.data_validation import check_language, check_corpus_external, check_corpus_new  # noqa: E501
 from webapp.utilities.state import LoadCorpusKeys, SessionKeys
 from webapp.utilities.corpus import get_corpus_manager
 from webapp.utilities.core import app_core
@@ -561,6 +562,201 @@ def handle_uploaded_text(
         df, exceptions = None, []
 
     return df, ready, exceptions
+
+
+def handle_uploaded_tabular(
+        uploaded_file,
+        check_size: bool,
+        max_size: int,
+        check_language_flag=False,
+        check_ref=False,
+        target_docs=None
+        ) -> tuple[pl.DataFrame | None, bool, list]:
+    """
+    Handle a single tabular corpus upload with doc_id and text columns.
+
+    Parameters
+    ----------
+    uploaded_file : UploadedFile or None
+        Uploaded Parquet, CSV, or TSV file containing corpus rows.
+    check_size : bool
+        Whether to check corpus size.
+    max_size : int
+        Maximum allowed corpus size.
+    check_language_flag : bool, optional
+        Whether to check language of documents.
+    check_ref : bool, optional
+        Whether to check for reference documents.
+    target_docs : list, optional
+        Target documents for duplicate checking.
+
+    Returns
+    -------
+    tuple[pl.DataFrame | None, bool, list]
+        Tuple of (dataframe, ready_to_process, exceptions).
+    """
+    if uploaded_file is None:
+        return None, False, []
+
+    file_name = uploaded_file.name
+    file_ext = os.path.splitext(file_name)[1].lower().lstrip(".")
+    file_bytes = uploaded_file.getvalue()
+
+    if file_ext not in {"parquet", "csv", "tsv"}:
+        st.error(
+            "Tabular corpora must be uploaded as Parquet, CSV, or TSV files.",
+            icon=":material/block:"
+        )
+        return None, False, []
+
+    try:
+        file_buffer = io.BytesIO(file_bytes)
+        if file_ext == "parquet":
+            df = pl.read_parquet(file_buffer)
+        elif file_ext == "tsv":
+            df = pl.read_csv(file_buffer, separator="\t")
+        else:
+            df = pl.read_csv(file_buffer)
+    except Exception as e:
+        st.error(f"Error processing tabular corpus file: {e}")
+        return None, False, []
+
+    required_cols = ["doc_id", "text"]
+    missing_cols = [col for col in required_cols if col not in df.columns]
+    if missing_cols:
+        st.error(
+            f"Your table must include these columns: {', '.join(missing_cols)}.",
+            icon=":material/block:"
+        )
+        return None, False, []
+
+    if df.is_empty():
+        st.error(
+            "Your tabular corpus does not contain any rows.",
+            icon=":material/block:"
+        )
+        return None, False, []
+
+    df = df.select(required_cols).with_columns(
+        pl.col("doc_id").cast(pl.String),
+        pl.col("text").cast(pl.String)
+    )
+
+    null_doc_ids = df.select(pl.col("doc_id").is_null().sum()).item()
+    null_texts = df.select(pl.col("text").is_null().sum()).item()
+    if null_doc_ids > 0 or null_texts > 0:
+        st.error(
[…]
