### 2026-06-30 afe8332
Update pyproject.toml
