+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "docuscope-ca-online"
-version = "0.4.1"
+version = "0.4.2"
 description = "An online version of the Docuscope CA tool for analyzing text."
 readme = "README.md"
 requires-python = "==3.11.*"
