### 2026-07-12 e6683ec
Create token-frequencies-classroom.yml
