### 2026-07-12 2c6e5e3
Create token-frequencies-max30.yml
