### 2026-07-12 39922e2
Create internal-target-ready.yml
