### 2026-07-12 d8461a3
Update test_headless_processing.py
