+++ tests/unit/ai/test_interpretation.py
@@ -1,0 +1,187 @@
+"""Tests for evidence-grounded interpretation assistant utilities."""
+
+import pandas as pd
+import polars as pl
+
+from webapp.utilities.ai.interpretation import (
+    build_interpretation_evidence,
+    build_interpretation_prompt,
+    get_teaching_lens_label,
+    get_teaching_lens_options,
+    run_interpretation_service,
+)
+
+
+class FakeInterpretationProvider:
+    """Capture request params and return deterministic interpretation text."""
+
+    def __init__(self, response="The table suggests a cautious pattern."):
+        self.response = response
+        self.request_params = None
+
+    def generate_text(self, api_key, request_params, cache_key=None):
+        self.request_params = request_params
+        return self.response
+
+
+def test_build_interpretation_evidence_summarizes_selected_table():
+    """Evidence should include bounded preview, column names, and summaries."""
+    df = pd.DataFrame({
+        "Tag": ["Academic", "Academic", "Narrative", "Narrative", "Narrative"],
+        "RF": [12.5, 9.5, 3.0, 2.5, 2.0],
+        "Category": ["A", "A", "B", "B", "B"],
+    })
+
+    evidence = build_interpretation_evidence(df, table_name="Tag Frequencies")
+
+    assert evidence.table_name == "Tag Frequencies"
+    assert evidence.row_count == 5
+    assert evidence.columns == ["Tag", "RF", "Category"]
+    assert evidence.preview_rows[0]["Tag"] == "Academic"
+    assert evidence.numeric_summaries["RF"]["count"] == 5
+    assert evidence.numeric_summaries["RF"]["max"] == 12.5
+    assert evidence.categorical_top_values["Tag"][0] == {
+        "value": "Narrative",
+        "count": 3,
+    }
+    assert any("RF is relative frequency" in rule for rule in evidence.measurement_rules)
+    assert any("DocuScope tags" in rule for rule in evidence.rhetorical_rules)
+    assert evidence.deterministic_note is not None
+    assert "Coaching guardrail" in evidence.deterministic_note
+    assert "Use RF, not AF" in evidence.deterministic_note
+    assert "rhetorical patterns" in evidence.deterministic_note
+
+
+def test_frequency_table_rules_warn_against_af_first_comparison():
+    """AF/RF tables should teach normalized comparison and rhetorical caution."""
+    evidence = build_interpretation_evidence(
+        pd.DataFrame({
+            "Tag": ["AcademicTerms", "InformationExposition"],
+            "AF": [250, 200],
+            "RF": [125.0, 100.0],
+        }),
+        table_name="Tag Frequencies",
+    )
+    prompt = build_interpretation_prompt(
+        "Which tags are most common?",
+        evidence,
+        teaching_lens="find_pattern",
+    )
+
+    assert "AF is absolute frequency" in prompt
+    assert "RF is relative frequency" in prompt
+    assert "do not recommend AF as the first comparison basis" in prompt
+    assert "High frequency means frequent" in prompt
+    assert "communicative patterns" in prompt
+    assert "model accuracy" in prompt
+    assert "Table-specific coaching observations" in prompt
+    assert "row-level RF rankings" in prompt
+    assert "do not foreground the mean of AF or RF" in prompt
+
+
+def test_docuscope_untagged_rows_are_filtered_from_evidence():
+    """Interpretation evidence should match DocuScope tools by excluding Untagged."""
+    evidence = build_interpretation_evidence(
+        pd.DataFrame({
+            "Tag": ["Untagged", "AcademicTerms", "InformationExposition"],
+            "AF": [125542, 34830, 22100],
+            "RF": [9.8, 2.7, 1.9],
+        }),
+        table_name="Tag Frequencies",
+    )
[…]
