### 2026-07-12 90444e7
Added scripts for model setup
