### 2026-07-23 fcad651
Update spaCy reference to Zenodo version
