+++ paper/paper.bib
@@ -89,13 +89,12 @@
   doi = {10.1109/IPCC.2005.1494244},
 }
 
-@software{spacy2020,
-  author       = {Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane},
-  title        = {{spaCy}: Industrial-strength Natural Language Processing in {P}ython},
-  year         = {2020},
-  publisher    = {Explosion},
+@software{spacy2023,
+  author       = {Montani, Ines and Honnibal, Matthew and Boyd, Adriane and Van Landeghem, Sofie and Peters, Henning},
+  title        = {{spaCy}},
+  year         = {2023},
   url          = {https://spacy.io},
-  version      = {3.x}
+  doi          = {10.5281/zenodo.1212303},
 }
 
 @software{streamlit2023,
