+++ webapp/persistence/registry.py
@@ -1,0 +1,980 @@
+"""Artifact registry services for shared and private analytical artifacts."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from datetime import datetime, timezone
+from functools import lru_cache
+import hashlib
+import json
+from pathlib import Path
+from typing import Any
+
+import polars as pl
+from sqlalchemy import select
+from sqlalchemy.exc import IntegrityError
+
+from webapp.persistence.database import create_session_factory
+from webapp.persistence.models import ArtifactJob, ArtifactRecord
+
+
+ARTIFACT_STORE_ROOT = Path("webapp/_artifacts")
+KEYNESS_ARTIFACT_TYPE = "keyness_bundle"
+KEYNESS_PARTS_ARTIFACT_TYPE = "keyness_parts_bundle"
+FREQUENCY_ARTIFACT_TYPE = "frequency_bundle"
+COLLOCATION_ARTIFACT_TYPE = "collocation_bundle"
+NGRAM_ARTIFACT_TYPE = "ngram_bundle"
+JSON_ARTIFACT_FILENAME = "payload.json"
+
+
+@dataclass(frozen=True)
+class ArtifactIdentity:
+    """Normalized identity for an artifact registry entry."""
+
+    artifact_type: str
+    scope: str
+    selector_hash: str
+    selector_payload: dict[str, Any]
+    parameter_hash: str
+    parameter_payload: dict[str, Any]
+    pipeline_version: str
+    model_version: str
+    owner_principal_id: str | None = None
+
+
+@dataclass(frozen=True)
+class ReservationResult:
+    """Result of attempting to reserve an artifact identity for computation."""
+
+    state: str
+    artifact: ArtifactRecord | None
+    job: ArtifactJob | None = None
+
+
+def _canonical_json(payload: dict[str, Any]) -> str:
+    return json.dumps(payload, sort_keys=True, separators=(",", ":"))
+
+
+def _hash_payload(payload: dict[str, Any]) -> str:
+    return hashlib.sha256(_canonical_json(payload).encode("utf-8")).hexdigest()
+
+
+def _normalize_source_path(raw_path: str) -> str:
+    path = Path(raw_path).resolve()
+    try:
+        return path.relative_to(Path.cwd().resolve()).as_posix()
+    except ValueError:
+        return path.as_posix()
+
+
+def get_pipeline_version() -> str:
+    try:
+        from importlib.metadata import version
+        return version("docuscope-ca-online")
+    except Exception:
+        return "0.0.0+local"
+
+
+@lru_cache(maxsize=64)
+def _load_cached_keyness_bundle(
+    storage_uri: str,
+) -> tuple[pl.DataFrame, pl.DataFrame, pl.DataFrame, pl.DataFrame]:
+    """Load a keyness bundle once per process and reuse it across requests."""
+
+    artifact_dir = Path(storage_uri)
+    return (
+        pl.read_parquet(artifact_dir / "kw_pos.parquet"),
+        pl.read_parquet(artifact_dir / "kw_ds.parquet"),
+        pl.read_parquet(artifact_dir / "kt_pos.parquet"),
+        pl.read_parquet(artifact_dir / "kt_ds.parquet"),
+    )
+
+
+@lru_cache(maxsize=64)
+def _load_cached_frequency_bundle(
+    storage_uri: str,
+) -> tuple[pl.DataFrame, pl.DataFrame]:
+    """Load a frequency bundle once per process and reuse it across requests."""
+
+    artifact_dir = Path(storage_uri)
+    return (
+        pl.read_parquet(artifact_dir / "ft_pos.parquet"),
+        pl.read_parquet(artifact_dir / "ft_ds.parquet"),
+    )
+
+
+@lru_cache(maxsize=64)
[…]
