+++ webapp/queue/tasks.py
@@ -1,0 +1,557 @@
+"""RQ task functions for background-job smoke testing."""
+
+from __future__ import annotations
+
+from dataclasses import asdict
+from datetime import datetime, timezone
+import gzip
+import os
+import pickle
+from pathlib import Path
+
+import docuscospacy as ds
+import pandas as pd
+import polars as pl
+from polars import DataFrame
+from rq import get_current_job
+
+from webapp.persistence import (
+    ArtifactIdentity,
+    SharedArtifactWorkflow,
+    build_shared_collocation_identity,
+    build_shared_frequency_identity,
+    build_shared_keyness_identity,
+    build_shared_keyness_parts_identity,
+    build_shared_ngram_identity,
+    registry_service,
+)
+from webapp.utilities.ai.plotbot import run_plotbot_serialized_service
+from webapp.utilities.configuration.logging_config import get_logger
+
+
+logger = get_logger()
+
+
+def _get_shared_artifact_workflow() -> SharedArtifactWorkflow:
+    """Build a shared-artifact workflow against the current registry service."""
+
+    return SharedArtifactWorkflow(registry_service, logger)
+
+
+def _identity_from_job_row(job_row) -> ArtifactIdentity:
+    """Reconstruct a normalized artifact identity from a control-plane job row."""
+
+    return ArtifactIdentity(
+        artifact_type=job_row.artifact_type,
+        scope=job_row.scope,
+        owner_principal_id=job_row.requester_principal_id,
+        selector_hash=job_row.selector_hash,
+        selector_payload=job_row.selector_payload,
+        parameter_hash=job_row.parameter_hash,
+        parameter_payload=job_row.parameter_payload,
+        pipeline_version=job_row.pipeline_version,
+        model_version=job_row.model_version,
+    )
+
+
+def _load_internal_target_tokens(corpus_path: str) -> DataFrame:
+    """Load built-in target tokens from the file-backed internal corpus path."""
+
+    ds_tokens_path = Path(corpus_path) / "ds_tokens.gz"
+    with gzip.open(ds_tokens_path, "rb") as handle:
+        return pickle.load(handle)
+
+
+def _ensure_shared_frequency_artifact(corpus_path: str):
+    """Ensure a shared frequency artifact exists for a built-in target corpus."""
+
+    shared_artifact_workflow = _get_shared_artifact_workflow()
+    identity = build_shared_frequency_identity(target_source=corpus_path)
+    ready_artifact = registry_service.find_ready_artifact(identity)
+    if ready_artifact is not None:
+        return ready_artifact
+
+    decision = shared_artifact_workflow.reserve(
+        identity,
+        cache_name="frequency",
+        ready_loader=lambda: registry_service.find_ready_artifact(identity),
+        poll_attempts=20,
+        poll_interval_seconds=0.5,
+    )
+    if decision.state == "ready":
+        return decision.payload
+    if decision.state == "pending":
+        raise RuntimeError("Shared frequency artifact is still pending.")
+    if decision.state != "reserved":
+        raise RuntimeError(f"Unexpected frequency reservation state: {decision.state}")
+
+    ds_tokens = _load_internal_target_tokens(corpus_path)
+    ft_pos, ft_ds = ds.frequency_table(ds_tokens, count_by="both")
+    artifact = shared_artifact_workflow.store(
+        identity,
+        decision.job_id,
+        cache_name="frequency",
+        store_func=lambda artifact_identity: registry_service.store_frequency_bundle(
+            artifact_identity,
+            {"ft_pos": ft_pos, "ft_ds": ft_ds},
+        ),
+    )
+    if artifact is None:
+        raise RuntimeError("Failed to store shared frequency artifact.")
+    return artifact
+
+
[…]
