### 2026-07-27 6bc69d2
Create interpretation.py
