+++ webapp/utilities/ai/interpretation.py
@@ -1,0 +1,455 @@
+"""Evidence-grounded interpretation assistant utilities.
+
+This module avoids dataframe-agent execution. Application code builds compact,
+deterministic evidence from a selected table, then asks a chat provider to explain
+only that evidence in cautious student-facing language.
+"""
+
+from dataclasses import dataclass, field
+from typing import Any
+
+import pandas as pd
+import polars as pl
+
+from webapp.utilities.ai.providers import ChatCompletionProvider, get_default_chat_provider
+from webapp.utilities.ai.shared import LLM_MODEL
+
+
+DEFAULT_INTERPRETATION_MODEL = "qwen2.5:7b-instruct"
+DEFAULT_TEACHING_LENS = "understand_table"
+
+TEACHING_LENSES: dict[str, dict[str, str]] = {
+    "understand_table": {
+        "label": "Understand what this table shows",
+        "instruction": (
+            "Help the student understand what the selected table measures. "
+            "Explain useful columns or values, what the table can support, "
+            "and what it cannot show by itself."
+        ),
+        "sections": (
+            "What This Table Can Help You Notice; What To Be Careful About; "
+            "Questions To Ask Yourself; A Useful Next Step"
+        ),
+    },
+    "find_pattern": {
+        "label": "Find a pattern worth investigating",
+        "instruction": (
+            "Coach the student to identify a pattern that deserves follow-up. "
+            "Do not present the pattern as a final interpretation; frame it "
+            "as something to test with additional quantitative or context evidence."
+        ),
+        "sections": (
+            "Pattern To Investigate; Why It Might Matter; What Could Complicate It; "
+            "Questions To Ask Yourself; A Useful Next Step"
+        ),
+    },
+    "think_beyond_average": {
+        "label": "Think beyond the average",
+        "instruction": (
+            "Coach the student to reason beyond means, ranks, or top-line values. "
+            "Emphasize spread, outliers, median, document-level concentration, "
+            "sample size, and whether a few cases may drive the pattern."
+        ),
+        "sections": (
+            "Start With The Average; Check Variation; Possible Pitfalls; "
+            "Questions To Ask Yourself; A Better Quantitative Check"
+        ),
+    },
+    "check_claim": {
+        "label": "Check whether my claim is supported",
+        "instruction": (
+            "Evaluate the student's proposed claim against the supplied evidence. "
+            "Separate what the table supports, what it does not support, and what "
+            "would require additional context or quantitative checks."
+        ),
+        "sections": (
+            "What Your Claim Uses From The Table; What The Table Does Not Yet Show; "
+            "Quantitative Check; Context Check; A More Cautious Revision"
+        ),
+    },
+    "choose_next_step": {
+        "label": "Choose a next analysis",
+        "instruction": (
+            "Suggest a responsible next analysis step the student could take. "
+            "Prefer concrete corpus-analysis actions such as KWIC, collocations, "
+            "group comparison, plotting, or checking document-level distribution."
+        ),
+        "sections": (
+            "What You Have So Far; What You Still Need; Recommended Next Analysis; "
+            "Why That Step Helps; Question To Take Forward"
+        ),
+    },
+    "draft_cautious_claim": {
+        "label": "Draft cautious wording",
+        "instruction": (
+            "Help the student phrase a cautious evidence-based claim without doing "
+            "the interpretation for them. Make uncertainty and limits explicit."
+        ),
+        "sections": (
+            "What The Evidence Can Support; What To Avoid Overstating; "
+            "Possible Cautious Wording; How To Strengthen The Claim"
[…]
