### 2026-07-27 6ecbc0f
Update llm_core.py
