### 2026-07-27 9a52071
Create plotbot_benchmark.py
