+++ webapp/utilities/ai/plotbot_benchmark.py
@@ -1,0 +1,193 @@
+"""Small Plotbot benchmark harness for local model evaluation."""
+
+import json
+import re
+import time
+from dataclasses import asdict, dataclass, field
+
+import pandas as pd
+
+from webapp.utilities.ai.plotbot import generate_plotbot_code_and_result
+from webapp.utilities.ai.providers import ChatCompletionProvider
+
+
+DEFAULT_LLM_PARAMS = {
+    "temperature": 0.1,
+    "max_tokens": 500,
+    "top_p": 0.7,
+    "frequency_penalty": 0,
+    "presence_penalty": 0,
+}
+
+
+@dataclass(frozen=True)
+class PlotbotBenchmarkCase:
+    """A single Plotbot prompt benchmark case."""
+
+    name: str
+    prompt: str
+    plot_lib: str = "plotly.express"
+    expected_columns: set[str] = field(default_factory=set)
+    forbidden_columns: set[str] = field(default_factory=set)
+    code_chunk: str | None = None
+
+
+@dataclass(frozen=True)
+class PlotbotBenchmarkResult:
+    """Compact result for one Plotbot benchmark case."""
+
+    name: str
+    plot_lib: str
+    success: bool
+    result_type: str
+    latency_seconds: float
+    code_length: int
+    has_import: bool
+    creates_fig: bool
+    uses_expected_columns: bool
+    hallucinated_columns: list[str]
+    error: str | None = None
+
+
+DEFAULT_CASES = [
+    PlotbotBenchmarkCase(
+        name="basic_bar",
+        prompt="Make a bar chart of value by category.",
+        expected_columns={"category", "value"}
+    ),
+    PlotbotBenchmarkCase(
+        name="axis_labels",
+        prompt="Make a bar chart and label the x-axis Category and the y-axis Value.",
+        expected_columns={"category", "value"}
+    ),
+    PlotbotBenchmarkCase(
+        name="color_by_group",
+        prompt="Make a scatter plot of value by score and color the points by group.",
+        expected_columns={"score", "value", "group"}
+    ),
+    PlotbotBenchmarkCase(
+        name="gridlines_title",
+        prompt="Make a line chart with a clear title and visible gridlines.",
+        expected_columns={"category", "value"}
+    ),
+    PlotbotBenchmarkCase(
+        name="tag_rf_bar",
+        prompt="Make a bar plot of RF vs Tag.",
+        expected_columns={"Tag", "RF"}
+    ),
+    PlotbotBenchmarkCase(
+        name="top_tags_by_rf",
+        prompt="Show the top 3 tags by RF as a horizontal bar chart.",
+        expected_columns={"Tag", "RF"}
+    ),
+    PlotbotBenchmarkCase(
+        name="ignore_missing_column",
+        prompt="Make a bar plot of RF vs ImaginaryColumn.",
+        expected_columns={"RF"},
+        forbidden_columns={"ImaginaryColumn"}
+    ),
+    PlotbotBenchmarkCase(
+        name="refine_existing_plot",
+        prompt="Update the previous plot with a clearer title and light gridlines.",
+        expected_columns={"Tag", "RF"},
+        code_chunk="fig = px.bar(df, x='Tag', y='RF')"
+    ),
+]
+
+
+def sample_plotbot_dataframe() -> pd.DataFrame:
+    """Return a small dataframe with categorical and numeric plotting options."""
+    return pd.DataFrame({
+        "category": ["A", "B", "C", "D"],
+        "group": ["control", "control", "treatment", "treatment"],
+        "value": [10, 14, 9, 18],
+        "score": [0.2, 0.4, 0.7, 0.9],
[…]
