### 2026-07-27 e273c5c
Create plotbot_smoke.py
