### 2026-07-27 c9eb7ef
Update plotbot.py
