+++ webapp/utilities/ai/plotbot.py
@@ -7,6 +7,7 @@ executable plotting code and can refine it based on user feedback.
 """
 
 import hashlib
+from dataclasses import dataclass
 
 import pandas as pd
 import polars as pl
@@ -29,9 +30,8 @@ from webapp.utilities.ai.shared import (
     prune_message_thread, fig_to_svg, increment_session_quota
 )
 from webapp.utilities.ai.code_execution import is_code_safe, strip_imports
-from webapp.utilities.ai.enterprise_integration import (
-    make_protected_openai_call, determine_api_key_type
-)
+from webapp.utilities.ai.enterprise_integration import determine_api_key_type
+from webapp.utilities.ai.providers import ChatCompletionProvider, get_default_chat_provider
 from webapp.utilities.storage.backend_factory import get_session_backend
 from webapp.config.unified import get_ai_config
 from webapp.utilities.core import app_core
@@ -51,6 +51,36 @@ FORBIDDEN_PATTERNS = [
     r'^\s*sys\.',            # sys. usage at line start
     r'^\s*subprocess\.',     # subprocess. usage at line start
 ]
+
+
+@dataclass(frozen=True)
+class PlotbotServiceResult:
+    """UI-independent result for one Plotbot generation/execution request."""
+
+    code: str | None
+    result: dict
+    used_cached_code: bool = False
+
+    @property
+    def success(self) -> bool:
+        """Return True when code execution produced a plot."""
+        return self.result.get("type") == "plot"
+
+
+@dataclass(frozen=True)
+class PlotbotSerializedResult:
+    """Queue-friendly Plotbot result with no live figure object."""
+
+    code: str | None
+    result_type: str
+    error: str | None = None
+    plot_svg: str | None = None
+    used_cached_code: bool = False
+
+    @property
+    def success(self) -> bool:
+        """Return True when execution produced a plot."""
+        return self.result_type == "plot"
 
 # AI logging is automatically configured by importing the centralized logging system
 
@@ -240,7 +270,9 @@ def plotbot_code_generate_or_update(
     schema: str,
     api_key: str,
     llm_params: dict,
-    code_chunk: str = None
+    code_chunk: str = None,
+    chat_provider: ChatCompletionProvider | None = None,
+    track_quota: bool = True
 ) -> str:
     """
     Generate or update plotting code using the LLM.
@@ -300,6 +332,7 @@ def plotbot_code_generate_or_update(
     - Use plotly.express functions like px.bar(), px.scatter(), px.line(), etc.
     - Assign the result to a variable called 'fig'.
     - Do not call 'fig.show()'.
+    - For gridlines, use fig.update_xaxes(showgrid=True) and fig.update_yaxes(showgrid=True). Do not use fig.update_layout(grid_xaxis=...) or fig.update_layout(grid_yaxis=...).
 
     Example:
     fig = px.bar(df, x='col1', y='col2')
@@ -330,11 +363,14 @@ def plotbot_code_generate_or_update(
     - Do not include any import statements.
     - The DataFrame is called 'df'.
     - Use only columns that exist in the DataFrame. If the user mentions a column that does not exist, ignore it and use available columns instead.
+    - Preserve any valid columns named by the user. Replace only the missing column names.
+    - When replacing a missing column, choose one available non-numeric column for categorical axes and one available numeric column for values. If one requested column is valid, keep it and choose a compatible replacement column near it in the column list above.
     - If the request involves numeric data (like line charts, bar charts, histograms), use only numeric columns.
     - If the request involves categorical or non-numeric data (like pie charts or scatter plots with labels), you can use non-numeric columns.
     - Ensure the code is error-free and matches the DataFrame schema.
     - If you need to set axis labels or titles, use generic names if the user does not specify.
[…]
