+++ webapp/utilities/ai/providers.py
@@ -1,0 +1,226 @@
+"""
+Model provider boundaries for AI-assisted tools.
+
+These classes keep assistant code independent from a specific model runner while
+preserving the current protected OpenAI path as the default implementation.
+"""
+
+from dataclasses import dataclass, field
+from typing import Any, Protocol
+from functools import lru_cache
+import json
+import os
+from urllib.error import URLError
+from urllib.request import urlopen
+
+from webapp.utilities.ai.enterprise_integration import make_protected_openai_call
+
+
+LOCAL_PROVIDER_NAMES = {"openai-compatible", "openai_compatible", "local"}
+DEFAULT_LOCAL_BASE_URL = "http://localhost:11434/v1"
+DEFAULT_PLOTBOT_LOCAL_MODEL = "qwen2.5-coder:7b-instruct"
+
+
+@dataclass(frozen=True)
+class OpenAICompatibleProviderHealth:
+    """Readiness result for an OpenAI-compatible model endpoint."""
+
+    available: bool
+    base_url: str
+    configured_model: str | None = None
+    resolved_model: str | None = None
+    model_ids: list[str] = field(default_factory=list)
+    error: str | None = None
+
+
+def _collect_streamed_text(response: Any) -> str:
+    """Collect text from an OpenAI-compatible streamed response."""
+    full_response = ""
+    for chunk in response:
+        chunk_content = chunk.choices[0].delta.content
+        if chunk_content:
+            full_response += chunk_content
+    return full_response
+
+
+class ChatCompletionProvider(Protocol):
+    """Minimal synchronous chat-completion provider contract."""
+
+    def generate_text(
+        self,
+        api_key: str,
+        request_params: dict[str, Any],
+        cache_key: str | None = None
+    ) -> str | dict[str, str]:
+        """Return generated text or a normalized error dictionary."""
+
+
+class ProtectedOpenAIChatProvider:
+    """Chat provider backed by the existing protected OpenAI integration."""
+
+    def generate_text(
+        self,
+        api_key: str,
+        request_params: dict[str, Any],
+        cache_key: str | None = None
+    ) -> str | dict[str, str]:
+        response = make_protected_openai_call(
+            api_key=api_key,
+            request_params=request_params,
+            request_type="chat_completion",
+            cache_key=cache_key
+        )
+
+        if isinstance(response, dict) and response.get("type") == "error":
+            return response
+
+        return _collect_streamed_text(response)
+
+
+class OpenAICompatibleChatProvider:
+    """Chat provider for local or hosted OpenAI-compatible endpoints."""
+
+    def __init__(self, base_url: str, model: str | None = None):
+        self.base_url = base_url
+        self.model = model
+
+    def generate_text(
+        self,
+        api_key: str,
+        request_params: dict[str, Any],
+        cache_key: str | None = None
+    ) -> str | dict[str, str]:
+        import openai
+
+        try:
+            params = dict(request_params)
+            if self.model:
+                params["model"] = self.model
+
+            client = openai.OpenAI(
+                api_key=api_key or "local-model",
+                base_url=self.base_url
+            )
+            response = client.chat.completions.create(**params)
+            return _collect_streamed_text(response)
+        except Exception as exc:
+            return {
+                "type": "error",
+                "value": f"Model provider request failed: {str(exc)}"
+            }
+
+
+def check_openai_compatible_provider_health(
+    base_url: str | None = None,
+    model: str | None = None,
+    timeout_seconds: float = 1.0,
+) -> OpenAICompatibleProviderHealth:
[…]
