+++ webapp/utilities/analysis/corpus_generators.py
@@ -8,9 +8,240 @@ analyses including n-grams, collocations, and keyword-in-context searches.
 import streamlit as st
 import docuscospacy as ds
 
+from webapp.persistence import (
+    SharedArtifactWorkflow,
+    build_shared_collocation_identity,
+    build_shared_ngram_identity,
+    registry_service,
+)
 from webapp.utilities.core import app_core
+from webapp.utilities.configuration.logging_config import get_logger
 from webapp.utilities.state import CorpusKeys, TargetKeys, WarningKeys, SessionKeys
 from webapp.utilities.corpus import get_corpus_data_manager
+from webapp.utilities.session import safe_session_get
+
+
+logger = get_logger()
+shared_artifact_workflow = SharedArtifactWorkflow(registry_service, logger)
+
+
+def _get_shared_collocation_identity(
+        user_session_id: str,
+        node_word: str,
+        node_tag: str,
+        to_left: int,
+        to_right: int,
+        stat_mode: str,
+        count_by: str
+):
+    """Return a shared collocation identity for built-in target corpora."""
+
+    session = st.session_state.get(user_session_id, {})
+    target_db = safe_session_get(session, SessionKeys.TARGET_DB, "")
+    if not target_db:
+        return None
+
+    return build_shared_collocation_identity(
+        target_source=target_db,
+        node_word=node_word,
+        node_tag=node_tag,
+        to_left=to_left,
+        to_right=to_right,
+        stat_mode=stat_mode,
+        count_by=count_by,
+    )
+
+
+def _get_shared_ngram_identity(
+        user_session_id: str,
+        analysis_type: str,
+        ngram_span: int,
+        count_by: str,
+        from_anchor: str | None = None,
+        node_word: str | None = None,
+        tag: str | None = None,
+        position: int | None = None,
+        search_type: str | None = None,
+):
+    """Return a shared n-gram/cluster identity for built-in target corpora."""
+
+    session = st.session_state.get(user_session_id, {})
+    target_db = safe_session_get(session, SessionKeys.TARGET_DB, "")
+    if not target_db:
+        return None
+
+    return build_shared_ngram_identity(
+        target_source=target_db,
+        analysis_type=analysis_type,
+        ngram_span=ngram_span,
+        count_by=count_by,
+        from_anchor=from_anchor,
+        node_word=node_word,
+        tag=tag,
+        position=position,
+        search_type=search_type,
+    )
+
+
+def _ngram_parameters(
+        analysis_type: str,
+        ngram_span: int,
+        count_by: str,
+        from_anchor: str | None = None,
+        node_word: str | None = None,
+        tag: str | None = None,
+        position: int | None = None,
+        search_type: str | None = None,
+) -> dict:
+    """Return stable parameters for shared n-gram/cluster generation."""
+
+    return {
+        "analysis_type": analysis_type,
+        "ngram_span": ngram_span,
+        "count_by": count_by,
+        "from_anchor": from_anchor,
+        "node_word": node_word.strip() if isinstance(node_word, str) else node_word,
+        "tag": tag,
+        "position": position,
+        "search_type": search_type,
+    }
+
+
+def attach_ngram_artifact(
+        user_session_id: str,
+        artifact_id: int,
+        artifact_type: str | None = None,
+) -> bool:
+    """Attach a ready n-gram/cluster artifact to the target corpus session."""
+
+    if artifact_type is None:
+        artifact = registry_service.get_artifact_by_id(artifact_id)
+        if artifact is None or artifact.status != "ready":
+            return False
+        artifact_type = artifact.artifact_type
+
+    manager = get_corpus_data_manager(user_session_id, CorpusKeys.TARGET)
[…]
