### 2026-07-27 d8bfb03
Create dfm_statistics.py
