+++ webapp/utilities/analysis/dfm_statistics.py
@@ -1,0 +1,273 @@
+"""Deterministic statistics for document-feature matrices."""
+
+from dataclasses import dataclass
+from typing import Literal
+
+import docuscospacy as ds
+import polars as pl
+
+
+DfmStatistic = Literal[
+    "mean_rf",
+    "median_rf",
+    "std_rf",
+    "cv_rf",
+    "range_pct",
+    "dp_dispersion",
+    "total_af",
+]
+RankAxis = Literal["features", "groups"]
+
+STATISTIC_LABELS: dict[DfmStatistic, str] = {
+    "mean_rf": "Mean RF",
+    "median_rf": "Median RF",
+    "std_rf": "Standard deviation of RF",
+    "cv_rf": "Coefficient of variation of RF",
+    "range_pct": "Document range",
+    "dp_dispersion": "DP dispersion",
+    "total_af": "Total AF",
+}
+
+STATISTIC_NOTES: dict[DfmStatistic, str] = {
+    "mean_rf": "Mean RF is the average normalized feature frequency across documents.",
+    "median_rf": "Median RF is less affected by unusually high-frequency documents than the mean.",
+    "std_rf": "Standard deviation of RF shows how much normalized frequency varies across documents.",
+    "cv_rf": "Coefficient of variation divides RF standard deviation by mean RF; it helps compare variability for features with different average frequencies.",
+    "range_pct": "Document range is the percent of documents where the feature appears at least once.",
+    "dp_dispersion": "DP dispersion estimates how unevenly a feature is distributed across documents. Higher values suggest stronger concentration in fewer documents.",
+    "total_af": "Total AF is the raw count across documents and is sensitive to corpus or group size.",
+}
+
+
+@dataclass(frozen=True)
+class DfmStatisticsResult:
+    """Computed DFM statistics and a short deterministic explanation."""
+
+    table: pl.DataFrame
+    statistic: DfmStatistic
+    rank_axis: RankAxis
+    note: str
+
+
+def get_dfm_statistic_options() -> dict[DfmStatistic, str]:
+    """Return available DFM statistics for UI controls."""
+    return dict(STATISTIC_LABELS)
+
+
+def get_dfm_statistic_note(statistic: DfmStatistic) -> str:
+    """Return a short explanation for a DFM statistic."""
+    return STATISTIC_NOTES.get(statistic, STATISTIC_NOTES["mean_rf"])
+
+
+def _feature_columns(dfm: pl.DataFrame) -> list[str]:
+    return [column for column in dfm.columns if column not in {"doc_id", "Group"}]
+
+
+def _with_group_column(dfm: pl.DataFrame, groups: list[str] | None) -> pl.DataFrame:
+    if groups is None:
+        return dfm
+    if len(groups) != dfm.height:
+        raise ValueError("Group labels must align with DFM rows.")
+    if "Group" in dfm.columns:
+        return dfm.with_columns(pl.Series("Group", groups))
+    return dfm.insert_column(1, pl.Series("Group", groups))
+
+
+def _filter_output_features(dfm: pl.DataFrame, tagset: str) -> pl.DataFrame:
+    if tagset == "DocuScope" and "Untagged" in dfm.columns:
+        return dfm.drop("Untagged")
+    if tagset == "Parts-of-Speech" and "FU" in dfm.columns:
+        return dfm.drop("FU")
+    return dfm
+
+
+def available_dfm_features(dfm: pl.DataFrame, tagset: str) -> list[str]:
+    """Return selectable feature columns after tool-specific exclusions."""
+    return _feature_columns(_filter_output_features(dfm, tagset))
+
+
[…]
