+++ webapp/utilities/analysis/statistical_analysis.py
@@ -8,13 +8,403 @@ keyness analysis, and statistical comparisons between corpora.
 import polars as pl
 import streamlit as st
 import docuscospacy as ds
-from scipy.stats import pearsonr
+from time import perf_counter
 from webapp.utilities.core import app_core
-from webapp.utilities.state import (
-    CorpusKeys, TargetKeys, ReferenceKeys, WarningKeys
+from webapp.utilities.configuration.logging_config import get_logger
+from webapp.utilities.analysis.correlation import pearson_correlation
+from webapp.persistence import (
+    SharedArtifactWorkflow,
+    build_shared_keyness_identity,
+    build_shared_keyness_parts_identity,
+    registry_service,
 )
-from webapp.utilities.corpus import get_corpus_data, set_corpus_data
+from webapp.utilities.state import (
+    CorpusKeys, TargetKeys, ReferenceKeys, WarningKeys, SessionKeys
+)
+from webapp.utilities.corpus import (
+    get_corpus_data,
+    get_corpus_data_manager,
+    set_corpus_data,
+)
 from webapp.utilities.session import safe_session_get, get_or_init_user_session
+
+
+logger = get_logger()
+shared_artifact_workflow = SharedArtifactWorkflow(registry_service, logger)
+SLOW_GENERATION_TRIGGER_MS = 25
+FREQUENCY_RENDER_TRACE_KEY = "_frequency_render_trace"
+
+
+def _log_slow_generation_trigger(
+    operation: str,
+    start_time: float,
+    session_id: str,
+) -> None:
+    """Log slow pre-rerun generation trigger steps."""
+
+    elapsed_ms = (perf_counter() - start_time) * 1000
+    if elapsed_ms >= SLOW_GENERATION_TRIGGER_MS:
+        logger.warning(
+            "Slow generation trigger op={} session={} duration_ms={:.2f}",
+            operation,
+            session_id,
+            elapsed_ms,
+        )
+
+
+def _get_shared_keyness_identity(user_session_id: str, threshold: float, swap_target: bool):
+    """Return a shared keyness identity for built-in corpus comparisons."""
+
+    _, session = get_or_init_user_session()
+    target_db = safe_session_get(session, SessionKeys.TARGET_DB, "")
+    reference_db = safe_session_get(session, SessionKeys.REFERENCE_DB, "")
+    if not target_db or not reference_db:
+        return None
+
+    return build_shared_keyness_identity(
+        target_source=target_db,
+        reference_source=reference_db,
+        threshold=threshold,
+        swap_target=swap_target,
+    )
+
+
+def _normalize_category_selection(categories: list) -> list[str]:
+    """Return a stable category selection for artifact identity matching."""
+
+    return sorted(str(category) for category in categories)
+
+
+def _get_keyness_parts_metadata(
+    tar_list: list[str],
+    ref_list: list[str],
+    tar_tokens_pos: int,
+    ref_tokens_pos: int,
+    tar_tokens_ds: int,
+    ref_tokens_ds: int,
+    tar_ndocs: int,
+    ref_ndocs: int,
+) -> dict[str, list[str]]:
+    """Build the metadata payload expected by the corpus-parts results page."""
+
+    return {
+        "keyness_parts": [
+            list(tar_list),
+            list(ref_list),
+            str(tar_tokens_pos),
+            str(ref_tokens_pos),
+            str(tar_tokens_ds),
+            str(ref_tokens_ds),
+            str(tar_ndocs),
+            str(ref_ndocs),
+        ]
+    }
+
+
+def _get_shared_keyness_parts_identity(
+    user_session_id: str,
+    threshold: float,
+    swap_target: bool,
+):
+    """Return a shared keyness-parts identity for built-in target corpora."""
+
+    _, session = get_or_init_user_session()
[…]
