### 2026-07-27 79ed5a9
Update user_authorization.py
