+++ webapp/utilities/auth/user_authorization.py
@@ -2,7 +2,7 @@
 User Authorization System
 
 This module provides role-based access control for the corpus analysis application.
-It manages user permissions through a local SQLite database and integrates with
+It manages user permissions through the Postgres control plane and integrates with
 Google OAuth authentication.
 
 Features:
@@ -10,19 +10,17 @@ Features:
 - Permission management
 - Session-based authorization caching
 - Audit trail for access attempts
-- Automatic database initialization
+- Automatic control-plane initialization
 """
 
-import sqlite3
-import json
 from datetime import datetime, timezone
 from functools import wraps
 from typing import List, Dict, Any, Optional, Callable
-from pathlib import Path
 
 import streamlit as st
 
 from webapp.config.unified import get_config
+from webapp.persistence.authorization_service import AuthorizationService
 from webapp.utilities.configuration.logging_config import get_logger
 
 logger = get_logger()
@@ -50,20 +48,24 @@ DEFAULT_ROLES = {
 
 # Cache for authorization checks during session
 _auth_cache = {}
+_authorization_service: AuthorizationService | None = None
 
 
-def get_auth_db_path() -> Path:
-    """Get the path to the authorization database."""
-    users_dir = Path("webapp/_users")
-    users_dir.mkdir(exist_ok=True)
-    return users_dir / "authorization.db"
+def _get_authorization_service() -> AuthorizationService:
+    """Return the shared Postgres-backed authorization service."""
+
+    global _authorization_service
+    if _authorization_service is None:
+        _authorization_service = AuthorizationService()
+    return _authorization_service
 
 
 def is_authorization_enabled() -> bool:
     """Check if user authorization is enabled based on configuration."""
-    # Authorization is disabled in desktop mode
+    # Authorization is disabled in desktop mode and test mode.
     desktop_mode = get_config('desktop_mode', 'global')
-    if desktop_mode:
+    test_mode = get_config('test_mode', 'global', False)
+    if desktop_mode or test_mode:
         return False
 
     # Check if explicitly enabled in config
@@ -71,93 +73,17 @@ def is_authorization_enabled() -> bool:
 
 
 def initialize_authorization_db() -> None:
-    """Initialize the authorization database with required tables and default data."""
+    """Initialize the authorization tables with required defaults."""
     if not is_authorization_enabled():
         return
 
-    db_path = get_auth_db_path()
-
     try:
-        with sqlite3.connect(db_path) as conn:
-            cursor = conn.cursor()
-
-            # Create user roles table
-            cursor.execute("""
-                CREATE TABLE IF NOT EXISTS user_roles (
-                    role_name TEXT PRIMARY KEY,
-                    description TEXT NOT NULL,
-                    permissions TEXT NOT NULL,
-                    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-                )
-            """)
-
-            # Create authorized users table
-            cursor.execute("""
-                CREATE TABLE IF NOT EXISTS authorized_users (
-                    email TEXT PRIMARY KEY,
-                    role TEXT NOT NULL DEFAULT 'user',
-                    permissions TEXT,
-                    added_by TEXT,
-                    added_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                    last_accessed TIMESTAMP,
-                    active BOOLEAN DEFAULT TRUE,
-                    FOREIGN KEY (role) REFERENCES user_roles (role_name)
-                )
-            """)
-
-            # Create access log table for auditing
-            cursor.execute("""
-                CREATE TABLE IF NOT EXISTS access_log (
-                    id INTEGER PRIMARY KEY AUTOINCREMENT,
-                    email TEXT NOT NULL,
[…]
