### 2026-07-27 789e388
Update config_manager.py
