### 2026-07-27 879915c
Update data_manager.py
