+++ webapp/utilities/corpus/data_manager.py
@@ -9,6 +9,12 @@ This module provides a centralized system for managing corpus data with:
 """
 
 import gc
+import gzip
+import os
+import pickle
+from collections import OrderedDict
+from threading import Lock
+from time import perf_counter
 from typing import Dict, Optional, Any
 import polars as pl
 import streamlit as st
@@ -16,9 +22,110 @@ import docuscospacy as ds
 
 from webapp.utilities.memory import DataFrameCache, lazy_computation
 from webapp.utilities.configuration.logging_config import get_logger
+from webapp.persistence import (
+    SharedArtifactWorkflow,
+    build_shared_frequency_identity,
+    registry_service,
+)
 from webapp.utilities.session.session_persistence import auto_persist_session
+from webapp.utilities.session.session_persistence import mark_session_dirty
 
 logger = get_logger()
+shared_artifact_workflow = SharedArtifactWorkflow(registry_service, logger)
+ARTIFACT_REF_KEY = "_artifact_refs"
+SLOW_CORPUS_OPERATION_MS = 50
+ARTIFACT_FRAME_CACHE_MAX_ITEMS = 32
+ArtifactFrameCacheOwner = int | str
+_artifact_frame_cache: OrderedDict[
+    tuple[ArtifactFrameCacheOwner, str],
+    pl.DataFrame,
+] = OrderedDict()
+_artifact_frame_cache_locks: dict[tuple[ArtifactFrameCacheOwner, str], Lock] = {}
+_artifact_frame_cache_locks_guard = Lock()
+
+
+def _get_cached_artifact_frame(
+    cache_owner: ArtifactFrameCacheOwner,
+    key: str,
+) -> Optional[pl.DataFrame]:
+    """Return an ephemeral in-process cache hit for an artifact-backed frame."""
+
+    cache_key = (cache_owner, key)
+    cached = _artifact_frame_cache.get(cache_key)
+    if cached is None:
+        return None
+
+    _artifact_frame_cache.move_to_end(cache_key)
+    return cached
+
+
+def _set_cached_artifact_frame(
+    cache_owner: ArtifactFrameCacheOwner,
+    key: str,
+    value: pl.DataFrame,
+) -> None:
+    """Store an artifact-backed frame in the ephemeral in-process cache."""
+
+    cache_key = (cache_owner, key)
+    _artifact_frame_cache[cache_key] = value
+    _artifact_frame_cache.move_to_end(cache_key)
+
+    while len(_artifact_frame_cache) > ARTIFACT_FRAME_CACHE_MAX_ITEMS:
+        _artifact_frame_cache.popitem(last=False)
+
+
+def _get_artifact_frame_lock(cache_owner: ArtifactFrameCacheOwner, key: str) -> Lock:
+    """Return the shared lock guarding first-load cache misses for one frame."""
+
+    cache_key = (cache_owner, key)
+    with _artifact_frame_cache_locks_guard:
+        lock = _artifact_frame_cache_locks.get(cache_key)
+        if lock is None:
+            lock = Lock()
+            _artifact_frame_cache_locks[cache_key] = lock
+        return lock
+
+
+def _build_shared_frequency_cache_owner(identity) -> str:
+    """Build a stable in-process cache owner key for shared frequency identities."""
+
+    return (
+        "shared-frequency:"
+        f"{identity.selector_hash}:"
+        f"{identity.parameter_hash}:"
+        f"{identity.pipeline_version}:"
+        f"{identity.model_version}"
+    )
+
+
+def _build_file_backed_cache_owner(path: str) -> str:
+    """Build a stable cache owner key for a file-backed artifact path."""
+
+    try:
+        stat_result = os.stat(path)
+        return (
+            "file-backed:"
[…]
