### 2026-07-27 6be67eb
Update scatterplots.py
