+++ webapp/utilities/processing/corpus_processing.py
@@ -5,8 +5,12 @@ This module provides functions for processing internal corpora, external corpora
 newly uploaded text, and corpus finalization.
 """
 
+import gzip
+import hashlib
 import os
-import io
+import pickle
+from pathlib import Path
+from time import perf_counter
 import unidecode
 import polars as pl
 import streamlit as st
@@ -15,12 +19,28 @@ import docuscospacy as ds
 # Module-specific imports
 from webapp.utilities.processing.corpus_loading import load_corpus_internal, load_corpus_new
 from webapp.utilities.session import (
-    init_metadata_target, init_metadata_reference
+    build_corpus_metadata_descriptor,
+    init_metadata_target,
+    init_metadata_reference,
+    set_session_persistence_policy,
+    write_metadata_descriptor_sidecar,
 )
-from webapp.utilities.analysis.data_validation import check_language, check_corpus_external, check_corpus_new  # noqa: E501
-from webapp.utilities.state import LoadCorpusKeys, SessionKeys
+from webapp.utilities.session.session_persistence import (
+    auto_persist_session,
+    mark_session_dirty,
+)
+from webapp.utilities.analysis.data_validation import check_corpus_external, check_corpus_new  # noqa: E501
+from webapp.utilities.state import (
+    CorpusPersistencePolicy,
+    LoadCorpusKeys,
+    SessionKeys,
+)
 from webapp.utilities.corpus import get_corpus_manager
 from webapp.utilities.core import app_core
+from webapp.utilities.configuration.logging_config import get_logger
+from webapp.config.unified import get_config
+from webapp.persistence import registry_service
+from webapp.persistence.registry import FREQUENCY_ARTIFACT_TYPE
 
 # Warning constants for corpus processing
 WARNING_CORRUPT_TARGET = 10
@@ -28,11 +48,227 @@ WARNING_CORRUPT_REFERENCE = 11
 WARNING_DUPLICATE_REFERENCE = 21
 WARNING_EXCLUDED_TARGET = 40
 WARNING_EXCLUDED_REFERENCE = 41
+TEST_MODE = get_config('test_mode', 'global', False)
+PROCESS_TARGET_PROBE_ENV = "DOCUSCOPE_PROCESS_TARGET_PROBE"
+PROCESS_TARGET_PROBE_NONE = "full"
+PROCESS_TARGET_PROBE_NO_METADATA = "no_metadata"
+PROCESS_TARGET_PROBE_METADATA_NO_PERSIST = "metadata_no_persist"
+PROCESS_TARGET_PROBE_NO_RERUN = "no_rerun"
+PROCESS_TARGET_PROBE_SPLIT_READY = "split_ready"
+SLOW_INTERNAL_PROCESS_STAGE_MS = 50
+SESSION_CORPUS_ARTIFACTS_DIRNAME = "corpora"
+SESSION_CORE_DATA_FILENAME = "ds_tokens.gz"
+PROCESS_TARGET_PROBE_STAGE_KEY = "load_test_process_target_probe_stage"
+PROCESS_TARGET_PROBE_TIMINGS_KEY = "load_test_process_target_probe_stage_timings_ms"
+PROCESS_TARGET_PROBE_FLAT_TIMINGS_KEY = "load_test_process_target_probe_stage_timings_flat"
+PROCESS_TARGET_PROBE_SUMMARY_KEY = "load_test_process_target_probe_stage_summary"
+logger = get_logger()
 
 
-def finalize_corpus_load(ds_tokens, user_session_id: str, corpus_type: str) -> None:
+def _log_slow_process_internal_stage(
+    user_session_id: str,
+    corpus_type: str,
+    stage: str,
+    start_time: float,
+    probe_mode: str,
+) -> None:
[…]
