### 2026-07-27 db925da
Update metadata_handlers.py
