### 2026-07-27 ec5ecf5
Update session_core.py
