+++ webapp/utilities/session/session_core.py
@@ -2,15 +2,223 @@
 This module contains the core session initialization, update, and management.
 """
 
+from functools import lru_cache
+from pathlib import Path
+import gzip
+import json
+import pickle
+
 import polars as pl
 import streamlit as st
+from time import perf_counter
 
-from webapp.utilities.state import SessionKeys, MetadataKeys
+from webapp.utilities.corpus import get_corpus_manager
+from webapp.utilities.state import (
+    CorpusPersistencePolicy,
+    MetadataKeys,
+    SessionKeys,
+)
 from webapp.utilities.common import get_doc_cats
+from webapp.utilities.configuration.logging_config import get_logger
 from webapp.utilities.session.session_persistence import (
     load_persistent_session,
-    auto_persist_session
+    auto_persist_session,
+    mark_session_dirty,
 )
+
+
+COMMON_DICTIONARY_TAGS = {
+    'Actors', 'Organization', 'Planning', 'Sentiment', 'Signposting', 'Stance'
+}
+METADATA_DESCRIPTOR_FILE = "metadata_descriptor.json"
+SLOW_SESSION_UPDATE_MS = 25
+SLOW_METADATA_DESCRIPTOR_MS = 50
+logger = get_logger()
+
+
+def _log_slow_session_update(
+    session_id: str,
+    key: str,
+    start_time: float,
+    persist_ms: float,
+) -> None:
+    """Log slow session-state mutations on the analysis hot path."""
+
+    elapsed_ms = (perf_counter() - start_time) * 1000
+    if elapsed_ms >= SLOW_SESSION_UPDATE_MS:
+        logger.warning(
+            (
+                "Slow session update session={} key={} persist_ms={:.2f} "
+                "duration_ms={:.2f}"
+            ),
+            session_id,
+            key,
+            persist_ms,
+            elapsed_ms,
+        )
+
+
+def _log_slow_metadata_descriptor(
+    session_id: str,
+    corpus_type: str,
+    source: str,
+    start_time: float,
+) -> None:
+    """Log slow metadata descriptor resolution on the corpus-load path."""
+
+    elapsed_ms = (perf_counter() - start_time) * 1000
+    if elapsed_ms >= SLOW_METADATA_DESCRIPTOR_MS:
+        logger.warning(
+            (
+                "Slow metadata descriptor session={} corpus_type={} source={} "
+                "duration_ms={:.2f}"
+            ),
+            session_id,
+            corpus_type,
+            source,
+            elapsed_ms,
+        )
+
+
+def build_corpus_metadata_descriptor(df: pl.DataFrame) -> dict:
+    """Build the minimal eager metadata contract shared across corpus sources."""
+
+    ds_tag_series = df.get_column("ds_tag")
+    pos_tag_series = df.get_column("pos_tag")
+    doc_id_series = df.get_column("doc_id")
+
+    ds_tags = sorted(
+        tag for tag in ds_tag_series.unique().to_list()
+        if tag != "Untagged"
+    )
+    tags_pos = sorted(
+        tag for tag in pos_tag_series.unique().to_list()
+        if tag != "Y"
+    )
+    doc_ids = sorted(doc_id_series.unique().to_list())
+
+    model = (
+        'Common Dictionary'
+        if any(tag in ds_tag for ds_tag in ds_tags for tag in COMMON_DICTIONARY_TAGS)
+        else 'Large Dictionary'
+    )
+
+    return {
+        MetadataKeys.TOKENS_POS: df.group_by(
+            ["doc_id", "pos_id", "pos_tag"]
+        ).agg(pl.col("token").str.join("")).filter(
+            pl.col("pos_tag") != "Y"
+        ).height,
+        MetadataKeys.TOKENS_DS: df.group_by(
[…]
