### 2026-07-27 75deea4
Update session_management.py
