### 2026-07-27 bc25285
Update session_persistence.py
