+++ webapp/utilities/session/session_persistence.py
@@ -5,15 +5,407 @@ This module provides automatic session persistence by hooking into Streamlit's
 session state management and syncing with the SQLite session backend.
 """
 
+import polars as pl
 import streamlit as st
 from typing import Dict, Any
 from datetime import datetime, timezone
 import hashlib
 import json
+from time import perf_counter
 
 from webapp.utilities.configuration.logging_config import get_logger
+from webapp.utilities.state.session_keys import (
+    CorpusPersistencePolicy,
+    MetadataKeys,
+    SessionKeys,
+)
 
 logger = get_logger()
+SLOW_PERSISTENCE_OPERATION_MS = 25
+MIN_PERSIST_INTERVAL_SECONDS = 0.5
+ARTIFACT_REF_KEY = "_artifact_refs"
+PERSISTED_CORPUS_KEYS = ("target", "reference")
+CORPUS_POLICY_KEYS = {
+    "target": SessionKeys.TARGET_PERSISTENCE_POLICY,
+    "reference": SessionKeys.REFERENCE_PERSISTENCE_POLICY,
+}
+PERSISTED_SESSION_KEYS = (
+    SessionKeys.CORPUS_PERSISTENCE_POLICY,
+    SessionKeys.TARGET_PERSISTENCE_POLICY,
+    SessionKeys.REFERENCE_PERSISTENCE_POLICY,
+    SessionKeys.HAS_TARGET,
+    SessionKeys.TARGET_DB,
+    SessionKeys.HAS_META,
+    SessionKeys.HAS_REFERENCE,
+    SessionKeys.REFERENCE_DB,
+)
+PERSISTED_METADATA_KEYS = (
+    MetadataKeys.TOKENS_POS,
+    MetadataKeys.TOKENS_DS,
+    MetadataKeys.NDOCS,
+    MetadataKeys.MODEL,
+    MetadataKeys.DOCIDS,
+    MetadataKeys.TAGS_DS,
+    MetadataKeys.TAGS_POS,
+    MetadataKeys.DOCCATS,
+)
+NON_DURABLE_PERSISTENCE_POLICIES = {
+    CorpusPersistencePolicy.TEMPORARY_SESSION_ONLY,
+    CorpusPersistencePolicy.LOCAL_EXPORT_ONLY,
+}
+SUPPORTED_PERSISTENCE_POLICIES = {
+    CorpusPersistencePolicy.SERVER_SAVED,
+    CorpusPersistencePolicy.TEMPORARY_SESSION_ONLY,
+    CorpusPersistencePolicy.LOCAL_EXPORT_ONLY,
+}
+
+
+def _normalize_session_mapping(raw_value: Any) -> Dict[str, Any]:
+    """Normalize DataFrame-backed session payloads into plain dictionaries."""
+
+    if raw_value is None:
+        return {}
+
+    if hasattr(raw_value, "to_dict") and hasattr(raw_value, "columns"):
+        return raw_value.to_dict(as_series=False)
+
+    if isinstance(raw_value, dict):
+        return raw_value.copy()
+
+    return {}
+
+
+def _normalize_scalar(value: Any) -> Any:
+    """Collapse DataFrame-style single-item lists into scalar values."""
+
+    if isinstance(value, list):
+        return value[0] if value else None
+    return value
+
+
+def _normalize_persistence_policy(value: Any) -> str:
+    """Return a supported persistence policy or the default durable mode."""
+
+    if value in SUPPORTED_PERSISTENCE_POLICIES:
+        return value
+    return CorpusPersistencePolicy.SERVER_SAVED
+
+
+def _get_corpus_policy_from_mapping(
+    normalized: Dict[str, Any],
+    corpus_type: str,
+) -> str:
+    """Resolve a corpus policy from specific keys with legacy fallback."""
+
[…]
