+++ webapp/utilities/state/__init__.py
@@ -26,6 +26,7 @@ from webapp.utilities.state.session_keys import (
     SessionKeys,
     MetadataKeys,
     CorpusKeys,
+    CorpusPersistencePolicy,
     TargetKeys,
     ReferenceKeys,
     WarningKeys,
@@ -53,6 +54,7 @@ __all__ = [
     'SessionKeys',
     'MetadataKeys',
     'CorpusKeys',
+    'CorpusPersistencePolicy',
     'TargetKeys',
     'ReferenceKeys',
     'WarningKeys',
