### 2026-07-27 bd4f3d5
Update cache_management.py
