+++ webapp/utilities/storage/postgres_session_backend.py
@@ -1,0 +1,370 @@
+"""Postgres-backed session storage for the shared control plane."""
+
+from __future__ import annotations
+
+from datetime import datetime, timedelta, timezone
+from pathlib import Path
+import pickle
+from typing import Any, Dict, Optional
+
+from sqlalchemy import delete, func, select
+
+from webapp.persistence.database import create_session_factory, initialize_database_schema
+from webapp.persistence.models import SessionRecord, UserQuery
+from webapp.utilities.configuration.logging_config import get_logger
+
+
+logger = get_logger()
+
+
+class PostgresSessionBackend:
+    """Session backend backed by the SQLAlchemy/Postgres control plane."""
+
+    def __init__(self) -> None:
+        initialize_database_schema()
+        self._session_factory = create_session_factory()
+
+    def _session_artifact_root(self) -> Path:
+        """Return the root directory used for session-scoped corpus artifacts."""
+
+        return Path("webapp/_session") / "corpora"
+
+    def _collect_session_artifact_paths(self, data: Dict[str, Any]) -> list[Path]:
+        """Collect session-scoped file-backed artifact paths from persisted data."""
+
+        artifact_paths: list[Path] = []
+        artifact_root = self._session_artifact_root()
+
+        for corpus_key in ("target", "reference"):
+            corpus_state = data.get(corpus_key)
+            if not isinstance(corpus_state, dict):
+                continue
+
+            refs = corpus_state.get("_artifact_refs")
+            if not isinstance(refs, dict):
+                continue
+
+            for ref in refs.values():
+                if not isinstance(ref, dict):
+                    continue
+                if ref.get("storage_type") != "gzip_pickle":
+                    continue
+
+                path_value = ref.get("path")
+                if not path_value:
+                    continue
+
+                artifact_path = Path(path_value)
+                try:
+                    resolved_path = artifact_path.resolve(strict=False)
+                    resolved_root = artifact_root.resolve(strict=False)
+                    resolved_path.relative_to(resolved_root)
+                except Exception:
+                    continue
+
+                artifact_paths.append(artifact_path)
+
+        deduped_paths: list[Path] = []
+        seen_paths: set[Path] = set()
+        for path in artifact_paths:
+            if path in seen_paths:
+                continue
+            seen_paths.add(path)
+            deduped_paths.append(path)
+        return deduped_paths
+
+    def _cleanup_session_artifacts(self, artifact_paths: list[Path]) -> None:
+        """Remove session-scoped artifact files and empty parent directories."""
+
+        artifact_root = self._session_artifact_root()
+
+        for artifact_path in artifact_paths:
+            try:
+                if artifact_path.exists():
+                    artifact_path.unlink()
+
+                sidecar_path = artifact_path.with_name("metadata_descriptor.json")
+                if sidecar_path.exists():
+                    sidecar_path.unlink()
+
+                parent = artifact_path.parent
+                resolved_root = artifact_root.resolve(strict=False)
+                while parent.exists() and parent != artifact_root:
+                    try:
+                        parent.resolve(strict=False).relative_to(resolved_root)
+                    except Exception:
+                        break
+
+                    try:
+                        parent.rmdir()
+                    except OSError:
+                        break
+                    parent = parent.parent
+            except Exception as exc:
+                logger.warning(
+                    f"Failed to clean session artifact {artifact_path}: {exc}"
+                )
+
+    def save_session(
+        self,
+        session_id: str,
+        data: Dict[str, Any],
+        user_id: str = None,
+    ) -> bool:
+        """Save a lightweight session envelope to the control plane."""
+
+        try:
+            from webapp.utilities.storage.cache_management import persistent_hash
+
[…]
