+++ webapp/utilities/storage/sharded_session_backend.py
@@ -1,803 +1,0 @@
-"""
-Sharded SQLite Session Backend for Enterprise Scale
-
-This module implements a session backend that uses the ShardedDatabaseManager
-to distribute session data across multiple SQLite databases for enterprise
-scale deployments (500+ concurrent users).
-"""
-
-import pickle
-import hashlib
-import threading
-import time
-from datetime import datetime, timezone, timedelta
-from typing import Dict, Any, Optional
-
-from webapp.config.unified import get_config
-from webapp.utilities.configuration.logging_config import get_logger
-from webapp.utilities.storage.sqlite_sharding import get_sharded_db_manager
-
-logger = get_logger()
-
-
-class ShardedSQLiteSessionBackend:
-    """
-    Enterprise-scale session backend using sharded SQLite databases.
-
-    This backend distributes session data across multiple SQLite databases
-    based on user hash, providing horizontal scaling capabilities.
-    """
-
-    def __init__(self):
-        """Initialize the sharded session backend."""
-        self.shard_manager = get_sharded_db_manager()
-
-        # Initialize database schemas for all shards
-        self._initialize_all_schemas()
-
-        # Start background cleanup
-        self._start_cleanup_thread()
-
-        # Health monitoring
-        self._last_health_check = datetime.now(timezone.utc)
-        self._health_check_interval = get_config(
-            'pool_health_check_interval', 'session', 30
-        )
-
-    def _initialize_all_schemas(self):
-        """Initialize database schemas for all shards and shared databases."""
-        # Initialize session table schema for each shard directly
-        for shard_id in range(self.shard_manager.shard_count):
-            try:
-                # Access session pool directly to ensure we initialize this specific shard
-                session_pool = self.shard_manager.session_pools[shard_id]
-                with session_pool.get_connection() as conn:
-                    self._create_session_schema(conn)
-
-                # Access cache pool directly to ensure we initialize this specific shard
-                cache_pool = self.shard_manager.cache_pools[shard_id]
-                with cache_pool.get_connection() as conn:
-                    self._create_cache_schema(conn)
-
-            except Exception as e:
-                logger.error(f"Failed to initialize shard {shard_id}: {e}")
-
-        # Initialize shared database schemas
-        try:
-            with self.shard_manager.get_analytics_connection() as conn:
-                self._create_analytics_schema(conn)
-
-            with self.shard_manager.get_health_connection() as conn:
-                self._create_health_schema(conn)
-
-        except Exception as e:
-            logger.error(f"Failed to initialize shared databases: {e}")
-
-    def _create_session_schema(
-            self,
-            conn
-    ):
-        """Create session tables and indexes for a shard."""
-        cursor = conn.cursor()
-
-        # Sessions table
-        cursor.execute("""
-            CREATE TABLE IF NOT EXISTS sessions (
-                session_id TEXT PRIMARY KEY,
-                user_id TEXT NOT NULL,
-                data BLOB NOT NULL,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                expires_at TIMESTAMP NOT NULL,
-                size_bytes INTEGER DEFAULT 0,
-                access_count INTEGER DEFAULT 0,
-                checksum TEXT
-            )
-        """)
-
-        # User queries table for quota tracking
-        cursor.execute("""
-            CREATE TABLE IF NOT EXISTS user_queries (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                user_id TEXT NOT NULL,
-                session_id TEXT,
-                query_timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                assistant_type TEXT,
[…]
