### 2026-07-27 7eedf27
Update sqlite_session_backend.py
