+++ webapp/utilities/storage/sqlite_session_backend.py
@@ -9,6 +9,7 @@ import sqlite3
 import pickle
 import threading
 import time
+import os
 from datetime import datetime, timezone, timedelta
 from pathlib import Path
 from typing import Dict, Any, Optional
@@ -19,6 +20,33 @@ from webapp.config.unified import get_config
 from webapp.utilities.configuration.logging_config import get_logger
 
 logger = get_logger()
+SLOW_SQLITE_SESSION_SAVE_MS = 25
+
+
+def _sqlite_session_save_probe_enabled() -> bool:
+    """Return whether detailed SQLite session-save timing is enabled."""
+
+    return os.getenv("DOCUSCOPE_SQLITE_SESSION_SAVE_PROBE", "").strip() == "1"
+
+
+def _log_slow_sqlite_session_save(
+    session_id: str,
+    start_time: float,
+    detail: str,
+) -> None:
+    """Log slow SQLite session-save phases with a detailed timing breakdown."""
+
+    if not _sqlite_session_save_probe_enabled():
+        return
+
+    elapsed_ms = (time.perf_counter() - start_time) * 1000
+    if elapsed_ms >= SLOW_SQLITE_SESSION_SAVE_MS:
+        logger.warning(
+            "Slow sqlite session save session={} {} duration_ms={:.2f}",
+            session_id,
+            detail,
+            elapsed_ms,
+        )
 
 
 class SQLiteConnectionPool:
@@ -191,6 +219,87 @@ class SQLiteSessionBackend:
 
             conn.commit()
 
+    def _session_artifact_root(self) -> Path:
+        """Return the root directory used for session-scoped corpus artifacts."""
+
+        return self.storage_path / "corpora"
+
+    def _collect_session_artifact_paths(self, data: Dict[str, Any]) -> list[Path]:
+        """Collect session-scoped file-backed artifact paths from persisted data."""
+
+        artifact_paths: list[Path] = []
+        artifact_root = self._session_artifact_root()
+
+        for corpus_key in ("target", "reference"):
+            corpus_state = data.get(corpus_key)
+            if not isinstance(corpus_state, dict):
+                continue
+
+            refs = corpus_state.get("_artifact_refs")
+            if not isinstance(refs, dict):
+                continue
+
+            for ref in refs.values():
+                if not isinstance(ref, dict):
+                    continue
+                if ref.get("storage_type") != "gzip_pickle":
+                    continue
+
+                path_value = ref.get("path")
+                if not path_value:
+                    continue
+
+                artifact_path = Path(path_value)
+                try:
+                    resolved_path = artifact_path.resolve(strict=False)
+                    resolved_root = artifact_root.resolve(strict=False)
+                    resolved_path.relative_to(resolved_root)
+                except Exception:
+                    continue
+
+                artifact_paths.append(artifact_path)
+
+        deduped_paths: list[Path] = []
+        seen_paths: set[Path] = set()
+        for path in artifact_paths:
+            if path in seen_paths:
+                continue
+            seen_paths.add(path)
+            deduped_paths.append(path)
+        return deduped_paths
+
+    def _cleanup_session_artifacts(self, artifact_paths: list[Path]) -> None:
+        """Remove session-scoped artifact files and empty parent directories."""
+
+        artifact_root = self._session_artifact_root()
+
+        for artifact_path in artifact_paths:
+            try:
+                if artifact_path.exists():
+                    artifact_path.unlink()
+
+                sidecar_path = artifact_path.with_name("metadata_descriptor.json")
+                if sidecar_path.exists():
+                    sidecar_path.unlink()
+
+                parent = artifact_path.parent
+                resolved_root = artifact_root.resolve(strict=False)
+                while parent.exists() and parent != artifact_root:
+                    try:
[…]
