### 2026-07-27 ea57047
Delete sqlite_sharding.py
