+++ webapp/utilities/storage/sqlite_sharding.py
@@ -1,263 +1,0 @@
-"""
-Enterprise Database Sharding for SQLite
-Handles 500+ concurrent users across multiple database shards.
-"""
-
-import hashlib
-import threading
-from typing import Dict, Any
-from pathlib import Path
-from contextlib import contextmanager
-
-from webapp.config.unified import get_config
-from webapp.utilities.configuration.logging_config import get_logger
-from webapp.utilities.storage.sqlite_session_backend import SQLiteConnectionPool
-
-logger = get_logger()
-
-
-class ShardedDatabaseManager:
-    """
-    Manages multiple SQLite database shards for enterprise scale.
-    Distributes load across multiple databases based on user hash.
-    """
-
-    def __init__(self):
-        # Check if enterprise mode is enabled (desktop_mode = false)
-        desktop_mode = get_config('desktop_mode', 'global', True)
-
-        if desktop_mode:
-            # Desktop mode: minimal sharding for compatibility
-            self.shard_count = 1
-            self.enable_sharding = False
-            logger.info("Desktop mode: using single database (no sharding)")
-        else:
-            # Enterprise mode: full sharding enabled
-            self.shard_count = get_config('shard_count', 'session', 8)
-            self.enable_sharding = True
-
-        self.shard_strategy = get_config('shard_strategy', 'session', 'user_hash')
-
-        # Connection pools for each shard
-        self.session_pools: Dict[int, SQLiteConnectionPool] = {}
-        self.cache_pools: Dict[int, SQLiteConnectionPool] = {}
-
-        # Shared databases (not sharded)
-        self.analytics_pool = None
-        self.health_pool = None
-
-        self._initialize_shards()
-        self._initialize_shared_databases()
-
-        # Shard health tracking
-        self.shard_health = {i: True for i in range(self.shard_count)}
-        self._health_lock = threading.RLock()
-
-    def _initialize_shards(self):
-        """Initialize connection pools for all shards."""
-        if not self.enable_sharding:
-            # Fallback to single database
-            self.shard_count = 1
-
-        sessions_pattern = get_config(
-            'sessions_shard_pattern',
-            'session',
-            'webapp/_session/sessions_shard_{}.db'
-        )
-        cache_pattern = get_config(
-            'cache_shard_pattern',
-            'session',
-            'webapp/_session/cache_shard_{}.db'
-        )
-
-        pool_size = get_config('connection_pool_size', 'session', 20)
-        cache_pool_size = get_config('cache_pool_size', 'session', 15)
-
-        for shard_id in range(self.shard_count):
-            # Session database pools
-            session_db_path = sessions_pattern.format(shard_id)
-            Path(session_db_path).parent.mkdir(parents=True, exist_ok=True)
-            self.session_pools[shard_id] = SQLiteConnectionPool(
-                session_db_path, pool_size
-            )
-
-            # Cache database pools
-            cache_db_path = cache_pattern.format(shard_id)
-            Path(cache_db_path).parent.mkdir(parents=True, exist_ok=True)
-            self.cache_pools[shard_id] = SQLiteConnectionPool(
-                cache_db_path, cache_pool_size
-            )
-
-    def _initialize_shared_databases(self):
-        """Initialize shared databases that are not sharded."""
-        analytics_path = get_config(
-            'analytics_db_path',
-            'session',
-            'webapp/_session/analytics.db'
-        )
-        health_path = get_config(
-            'health_db_path',
-            'session',
-            'webapp/_session/health.db'
-        )
-
-        analytics_pool_size = get_config('analytics_pool_size', 'session', 8)
-        health_pool_size = get_config('health_pool_size', 'session', 3)
-
-        self.analytics_pool = SQLiteConnectionPool(analytics_path, analytics_pool_size)
[…]
