### 2026-07-27 74459ec
Update corpus_display.py
