+++ webapp/utilities/ui/data_tables.py
@@ -7,17 +7,124 @@ dataframes, including column configurations and data transformations.
 
 import streamlit as st
 import polars as pl
-import pandas as pd
 from typing import Literal
+from time import perf_counter
 
+from webapp.utilities.configuration.logging_config import get_logger
 from webapp.utilities.ui.corpus_display import target_info
 from webapp.utilities.ui.form_controls import tag_filter_multiselect
 from webapp.utilities.exports import convert_to_excel
 
 
-def get_streamlit_column_config(
-        df: pl.DataFrame | pd.DataFrame
-        ) -> dict:
+SLOW_TABLE_OPERATION_MS = 100
+DEFAULT_TABLE_BATCH_SIZE = 500
+logger = get_logger()
+
+
+def _frame_shape(df) -> tuple[int, int]:
+    if df is None:
+        return 0, 0
+
+    height = getattr(df, "height", 0)
+    width = getattr(df, "width", 0)
+    return height, width
+
+
+def _log_table_operation(
+    operation: str,
+    start_time: float,
+    user_session_id: str | None,
+    df=None,
+) -> None:
+    elapsed_ms = (perf_counter() - start_time) * 1000
+    if elapsed_ms < SLOW_TABLE_OPERATION_MS:
+        return
+
+    height, width = _frame_shape(df)
+    logger.warning(
+        "Slow data table step op={} session={} rows={} cols={} duration_ms={:.2f}",
+        operation,
+        user_session_id or "unknown",
+        height,
+        width,
+        elapsed_ms,
+    )
+
+
+def _get_table_window_keys(
+    base_filename: str,
+    user_session_id: str | None,
+) -> tuple[str, str]:
+    """Return session-state keys for incremental table rendering."""
+
+    session_prefix = user_session_id or "global"
+    state_prefix = f"{session_prefix}_{base_filename}"
+    return (
+        f"{state_prefix}_visible_rows",
+        f"{state_prefix}_dataset_signature",
+    )
+
+
+def _table_dataset_signature(df: pl.DataFrame | None) -> tuple:
+    """Build a cheap signature to reset the visible window when data changes."""
+
+    if df is None or getattr(df, "height", 0) == 0:
+        return (0, 0, (), ())
+
+    first_row = ()
+    try:
+        first_row = tuple(df.head(1).row(0)) if df.height > 0 else ()
+    except Exception:
+        first_row = ()
+
+    return (df.height, df.width, tuple(df.columns), first_row)
+
+
+def _get_visible_table_slice(
+    df: pl.DataFrame,
+    base_filename: str,
+    user_session_id: str | None,
+    batch_size: int,
+) -> tuple[pl.DataFrame, int, bool]:
+    """Return the current visible slice and whether more rows remain."""
+
+    slice_start = perf_counter()
+
+    visible_key, signature_key = _get_table_window_keys(
+        base_filename,
+        user_session_id,
+    )
+
+    signature_start = perf_counter()
+    signature = _table_dataset_signature(df)
+    _log_table_operation(
+        "table_dataset_signature",
+        signature_start,
+        user_session_id,
+        df,
+    )
+    previous_signature = st.session_state.get(signature_key)
+
+    if previous_signature != signature:
+        st.session_state[signature_key] = signature
+        st.session_state[visible_key] = min(batch_size, df.height)
+
+    visible_rows = st.session_state.get(visible_key, min(batch_size, df.height))
+    visible_rows = min(max(int(visible_rows), batch_size), df.height)
+    st.session_state[visible_key] = visible_rows
+
+    visible_df = df.head(visible_rows)
+    has_more_rows = visible_rows < df.height
+    _log_table_operation(
+        "get_visible_table_slice",
+        slice_start,
[…]
