### 2026-07-27 020d418
Update error_boundaries.py
