### 2026-07-27 d787656
Update form_controls.py
