+++ webapp/utilities/ui/form_controls.py
@@ -9,7 +9,10 @@ import streamlit as st
 import polars as pl
 import operator
 import plotly.colors
+from collections import OrderedDict
 from functools import reduce
+from threading import Lock
+from time import perf_counter
 from typing import Dict, Optional, Callable, Tuple
 
 # Import widget key manager for centralized widget management
@@ -20,6 +23,16 @@ from webapp.utilities.state.widget_key_manager import (
 
 # Import corpus data manager for unified data access
 from webapp.utilities.corpus import get_corpus_data
+from webapp.utilities.configuration.logging_config import get_logger
+
+SLOW_FORM_CONTROL_MS = 100
+SIMPLIFIED_FRAME_CACHE_MAX_ITEMS = 16
+TAGSET_FRAME_CACHE_MAX_ITEMS = 64
+_simplified_frame_cache: OrderedDict[tuple[int, str], pl.DataFrame] = OrderedDict()
+_simplified_frame_cache_locks: dict[tuple[int, str], Lock] = {}
+_simplified_frame_cache_locks_guard = Lock()
+_tagset_frame_cache: OrderedDict[tuple, pl.DataFrame] = OrderedDict()
+logger = get_logger()
 
 # Register persistent widgets used across form controls
 # These widgets should persist across page loads
@@ -34,6 +47,141 @@ PERSISTENT_FORM_WIDGETS = [
 register_persistent_widgets(PERSISTENT_FORM_WIDGETS)
 
 
+def _frame_shape(df) -> tuple[int, int]:
+    if df is None:
+        return 0, 0
+
+    return getattr(df, "height", 0), getattr(df, "width", 0)
+
+
+def _log_slow_form_control(
+    operation: str,
+    start_time: float,
+    user_session_id: str | None,
+    df=None,
+) -> None:
+    elapsed_ms = (perf_counter() - start_time) * 1000
+    if elapsed_ms < SLOW_FORM_CONTROL_MS:
+        return
+
+    height, width = _frame_shape(df)
+    logger.warning(
+        "Slow form control step op={} session={} rows={} cols={} duration_ms={:.2f}",
+        operation,
+        user_session_id or "unknown",
+        height,
+        width,
+        elapsed_ms,
+    )
+
+
+def _get_cached_simplified_frame(cache_key: tuple[int, str]) -> pl.DataFrame | None:
+    cached = _simplified_frame_cache.get(cache_key)
+    if cached is None:
+        return None
+
+    _simplified_frame_cache.move_to_end(cache_key)
+    return cached
+
+
+def _set_cached_simplified_frame(
+    cache_key: tuple[int, str],
+    df: pl.DataFrame,
+) -> None:
+    _simplified_frame_cache[cache_key] = df
+    _simplified_frame_cache.move_to_end(cache_key)
+
+    while len(_simplified_frame_cache) > SIMPLIFIED_FRAME_CACHE_MAX_ITEMS:
+        _simplified_frame_cache.popitem(last=False)
+
+
+def _get_simplified_frame_lock(cache_key: tuple[int, str]) -> Lock:
+    with _simplified_frame_cache_locks_guard:
+        lock = _simplified_frame_cache_locks.get(cache_key)
+        if lock is None:
+            lock = Lock()
+            _simplified_frame_cache_locks[cache_key] = lock
+        return lock
+
+
+def _get_tagset_frame_signature(user_session_id: str, session_key: str) -> tuple:
+    session_container = st.session_state.get(user_session_id, {})
+    target_data = session_container.get("target", {})
+    if not isinstance(target_data, dict):
+        return (user_session_id, session_key, "missing")
+
+    refs = target_data.get("_artifact_refs")
+    ref_key = "ft_pos" if session_key == "ft_pos_general" else session_key
[…]
