### 2026-07-27 0b262d9
Update sidebar.py
