### 2026-07-27 5e858d3
Update text_display.py
