+++ webapp/pages/1_load_corpus.py
@@ -3,18 +3,22 @@ App for loading and managing corpora in the Corpus Tagger web application.
 
 This module provides functionality for:
 - Loading existing corpora from internal databases or user uploads.
-- Processing new corpora from text files or tabular uploads.
+- Processing new corpora from text files.
 - Uploading external corpora in Parquet format.
 - Managing corpus metadata and categories.
 - Resetting corpus data and clearing session state.
 """
 
+import os
+
 import spacy
 import streamlit as st
+from time import perf_counter
 
 # Core application utilities with standardized patterns
 from webapp.utilities.core import app_core
 from webapp.utilities.configuration.config_manager import config_manager
+from webapp.utilities.configuration.logging_config import get_logger
 
 # Module-specific imports
 from webapp.utilities.session import (
@@ -28,9 +32,16 @@ from webapp.utilities.analysis import (
     )
 from webapp.utilities.processing import (
     process_external, process_internal,
-    process_new, handle_uploaded_parquet, handle_uploaded_tabular,
+    process_new, handle_uploaded_parquet,
+    attach_queued_internal_target,
     handle_uploaded_text, sidebar_process_section
     )
+from webapp.persistence import registry_service
+from webapp.queue import (
+    enqueue_internal_target_preparation,
+    get_queue,
+    get_redis_queue_config,
+)
 from webapp.utilities.ui import (
     load_and_display_target_corpus, render_corpus_info_expanders
 )
@@ -38,6 +49,7 @@ from webapp.menu import (
     menu, require_login
     )
 from webapp.utilities.state import (
+    CorpusPersistencePolicy,
     CorpusKeys, LoadCorpusKeys,
     MetadataKeys, SessionKeys,
     WarningKeys
@@ -47,6 +59,8 @@ from webapp.utilities.state import (
 LOAD_CORPUS_PERSISTENT_WIDGETS = [
     "reffiles",           # File uploader for reference files
     "corpora_to_load",    # Selection for corpora to load
+    "target_persistence_policy",
+    "reference_persistence_policy",
 ]
 app_core.register_page_widgets(LOAD_CORPUS_PERSISTENT_WIDGETS)
 
@@ -64,9 +78,17 @@ MAX_POLARS = config_manager.max_polars_size
 
 TITLE = "Manage Corpus Data"
 ICON = ":material/database:"
+SLOW_LOAD_CORPUS_PAGE_MS = 100
+logger = get_logger()
+TEST_MODE = config_manager.test_mode
+PROCESS_TARGET_PROBE_STAGE_KEY = "load_test_process_target_probe_stage"
+PROCESS_TARGET_PROBE_TIMINGS_KEY = "load_test_process_target_probe_stage_timings_ms"
+PROCESS_TARGET_PROBE_FLAT_TIMINGS_KEY = "load_test_process_target_probe_stage_timings_flat"
+PROCESS_TARGET_PROBE_SUMMARY_KEY = "load_test_process_target_probe_stage_summary"
+INTERNAL_TARGET_QUEUE_STATE_KEY = "internal_target_queue_state"
 
 # Define labels and options for the app
-CORPUS_SOURCES = ["Internal", "External", "New", "Tabular"]
+CORPUS_SOURCES = ["Internal", "External", "New"]
 
 # Button and form labels
 LABEL_PROCESS_TARGET = "Process Target"
@@ -79,6 +101,147 @@ LABEL_RESET_CORPUS = "Reset Corpus"
 MODEL_LARGE_LABEL = "Large Dictionary"
 MODEL_SMALL_LABEL = "Common Dictionary"
[…]
