### 2026-07-27 710b686
Update 2_token_frequencies.py
