### 2026-07-27 f791548
Update 3_tag_frequencies.py
