+++ webapp/pages/4_ngrams.py
@@ -25,8 +25,10 @@ from webapp.utilities.ui import (
 )
 from webapp.utilities.analysis import (
     has_target_corpus, render_corpus_not_loaded_error,
-    generate_clusters, generate_ngrams
+    generate_clusters, generate_ngrams, attach_ngram_artifact
 )
+from webapp.persistence import registry_service
+from webapp.queue import enqueue_ngram_preparation, get_queue, get_redis_queue_config
 from webapp.menu import (
     menu, require_login
     )
@@ -53,11 +55,167 @@ SEARCH_MODE_CONFIG = {
     "Contains": "contains"
 }
 SPAN_OPTIONS = (2, 3, 4)
+NGRAM_QUEUE_STATE_KEY = "ngram_queue_state"
 
 st.set_page_config(
     page_title=TITLE, page_icon=ICON,
     layout="wide"
     )
+
+
+def _get_ngram_queue_state(user_session_id: str) -> dict | None:
+    """Return queued n-gram/cluster state for the current session, if any."""
+
+    queue_state = st.session_state.get(NGRAM_QUEUE_STATE_KEY)
+    if not isinstance(queue_state, dict):
+        return None
+    if queue_state.get("session_id") != user_session_id:
+        return None
+    return queue_state
+
+
+def _clear_ngram_queue_state(user_session_id: str) -> None:
+    """Clear queued n-gram/cluster state for this session."""
+
+    if _get_ngram_queue_state(user_session_id) is not None:
+        st.session_state.pop(NGRAM_QUEUE_STATE_KEY, None)
+
+
+def _can_queue_ngrams(session: dict) -> str | None:
+    """Return built-in target path when n-gram/cluster generation can be queued."""
+
+    if not get_redis_queue_config().enabled:
+        return None
+
+    target_source = safe_session_get(session, SessionKeys.TARGET_DB, "")
+    if not target_source:
+        return None
+    return str(target_source)
+
+
+def _submit_ngram_job(
+    user_session_id: str,
+    target_source: str,
+    analysis_type: str,
+    ngram_span: int,
+    count_by: str,
+    from_anchor: str | None = None,
+    node_word: str | None = None,
+    tag: str | None = None,
+    position: int | None = None,
+    search_type: str | None = None,
+) -> None:
+    """Submit built-in n-gram/cluster generation through Redis/RQ."""
+
+    result = enqueue_ngram_preparation(
+        target_source=target_source,
+        analysis_type=analysis_type,
+        ngram_span=ngram_span,
+        count_by=count_by,
+        from_anchor=from_anchor,
+        node_word=node_word,
+        tag=tag,
+        position=position,
+        search_type=search_type,
+    )
+    if result.state == "ready" and result.artifact_id is not None:
+        _clear_ngram_queue_state(user_session_id)
+        if attach_ngram_artifact(user_session_id, result.artifact_id):
+            st.rerun()
+        st.error("Ready n-gram artifact could not be attached. Please retry.")
+        return
+
+    st.session_state[NGRAM_QUEUE_STATE_KEY] = {
+        "session_id": user_session_id,
+        "analysis_type": analysis_type,
+        "control_plane_job_id": result.control_plane_job_id,
+        "rq_job_id": result.rq_job_id,
+        "artifact_id": result.artifact_id,
+    }
+    st.rerun()
+
+
+@st.fragment(run_every="2s")
+def _render_ngram_queue_status(user_session_id: str) -> None:
+    """Poll queued n-gram/cluster status and attach it once complete."""
+
+    queue_state = _get_ngram_queue_state(user_session_id)
+    if queue_state is None:
+        return
+
[…]
