+++ webapp/pages/5_compare_corpora.py
@@ -39,8 +39,10 @@ from webapp.utilities.state.widget_key_manager import create_persist_function
 from webapp.utilities.analysis import (
     has_target_corpus, has_reference_corpus,
     render_corpus_not_loaded_error, generate_keyness_tables,
-    freq_simplify_pl
+    freq_simplify_pl, attach_keyness_artifact
 )
+from webapp.persistence import registry_service
+from webapp.queue import enqueue_keyness_preparation, get_queue, get_redis_queue_config
 from webapp.utilities.plotting import (
     plot_compare_corpus_bar, plot_download_link
 )
@@ -60,6 +62,7 @@ COMPARE_CORPORA_PERSISTENT_WIDGETS = [
 app_core.register_page_widgets(COMPARE_CORPORA_PERSISTENT_WIDGETS)
 
 TOKEN_LIMIT = 1_500_000
+KEYNESS_QUEUE_STATE_KEY = "keyness_queue_state"
 
 # Configuration constants
 KEYNESS_TOKEN_TAGSET_CONFIG = {
@@ -82,6 +85,154 @@ st.set_page_config(
     page_title=TITLE, page_icon=ICON,
     layout="wide"
     )
+
+
+def _get_keyness_queue_state(user_session_id: str) -> dict | None:
+    """Return queued keyness state for the current session, if any."""
+
+    queue_state = st.session_state.get(KEYNESS_QUEUE_STATE_KEY)
+    if not isinstance(queue_state, dict):
+        return None
+    if queue_state.get("session_id") != user_session_id:
+        return None
+    return queue_state
+
+
+def _clear_keyness_queue_state(user_session_id: str) -> None:
+    """Clear queued keyness state for the current session."""
+
+    if _get_keyness_queue_state(user_session_id) is not None:
+        st.session_state.pop(KEYNESS_QUEUE_STATE_KEY, None)
+
+
+def _can_queue_keyness(session: dict) -> tuple[str, str] | None:
+    """Return built-in target/reference paths when keyness can be queued."""
+
+    if not get_redis_queue_config().enabled:
+        return None
+
+    target_source = safe_session_get(session, SessionKeys.TARGET_DB, "")
+    reference_source = safe_session_get(session, SessionKeys.REFERENCE_DB, "")
+    if not target_source or not reference_source:
+        return None
+    return str(target_source), str(reference_source)
+
+
+def _submit_keyness_job(
+    user_session_id: str,
+    target_source: str,
+    reference_source: str,
+    threshold: float,
+    swap_target: bool,
+) -> None:
+    """Submit built-in keyness generation through Redis/RQ."""
+
+    result = enqueue_keyness_preparation(
+        target_source=target_source,
+        reference_source=reference_source,
+        threshold=threshold,
+        swap_target=swap_target,
+    )
+    if result.state == "ready" and result.artifact_id is not None:
+        _clear_keyness_queue_state(user_session_id)
+        if attach_keyness_artifact(user_session_id, result.artifact_id):
+            st.rerun()
+        st.error("Ready keyness artifact could not be attached. Please retry.")
+        return
+
+    st.session_state[KEYNESS_QUEUE_STATE_KEY] = {
+        "session_id": user_session_id,
+        "target_source": target_source,
+        "reference_source": reference_source,
+        "threshold": threshold,
+        "swap_target": swap_target,
+        "control_plane_job_id": result.control_plane_job_id,
+        "rq_job_id": result.rq_job_id,
+        "artifact_id": result.artifact_id,
+    }
+    st.rerun()
+
+
+@st.fragment(run_every="2s")
+def _render_keyness_queue_status(user_session_id: str) -> None:
[…]
