+++ webapp/pages/6_compare_corpus_parts.py
@@ -47,8 +47,10 @@ from webapp.utilities.state import (
 from webapp.utilities.state.widget_key_manager import create_persist_function
 from webapp.utilities.analysis import (
     has_target_corpus, render_corpus_not_loaded_error,
-    freq_simplify_pl, generate_keyness_parts
+    freq_simplify_pl, generate_keyness_parts, attach_keyness_parts_artifact
 )
+from webapp.persistence import registry_service
+from webapp.queue import enqueue_keyness_parts_preparation, get_queue, get_redis_queue_config
 from webapp.menu import (   # noqa: E402
     menu,
     require_login
@@ -67,10 +69,168 @@ COMPARE_CORPUS_PARTS_PERSISTENT_WIDGETS = [
 ]
 app_core.register_page_widgets(COMPARE_CORPUS_PARTS_PERSISTENT_WIDGETS)
 
+KEYNESS_PARTS_QUEUE_STATE_KEY = "keyness_parts_queue_state"
+
 st.set_page_config(
     page_title=TITLE, page_icon=ICON,
     layout="wide"
     )
+
+
+def _normalize_category_selection(categories: list) -> list[str]:
+    """Return a stable category list for queue/artifact identity matching."""
+
+    return sorted(str(category) for category in categories)
+
+
+def _get_keyness_parts_queue_state(user_session_id: str) -> dict | None:
+    """Return queued corpus-parts keyness state for this session, if any."""
+
+    queue_state = st.session_state.get(KEYNESS_PARTS_QUEUE_STATE_KEY)
+    if not isinstance(queue_state, dict):
+        return None
+    if queue_state.get("session_id") != user_session_id:
+        return None
+    return queue_state
+
+
+def _clear_keyness_parts_queue_state(user_session_id: str) -> None:
+    """Clear queued corpus-parts keyness state for this session."""
+
+    if _get_keyness_parts_queue_state(user_session_id) is not None:
+        st.session_state.pop(KEYNESS_PARTS_QUEUE_STATE_KEY, None)
+
+
+def _can_queue_keyness_parts(session: dict) -> str | None:
+    """Return built-in target path when corpus-parts keyness can be queued."""
+
+    if not get_redis_queue_config().enabled:
+        return None
+
+    target_source = safe_session_get(session, SessionKeys.TARGET_DB, "")
+    if not target_source:
+        return None
+    return str(target_source)
+
+
+def _submit_keyness_parts_job(
+    user_session_id: str,
+    target_source: str,
+    target_categories: list[str],
+    reference_categories: list[str],
+    threshold: float,
+    swap_target: bool,
+) -> None:
+    """Submit built-in corpus-parts keyness generation through Redis/RQ."""
+
+    result = enqueue_keyness_parts_preparation(
+        target_source=target_source,
+        target_categories=target_categories,
+        reference_categories=reference_categories,
+        threshold=threshold,
+        swap_target=swap_target,
+    )
+    if result.state == "ready" and result.artifact_id is not None:
+        _clear_keyness_parts_queue_state(user_session_id)
+        if attach_keyness_parts_artifact(user_session_id, result.artifact_id):
+            st.rerun()
+        st.error("Ready corpus-parts artifact could not be attached. Please retry.")
+        return
+
+    st.session_state[KEYNESS_PARTS_QUEUE_STATE_KEY] = {
+        "session_id": user_session_id,
+        "target_source": target_source,
+        "target_categories": target_categories,
+        "reference_categories": reference_categories,
+        "threshold": threshold,
+        "swap_target": swap_target,
+        "control_plane_job_id": result.control_plane_job_id,
+        "rq_job_id": result.rq_job_id,
[…]
