### 2026-07-27 670b3d6
Update 7_collocations.py
