+++ webapp/pages/7_collocations.py
@@ -34,8 +34,10 @@ from webapp.utilities.analysis import (
     has_target_corpus, render_corpus_not_loaded_error
 )
 from webapp.utilities.analysis import (
-    generate_collocations
+    attach_collocation_artifact, generate_collocations
 )
+from webapp.persistence import registry_service
+from webapp.queue import enqueue_collocation_preparation, get_queue, get_redis_queue_config
 from webapp.utilities.state import (
     safe_clear_widget_state,
     SessionKeys, CorpusKeys,
@@ -48,10 +50,181 @@ from webapp.menu import (
 
 TITLE = "Collocates"
 ICON = ":material/network_node:"
+COLLOCATION_QUEUE_STATE_KEY = "collocation_queue_state"
 
 st.set_page_config(
     page_title=TITLE, page_icon=ICON,
     layout="wide"
+    )
+
+
+def _collocation_parameters(
+    node_word: str,
+    node_tag: str | None,
+    to_left: int,
+    to_right: int,
+    stat_mode: str,
+    count_by: str,
+) -> dict:
+    """Return metadata parameters for a collocation table."""
+
+    return {
+        "node_word": node_word,
+        "node_tag": node_tag,
+        "to_left": to_left,
+        "to_right": to_right,
+        "stat_mode": stat_mode,
+        "count_by": count_by,
+    }
+
+
+def _get_collocation_queue_state(user_session_id: str) -> dict | None:
+    """Return queued collocation state for the current session, if any."""
+
+    queue_state = st.session_state.get(COLLOCATION_QUEUE_STATE_KEY)
+    if not isinstance(queue_state, dict):
+        return None
+    if queue_state.get("session_id") != user_session_id:
+        return None
+    return queue_state
+
+
+def _clear_collocation_queue_state(user_session_id: str) -> None:
+    """Clear queued collocation state for the current session."""
+
+    if _get_collocation_queue_state(user_session_id) is not None:
+        st.session_state.pop(COLLOCATION_QUEUE_STATE_KEY, None)
+
+
+def _can_queue_collocations(session: dict) -> str | None:
+    """Return built-in target path when collocations can be queued."""
+
+    if not get_redis_queue_config().enabled:
+        return None
+
+    target_source = safe_session_get(session, SessionKeys.TARGET_DB, "")
+    if not target_source:
+        return None
+    return str(target_source)
+
+
+def _submit_collocation_job(
+    user_session_id: str,
+    target_source: str,
+    node_word: str,
+    node_tag: str | None,
+    to_left: int,
+    to_right: int,
+    stat_mode: str,
+    count_by: str,
+) -> None:
+    """Submit built-in collocation generation through Redis/RQ."""
+
+    colloc_params = _collocation_parameters(
+        node_word,
+        node_tag,
+        to_left,
+        to_right,
+        stat_mode,
+        count_by,
+    )
+    result = enqueue_collocation_preparation(
+        target_source=target_source,
+        node_word=node_word,
+        node_tag=node_tag,
+        to_left=to_left,
+        to_right=to_right,
+        stat_mode=stat_mode,
+        count_by=count_by,
+    )
+    if result.state == "ready" and result.artifact_id is not None:
+        _clear_collocation_queue_state(user_session_id)
+        if attach_collocation_artifact(user_session_id, result.artifact_id, colloc_params):
+            st.rerun()
+        st.error("Ready collocation artifact could not be attached. Please retry.")
+        return
+
+    st.session_state[COLLOCATION_QUEUE_STATE_KEY] = {
+        "session_id": user_session_id,
+        "target_source": target_source,
+        "node_word": node_word,
+        "node_tag": node_tag,
+        "to_left": to_left,
[…]
