### 2026-07-27 dfac683
Update 11_assisted_plotting.py
