+++ webapp/pages/11_assisted_plotting.py
@@ -3,9 +3,7 @@ This app provides an interface for AI-assisted plotting
 from a loaded target corpus. Users can interact with Plotbot to create
 and refine plots based on their data.
 
-Users can:
-- Select a plotting library (Plotly, Matplotlib, Seaborn)
-- Interact with Plotbot to generate and refine plots
+Users can interact with Plotbot to generate and refine Plotly Express plots.
 """
 
 import base64
@@ -30,7 +28,15 @@ from webapp.utilities.ai import (
     render_data_selection_interface, render_data_preview_controls,
     render_quota_tracker, should_show_api_key_input,
     render_work_preservation_interface, should_show_work_preservation_interface,
-    export_conversation_history, clear_plotbot_table
+    export_conversation_history, clear_plotbot_table,
+    prune_message_thread
+)
+from webapp.utilities.ai.providers import get_openai_compatible_provider_config
+from webapp.persistence import registry_service
+from webapp.queue import (
+    enqueue_plotbot_generation,
+    get_queue,
+    get_redis_queue_config,
 )
 from webapp.utilities.analysis import (
     generate_tags_table
@@ -62,35 +68,185 @@ LLM_MODEL = AI_CONFIG['model']
 LLM_PARAMS = AI_CONFIG['parameters']
 QUOTA = AI_CONFIG['quota']
 
-# Register persistent widgets for this page
-app_core.register_page_widgets(["plot_radio"])
+PLOTBOT_QUEUE_STATE_KEY = "plotbot_queue_state"
+PLOTBOT_LIBRARY = "plotly.express"
 
 
-def render_plotting_library_selection(user_session_id: str) -> str:
-    """Render the plotting library selection interface."""
-    st.markdown(
-        body="### Plotting Library",
-        help=(
-            "To create plots, I can use the following libraries:\n\n"
-            "* [Plotly express](https://plotly.com/python/plotly-express/)\n"
-            "* [Matplotlib](https://matplotlib.org/)\n"
-            "* [Seaborn](https://seaborn.pydata.org/)\n\n"
-            "Each library has its own aesthetics and features. "
-            "If you're unfamiliar with them, you should check out their "
-            "documentation, as well as examples of their use."
+def _get_plotbot_queue_state(user_session_id: str) -> dict | None:
+    """Return pending Plotbot queue state for this Streamlit session."""
+    queue_state = st.session_state.get(PLOTBOT_QUEUE_STATE_KEY)
+    if not isinstance(queue_state, dict):
+        return None
+    if queue_state.get("session_id") != user_session_id:
+        return None
+    return queue_state
+
+
+def _clear_plotbot_queue_state(user_session_id: str) -> None:
+    """Clear pending Plotbot queue state for this Streamlit session."""
+    if _get_plotbot_queue_state(user_session_id) is not None:
+        st.session_state.pop(PLOTBOT_QUEUE_STATE_KEY, None)
+
+
+def _has_local_plotbot_provider() -> bool:
+    """Return True when Plotbot can use a configured local model provider."""
+    return get_openai_compatible_provider_config() is not None
+
+
+def _dataframe_to_plotbot_records(df) -> list[dict[str, object]]:
+    """Convert a selected Plotbot dataframe to JSON-friendly row records."""
+    if hasattr(df, "to_pandas"):
+        df = df.to_pandas()
+    return json.loads(df.to_json(orient="records", date_format="iso"))
+
+
+def _append_queued_plotbot_result(user_session_id: str, artifact_id: int) -> bool:
+    """Attach a completed queued Plotbot JSON result to the chat state."""
+    artifact = registry_service.get_artifact_by_id(artifact_id)
+    if artifact is None:
+        return False
+
[…]
