+++ webapp/pages/12_assisted_analysis.py
@@ -1,37 +1,21 @@
 """
-This app provides an interface for AI-assisted analysis of tabular data
-using Pandabot, a chat assistant that can answer questions and perform
-data analysis tasks.
+This app provides an interface for deterministic DFM statistics.
 """
 
-import json
 import streamlit as st
-from datetime import datetime, timezone
-
-# Core application utilities with standardized patterns
-from webapp.config.unified import get_ai_config
-from webapp.config.config_utils import get_runtime_setting
-
-# UI error boundaries (imported directly to avoid None fallback)
-from webapp.utilities.ui.error_boundaries import SafeComponentRenderer
+import docuscospacy as ds
 
 # Module-specific imports
 from webapp.utilities.session import (
     get_or_init_user_session, safe_session_get, load_metadata
 )
-from webapp.utilities.ai import (
-    clear_pandasai, pandabot_user_query,
-    setup_ai_session_state, get_api_key,
-    render_api_key_input, render_data_selection_interface,
-    render_data_preview_controls, render_quota_tracker,
-    should_show_api_key_input, render_work_preservation_interface,
-    should_show_work_preservation_interface, export_conversation_history,
-    clear_pandasai_table
-)
 from webapp.utilities.analysis import (
-    generate_tags_table
+    available_dfm_features,
+    compute_dfm_statistics,
+    generate_tags_table,
+    get_dfm_statistic_options,
 )
-from webapp.utilities.core import app_core
+from webapp.utilities.corpus import get_corpus_data_manager
 from webapp.utilities.ui import (
     sidebar_help_link, render_table_generation_interface
 )
@@ -42,244 +26,276 @@ from webapp.menu import (
     menu, require_login
 )
 
-TITLE = "AI-Assisted Analysis"
-ICON = ":material/smart_toy:"
+TITLE = "Matrix Explorer"
+ICON = ":material/table_view:"
 
 st.set_page_config(
     page_title=TITLE, page_icon=ICON,
     layout="wide"
 )
 
-# Get AI configuration using standardized access
-AI_CONFIG = get_ai_config()
-DESKTOP = AI_CONFIG['desktop_mode']
-CACHE = AI_CONFIG['cache_enabled']
-LLM_MODEL = AI_CONFIG['model']
-LLM_PARAMS = AI_CONFIG['parameters']
-QUOTA = AI_CONFIG['quota']
+DFM_SOURCES = {
+    "DocuScope DFM": {
+        "key": "dtm_ds",
+        "tagset": "DocuScope",
+        "transform": None,
+    },
+    "Parts-of-Speech DFM": {
+        "key": "dtm_pos",
+        "tagset": "Parts-of-Speech",
+        "transform": None,
+    },
+    "Parts-of-Speech DFM (general)": {
+        "key": "dtm_pos",
+        "tagset": "Parts-of-Speech",
+        "transform": ds.dtm_simplify,
+    },
+}
 
 
-def render_pandabot_chat_interface(
-    user_session_id: str,
-    api_key: str,
-    df,
-    selected_query: str
-) -> None:
-    """Render the chat interface for Pandabot."""
-    # Display chat history
-    for message in st.session_state[user_session_id]["pandasai"]:
-        with st.chat_message(message["role"]):
-            if message["type"] == "string":
-                st.markdown(message["value"])
-            elif message["type"] == "code":
-                st.code(message["value"], language="python")
-            elif message["type"] == "error":
-                st.error(message["value"], icon=":material/error:")
-            elif message["type"] == "plot":
-                # Display plot image with error boundary
-                SafeComponentRenderer.safe_image(
-                    message["value"], "Generated plot unavailable"
-                )
-            elif message["type"] == "dataframe":
-                st.dataframe(
[…]
