### 2026-07-27 1671146
Update 9_advanced_plotting.py
