### 2026-07-27 a301e96
Update menu.py
