+++ webapp/menu.py
@@ -11,24 +11,126 @@ Both timeouts are configurable via options.toml and include warning systems.
 """
 
 import base64
+import os
 import time
+from time import perf_counter
 
 import streamlit as st
 
 from webapp.utilities.storage import add_login
 from webapp.config.unified import get_config
 from webapp.utilities.core import safe_config_value
+from webapp.utilities.configuration.logging_config import get_logger, setup_debug_logging
 from webapp.utilities.session import get_or_init_user_session
 from webapp.utilities.auth import is_user_authorized
 from webapp.config.config_utils import get_runtime_setting
 
 GOOGLE_LOGO = get_config('google_logo_path', 'global', 'webapp/_static/web_light_rd_na.svg')
 DESKTOP = get_config('desktop_mode', 'global')
+TEST_MODE = get_config('test_mode', 'global', False)
+SLOW_MENU_STEP_MS = 25
+setup_debug_logging("navigation_probe")
+logger = get_logger()
+
+CORE_PAGE_LINKS = [
+    ("index.py", "Main Page", ":material/home:"),
+    ("pages/1_load_corpus.py", "Manage Corpus Data", ":material/database:"),
+    ("pages/2_token_frequencies.py", "Token Frequencies", ":material/table_view:"),
+    ("pages/3_tag_frequencies.py", "Tag Frequencies", ":material/table_view:"),
+    ("pages/4_ngrams.py", "Ngrams & Clusters", ":material/table_view:"),
+    ("pages/5_compare_corpora.py", "Compare Corpora", ":material/compare_arrows:"),
+    ("pages/6_compare_corpus_parts.py", "Compare Corpus Parts", ":material/compare_arrows:"),
+    ("pages/7_collocations.py", "Collocations", ":material/network_node:"),
+    ("pages/8_kwic.py", "Key Words in Context", ":material/network_node:"),
+    ("pages/9_advanced_plotting.py", "Advanced Plotting", ":material/line_axis:"),
+    ("pages/12_assisted_analysis.py", "Matrix Explorer", ":material/line_axis:"),
+    ("pages/10_single_document.py", "Single Document", ":material/find_in_page:"),
+    ("pages/11_assisted_plotting.py", "AI-Asissted Plotting", ":material/smart_toy:"),
+    ("pages/13_download_corpus.py", "Download Corpus Data", ":material/download:"),
+    ("pages/14_download_tagged_files.py", "Download Tagged Files", ":material/download:"),
+]
+
+ADMIN_PAGE_LINKS = [
+    ("pages/98_user_management.py", "User Management", ":material/admin_panel_settings:"),
+    ("pages/99_health_monitor.py", "Health Monitor", ":material/cardiology:"),
+]
+
+MINIMAL_PAGE_LINKS = CORE_PAGE_LINKS[:3]
+
+
+def _load_test_minimal_menu_enabled() -> bool:
+    """Return whether load-test-only minimal navigation should be used."""
+
+    return os.environ.get("DOCUSCOPE_LOAD_TEST_MINIMAL_MENU", "").strip() == "1"
+
+
+def navigation_experiment_enabled() -> bool:
+    """Legacy compatibility shim for removed navigation experiment."""
+
+    return False
+
+
+def _log_slow_menu_step(session_id: str, step: str, start_time: float) -> None:
+    """Log slow menu phases during page entry reruns."""
+
+    elapsed_ms = (perf_counter() - start_time) * 1000
+    if elapsed_ms >= SLOW_MENU_STEP_MS:
+        logger.warning(
+            "Slow menu step session={} step={} duration_ms={:.2f}",
+            session_id,
+            step,
+            elapsed_ms,
+        )
+
+
+def _render_page_links(page_links: list[tuple[str, str, str]]) -> None:
[…]
