+++ docker-compose.yml
@@ -1,14 +1,133 @@
 # Docker Compose v2 format (implicit version). The top-level `version` key is deprecated in v2 and intentionally omitted.
 # Tip: For local development with hot-reloading, mount the project and override the command (see commented lines below).
 services:
+  postgres:
+    image: postgres:16-alpine
+    environment:
+      POSTGRES_DB: docuscope_ca
+      POSTGRES_USER: docuscope
+      POSTGRES_PASSWORD: docuscope
+    ports:
+      - "5432:5432"
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+    healthcheck:
+      test: ["CMD-SHELL", "pg_isready -U docuscope -d docuscope_ca"]
+      interval: 10s
+      timeout: 5s
+      retries: 5
+
+  redis:
+    image: redis:7-alpine
+    ports:
+      - "6379:6379"
+    healthcheck:
+      test: ["CMD", "redis-cli", "ping"]
+      interval: 10s
+      timeout: 5s
+      retries: 5
+
   streamlit_app:
     build:
       context: .
       dockerfile: Dockerfile
+    depends_on:
+      postgres:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    environment:
+      DATABASE_URL: postgresql+psycopg://docuscope:docuscope@postgres:5432/docuscope_ca
+      REDIS_URL: redis://redis:6379/0
+      DOCUSCOPE_RQ_ENABLED: "1"
+      DOCUSCOPE_RQ_SIMPLE_WORKER: "1"
+      DOCUSCOPE_AI_PROVIDER: ${DOCUSCOPE_AI_PROVIDER:-}
+      DOCUSCOPE_AI_BASE_URL: ${DOCUSCOPE_AI_BASE_URL:-}
+      DOCUSCOPE_AI_MODEL: ${DOCUSCOPE_AI_MODEL:-}
     ports:
       - "8501:8501"
+    volumes:
+      - artifact_data:/app/webapp/_artifacts
+    command:
+      - streamlit
+      - run
+      - webapp/index.py
+      - --server.port
+      - "8501"
+      - --server.address
+      - 0.0.0.0
+      - --server.fileWatcherType
+      - none
     # For development with hot-reloading, uncomment the next two lines:
     # volumes:
     #   - ./:/app
     # command overrides Dockerfile CMD for dev; for prod, let Dockerfile CMD run
     # command: streamlit run webapp/index.py --server.port 8501 --server.address 0.0.0.0
+
+  rq_worker:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    depends_on:
+      postgres:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    environment:
+      DATABASE_URL: postgresql+psycopg://docuscope:docuscope@postgres:5432/docuscope_ca
+      REDIS_URL: redis://redis:6379/0
+      DOCUSCOPE_RQ_ENABLED: "1"
+      DOCUSCOPE_AI_PROVIDER: ${DOCUSCOPE_AI_PROVIDER:-}
+      DOCUSCOPE_AI_BASE_URL: ${DOCUSCOPE_AI_BASE_URL:-}
+      DOCUSCOPE_AI_MODEL: ${DOCUSCOPE_AI_MODEL:-}
+    volumes:
+      - artifact_data:/app/webapp/_artifacts
+    command: python -m webapp.queue.worker
+    healthcheck:
[…]
