+++ Dockerfile
@@ -64,5 +64,6 @@ HEALTHCHECK CMD curl --fail http://localhost:8501/_stcore/health || exit 1
 ENV STREAMLIT_SERVER_PORT=8501
 ENV STREAMLIT_SERVER_ADDRESS=0.0.0.0
 
-# Run the application
-CMD ["streamlit", "run", "webapp/index.py", "--server.port", "8501", "--server.address", "0.0.0.0"]
+# Run the application from the webapp entrypoint so multipage routing resolves
+# the bundled page files under webapp/pages.
+CMD ["streamlit", "run", "webapp/index.py", "--server.port", "8501", "--server.address", "0.0.0.0", "--server.fileWatcherType", "none"]
