+++ requirements.txt
@@ -1,24 +1,25 @@
-Authlib==1.5.2 ; python_version >= "3.11" and python_version < "3.13"
-docuscospacy==0.3.6 ; python_version >= "3.11" and python_version < "3.13"
-google-cloud-firestore==2.20.2 ; python_version >= "3.11" and python_version < "3.13"
-lingua-language-detector==2.0.2 ; python_version >= "3.11" and python_version < "3.13"
-loguru==0.7.3 ; python_version >= "3.11" and python_version < "3.13"
-kaleido==0.2.1 ; python_version >= "3.11" and python_version < "3.13"
-matplotlib==3.7.5 ; python_version >= "3.11" and python_version < "3.13"
-openai==1.66.5 ; python_version >= "3.11" and python_version < "3.13"
-numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.13"
-pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.13"
-pandasai==3.0.0b18 ; python_version >= "3.11" and python_version < "3.13"
-pandasai-openai==0.1.5 ; python_version >= "3.11" and python_version < "3.13"
-plotly==6.0.1 ; python_version >= "3.11" and python_version < "3.13"
-polars==1.31.0 ; python_version >= "3.11" and python_version < "3.13"
-psutil==7.0.0 ; python_version >= "3.11" and python_version < "3.13"
-python-docx==1.1.2 ; python_version >= "3.11" and python_version < "3.13"
-RestrictedPython==8.0.0 ; python_version >= "3.11" and python_version < "3.13"
-scikit-learn==1.6.1 ; python_version >= "3.11" and python_version < "3.13"
-scipy==1.10.1 ; python_version >= "3.11" and python_version < "3.13"
-seaborn==0.12.2 ; python_version >= "3.11" and python_version < "3.13"
-spacy==3.8.7 ; python_version >= "3.11" and python_version < "3.13"
-streamlit==1.45.1 ; python_version >= "3.11" and python_version < "3.13"
-Unidecode==1.3.8 ; python_version >= "3.11" and python_version < "3.13"
-XlsxWriter==3.2.2 ; python_version >= "3.11" and python_version < "3.13"
+Authlib==1.5.2 ; python_version >= "3.11" and python_version < "3.14"
+docuscospacy==0.3.7 ; python_version >= "3.11" and python_version < "3.14"
+google-cloud-firestore==2.20.2 ; python_version >= "3.11" and python_version < "3.14"
+lingua-language-detector==2.0.2 ; python_version >= "3.11" and python_version < "3.14"
+loguru==0.7.3 ; python_version >= "3.11" and python_version < "3.14"
+kaleido==0.2.1 ; python_version >= "3.11" and python_version < "3.14"
+matplotlib==3.7.5 ; python_version >= "3.11" and python_version < "3.14"
+openai==1.66.5 ; python_version >= "3.11" and python_version < "3.14"
+numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.14"
+pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.14"
+plotly==6.0.1 ; python_version >= "3.11" and python_version < "3.14"
+polars==1.42.1 ; python_version >= "3.11" and python_version < "3.14"
+psutil==7.0.0 ; python_version >= "3.11" and python_version < "3.14"
+SQLAlchemy==2.0.41 ; python_version >= "3.11" and python_version < "3.14"
+alembic==1.16.1 ; python_version >= "3.11" and python_version < "3.14"
+psycopg[binary]==3.2.9 ; python_version >= "3.11" and python_version < "3.14"
[…]
