+++ pyproject.toml
@@ -1,9 +1,9 @@
 [project]
 name = "docuscope-ca-online"
-version = "0.4.2"
+version = "0.5.0"
 description = "An online version of the Docuscope CA tool for analyzing text."
 readme = "README.md"
-requires-python = "==3.11.*"
+requires-python = ">=3.11,<3.14"
 keywords = ["nlp", "language"]
 authors = [
     { name = "David Brown", email = "dwb2@andrew.cmu.edu" }
@@ -15,11 +15,13 @@ license = { text = "Apache-2.0" }
 classifiers = [
     "Programming Language :: Python :: 3 :: Only",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "License :: OSI Approved :: Apache Software License",
 ]
 dependencies = [
     "Authlib==1.5.2",
-    "docuscospacy==0.3.6",
+    "docuscospacy==0.3.7",
     "google-cloud-firestore==2.20.2",
     "lingua-language-detector==2.0.2",
     "loguru==0.7.3",
@@ -28,18 +30,19 @@ dependencies = [
     "numpy==1.26.4",
     "openai==1.66.5",
     "pandas==2.2.3",
-    "pandasai==3.0.0b18",
-    "pandasai-openai==0.1.5",
     "plotly==6.0.1",
-    "polars==1.31.0",
+    "polars==1.42.1",
     "psutil==7.0.0",
+    "SQLAlchemy==2.0.41",
+    "alembic==1.16.1",
+    "psycopg[binary]==3.2.9",
+    "redis==5.2.1",
+    "rq==1.16.2",
     "python-docx==1.1.2",
     "RestrictedPython==8.0.0",
-    "scikit-learn==1.6.1",
-    "scipy==1.10.1",
     "seaborn==0.12.2",
-    "spacy==3.8.7",
-    "streamlit==1.45.1",
+    "spacy==3.8.14",
+    "streamlit==1.58.0",
     "Unidecode==1.3.8",
     "XlsxWriter==3.2.2",
 ]
@@ -63,6 +66,7 @@ find = { namespaces = false }
 
 [project.scripts]
 docuscope-ca = "docuscope_ca.cli:main"
+docuscope-rq-smoke = "webapp.queue.smoke_test:main"
 
 [tool.setuptools_scm]
 
