+++ webapp/config/runtime_config.py
@@ -15,8 +15,8 @@ from typing import Dict, Any, Optional
 
 from webapp.config.static_config import get_static_value
 from webapp.config.unified import get_config
+from webapp.persistence.runtime_config_service import RuntimeConfigService
 from webapp.utilities.configuration.logging_config import get_logger
-from webapp.utilities.storage.backend_factory import get_session_backend
 
 logger = get_logger()
 
@@ -36,6 +36,7 @@ class RuntimeConfigManager:
         self._lock = threading.RLock()
         self._cache_ttl = 5.0  # Cache for 5 seconds to reduce DB hits
         self._initialized = False
+        self._service = None
 
         # Check if running in desktop mode - disable runtime config for
         # lightweight operation
@@ -44,7 +45,7 @@ class RuntimeConfigManager:
             self._initialized = True  # Mark as initialized to skip database operations
 
     def _ensure_initialized(self):
-        """Ensure the runtime config tables exist in the health database."""
+        """Ensure the runtime config tables exist in the control plane."""
         if self._initialized:
             return
 
@@ -54,41 +55,19 @@ class RuntimeConfigManager:
             return
 
         try:
-            backend = get_session_backend()
-            with backend.get_health_connection() as conn:
-                cursor = conn.cursor()
-
-                # Create runtime configuration table
-                cursor.execute("""
-                    CREATE TABLE IF NOT EXISTS runtime_config (
-                        config_key TEXT PRIMARY KEY,
-                        config_value TEXT NOT NULL,
-                        config_type TEXT NOT NULL,
-                        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                        updated_by TEXT,
-                        description TEXT
-                    )
-                """)
-
-                # Create configuration audit log
-                cursor.execute("""
-                    CREATE TABLE IF NOT EXISTS config_audit_log (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        config_key TEXT NOT NULL,
-                        old_value TEXT,
-                        new_value TEXT,
-                        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                        updated_by TEXT,
-                        instance_id TEXT
-                    )
-                """)
-
-                conn.commit()
-                self._initialized = True
+            self._get_service()
+            self._initialized = True
 
         except Exception as e:
             logger.error(f"Failed to initialize runtime config tables: {e}")
             self._initialized = False
+
+    def _get_service(self) -> RuntimeConfigService:
+        """Return the shared runtime config service."""
+
+        if self._service is None:
+            self._service = RuntimeConfigService()
+        return self._service
 
     def _get_from_database(self, config_key: str) -> Optional[Any]:
         """Get configuration value from shared database."""
@@ -97,29 +76,7 @@ class RuntimeConfigManager:
             return None
 
         try:
-            backend = get_session_backend()
-            with backend.get_health_connection() as conn:
-                cursor = conn.cursor()
-                cursor.execute(
-                    "SELECT config_value, config_type FROM runtime_config "
-                    "WHERE config_key = ?",
-                    (config_key,)
-                )
-                result = cursor.fetchone()
-
-                if result:
-                    value_str, value_type = result
-                    # Convert back to appropriate type
-                    if value_type == 'bool':
-                        return value_str.lower() == 'true'
-                    elif value_type == 'int':
-                        return int(value_str)
-                    elif value_type == 'float':
-                        return float(value_str)
-                    else:
-                        return value_str
-
-                return None
+            return self._get_service().get_value(config_key)
 
         except Exception as e:
             logger.error(f"Failed to get runtime config {config_key}: {e}")
[…]
