### 2026-08-03 865c689
Update plotbot.py
