+++ webapp/utilities/ai/plotbot.py
@@ -11,9 +11,7 @@ from dataclasses import dataclass
 
 import pandas as pd
 import polars as pl
-import matplotlib.pyplot as plt
 import streamlit as st
-import seaborn as sns
 import plotly.express as px
 from RestrictedPython import compile_restricted
 from RestrictedPython.Guards import safe_builtins, guarded_unpack_sequence
@@ -284,7 +282,7 @@ def plotbot_code_generate_or_update(
     user_request : str
         User's plotting request.
     plot_lib : str
-        The plotting library to use ('matplotlib', 'seaborn', 'plotly.express').
+        The plotting library to use ('plotly.express').
     schema : str
         String representation of the dataframe schema.
     api_key : str
@@ -304,45 +302,17 @@ def plotbot_code_generate_or_update(
     non_numeric_columns = df.select_dtypes(exclude=['number']).columns.tolist()
 
     if code_chunk is None:
-        # Library-specific instructions and examples
-        if plot_lib == "matplotlib":
-            lib_instructions = """
-    - Use the format 'fig, ax = plt.subplots()' to create the figure.
-    - Do not call 'fig.show()' or 'plt.show()'.
-    - Use matplotlib functions like ax.plot(), ax.bar(), ax.scatter(), etc.
-
-    Example:
-    fig, ax = plt.subplots()
-    ax.plot(df['col1'], df['col2'])
-    ax.set_xlabel('Column 1')
-    ax.set_ylabel('Column 2')"""
-        elif plot_lib == "seaborn":
-            lib_instructions = """
-    - Use seaborn functions like sns.barplot(), sns.scatterplot(), sns.lineplot(), etc.
-    - Always create a figure first with 'fig, ax = plt.subplots()'.
-    - Pass the 'ax' parameter to seaborn functions.
-    - Do not call 'plt.show()'.
-
-    Example:
-    fig, ax = plt.subplots()
-    sns.barplot(data=df, x='col1', y='col2', ax=ax)
-    ax.set_title('My Plot')"""
-        elif plot_lib == "plotly.express":
-            lib_instructions = """
+        lib_instructions = """
     - Use plotly.express functions like px.bar(), px.scatter(), px.line(), etc.
     - Assign the result to a variable called 'fig'.
     - Do not call 'fig.show()'.
-    - For gridlines, use fig.update_xaxes(showgrid=True) and fig.update_yaxes(showgrid=True). Do not use fig.update_layout(grid_xaxis=...) or fig.update_layout(grid_yaxis=...).
+        - For gridlines, use fig.update_xaxes(showgrid=True) and
+            fig.update_yaxes(showgrid=True). Do not use
+            fig.update_layout(grid_xaxis=...) or fig.update_layout(grid_yaxis=...).
 
     Example:
     fig = px.bar(df, x='col1', y='col2')
     fig.update_layout(title='My Plot')"""
-        else:
-            # Default to matplotlib
-            lib_instructions = """
-    - Use the format 'fig, ax = plt.subplots()' to create the figure.
-    - Do not call 'fig.show()' or 'plt.show()'.
-    - Use matplotlib functions like ax.plot(), ax.bar(), ax.scatter(), etc."""
 
         prompt = f"""
     You are a Python plotting assistant.
@@ -378,45 +348,17 @@ def plotbot_code_generate_or_update(
     """  # noqa: E501
 
     else:
-        # Library-specific instructions for updates
-        if plot_lib == "matplotlib":
-            lib_instructions = """
-    - Use the format 'fig, ax = plt.subplots()' to create the figure.
-    - Do not call 'fig.show()' or 'plt.show()'.
[…]
