### 2026-08-03 ea1ff8d
Update shared.py
