### 2026-08-03 ae363ff
Update 11_assisted_plotting.py
