+++ requirements.txt
@@ -4,7 +4,6 @@ google-cloud-firestore==2.20.2 ; python_version >= "3.11" and python_version < "
 lingua-language-detector==2.0.2 ; python_version >= "3.11" and python_version < "3.14"
 loguru==0.7.3 ; python_version >= "3.11" and python_version < "3.14"
 kaleido==0.2.1 ; python_version >= "3.11" and python_version < "3.14"
-matplotlib==3.7.5 ; python_version >= "3.11" and python_version < "3.14"
 openai==1.66.5 ; python_version >= "3.11" and python_version < "3.14"
 numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.14"
 pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.14"
@@ -18,7 +17,6 @@ redis==5.2.1 ; python_version >= "3.11" and python_version < "3.14"
 rq==1.16.2 ; python_version >= "3.11" and python_version < "3.14"
 python-docx==1.1.2 ; python_version >= "3.11" and python_version < "3.14"
 RestrictedPython==8.0.0 ; python_version >= "3.11" and python_version < "3.14"
-seaborn==0.12.2 ; python_version >= "3.11" and python_version < "3.14"
 spacy==3.8.14 ; python_version >= "3.11" and python_version < "3.14"
 streamlit==1.58.0 ; python_version >= "3.11" and python_version < "3.14"
 Unidecode==1.3.8 ; python_version >= "3.11" and python_version < "3.14"
