### 2026-08-03 ebbeb3f
Update pyproject.toml
