+++ pyproject.toml
@@ -26,7 +26,6 @@ dependencies = [
     "lingua-language-detector==2.0.2",
     "loguru==0.7.3",
     "kaleido==0.2.1",
-    "matplotlib==3.7.5",
     "numpy==1.26.4",
     "openai==1.66.5",
     "pandas==2.2.3",
@@ -40,7 +39,6 @@ dependencies = [
     "rq==1.16.2",
     "python-docx==1.1.2",
     "RestrictedPython==8.0.0",
-    "seaborn==0.12.2",
     "spacy==3.8.14",
     "streamlit==1.58.0",
     "Unidecode==1.3.8",
