### 2026-08-03 49c8951
Update logging_config.py
