### 2026-08-03 cb4bcea
Update validation_enhanced.py
