### 2026-08-03 295d1d5
Update cache_management.py
