### 2026-08-03 9f8bf69
Update sqlite_session_backend.py
