+++ requirements.txt
@@ -1,7 +1,7 @@
 Authlib==1.5.2 ; python_version >= "3.11" and python_version < "3.14"
 docuscospacy==0.3.7 ; python_version >= "3.11" and python_version < "3.14"
 google-cloud-firestore==2.20.2 ; python_version >= "3.11" and python_version < "3.14"
-lingua-language-detector==2.0.2 ; python_version >= "3.11" and python_version < "3.14"
+lingua-language-detector==2.1.1 ; python_version >= "3.11" and python_version < "3.14"
 loguru==0.7.3 ; python_version >= "3.11" and python_version < "3.14"
 kaleido==0.2.1 ; python_version >= "3.11" and python_version < "3.14"
 openai==1.66.5 ; python_version >= "3.11" and python_version < "3.14"
