### 2026-08-03 e73f34e
Update pyproject.toml
