### 2026-08-03 70f6c52
Update test_authorization.py
