### 2026-08-03 2b13dc4
Update test_configuration.py
