### 2026-08-03 366eacd
Update test_configuration.py
