### 2026-08-03 b80730b
Update test_configuration.py
