### 2026-08-03 8607468
Create test_postgres_control_plane.py
