### 2026-08-03 e6867a4
Update CHANGELOG.md
