+++ .github/workflows/test.yml
@@ -10,6 +10,12 @@ on:
     tags:
       - "v*.*.*"
   workflow_dispatch:
+    inputs:
+      run-postgres-smoke:
+        description: "Run the real Postgres control-plane smoke test"
+        required: false
+        type: boolean
+        default: true
   workflow_call:
 
 jobs:
@@ -202,6 +208,47 @@ jobs:
         sleep 30
         curl -f http://localhost:8501/ || exit 1
         docker stop docuscope-test
+
+  postgres-control-plane-smoke:
+    runs-on: ubuntu-latest
+    needs: test
+    if: github.event_name == 'workflow_dispatch' && inputs.run-postgres-smoke
+
+    services:
+      postgres:
+        image: postgres:16-alpine
+        env:
+          POSTGRES_DB: docuscope_ca
+          POSTGRES_USER: docuscope
+          POSTGRES_PASSWORD: docuscope
+        ports:
+          - 5432:5432
+        options: >-
+          --health-cmd "pg_isready -U docuscope -d docuscope_ca"
+          --health-interval 10s
+          --health-timeout 5s
+          --health-retries 5
+
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: "3.11"
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+        pip install pytest
+
+    - name: Run Postgres control-plane smoke
+      env:
+        DOCUSCOPE_POSTGRES_INTEGRATION: "1"
+        DATABASE_URL: postgresql+psycopg://docuscope:docuscope@localhost:5432/docuscope_ca
+      run: |
+        python -m pytest tests/integration/test_postgres_control_plane.py -v --tb=short -ra
 
   desktop-test:
     runs-on: ${{ matrix.os }}
