### 2026-08-03 74fe7dd
Create compose-rq-smoke.sh
