### 2026-08-03 8ccac0c
Update CONTRIBUTING.md
