+++ webapp/pages/1_load_corpus.py
@@ -10,10 +10,10 @@ This module provides functionality for:
 """
 
 import os
+from functools import partial
 
 import spacy
 import streamlit as st
-from time import perf_counter
 
 # Core application utilities with standardized patterns
 from webapp.utilities.core import app_core
@@ -33,6 +33,7 @@ from webapp.utilities.analysis import (
 from webapp.utilities.processing import (
     process_external, process_internal,
     process_new, handle_uploaded_parquet,
+    handle_uploaded_tabular,
     attach_queued_internal_target,
     handle_uploaded_text, sidebar_process_section
     )
@@ -383,13 +384,15 @@ def main() -> None:
                     'Select a saved corpus to load:',
                     sorted(saved_corpora)
                 )
-            process_fn = lambda: process_internal(
+            process_fn = partial(
+                process_internal,
                 saved_corpora.get(to_load),
                 user_session_id,
                 CorpusKeys.TARGET
             )
             if get_redis_queue_config().enabled:
-                process_fn = lambda: _submit_internal_target_job(
+                process_fn = partial(
+                    _submit_internal_target_job,
                     saved_corpora.get(to_load),
                     user_session_id,
                 )
@@ -413,7 +416,6 @@ def main() -> None:
 
     # If a target corpus is already loaded
     if has_target:
-        target_branch_start = perf_counter()
         probe_state = st.session_state.get(PROCESS_TARGET_PROBE_STAGE_KEY)
         if (
             isinstance(probe_state, dict) and
@@ -444,18 +446,15 @@ def main() -> None:
         # Load and display corpus information
         load_and_display_target_corpus(session, user_session_id)
 
-
         if not has_meta:
             if metadata_target is None:
                 metadata_target = load_metadata("target", user_session_id)
 
-            target_metadata_start = perf_counter()
             handle_target_metadata_processing(metadata_target, user_session_id)
 
         # If reference corpus is loaded, show info and warnings
         if not has_reference:
             # Reference corpus not loaded: offer options to load one
-            reference_prompt_start = perf_counter()
             st.markdown("---")
             st.markdown('##### Reference corpus:')
             load_ref = st.radio(
@@ -477,7 +476,6 @@ def main() -> None:
 
             if load_ref == 'Yes':
                 if metadata_target is None:
-                    metadata_start = perf_counter()
                     metadata_target = load_metadata("target", user_session_id)
 
                 # Choose reference corpus source
@@ -676,6 +674,86 @@ def main() -> None:
                                 ref_persistence_policy,
                             ))
 
+                # Option 4: Process new reference corpus from a tabular file
+                if ref_corpus_source == 'Tabular':
+                    st.markdown(
+                        """
+                        :material/priority:
+                        Upload a single **Parquet**, **CSV**, or **TSV** file
+                        with columns named **doc_id** and **text**.
+
+                        :material/priority:
+                        Extra columns are allowed, but only **doc_id** and
+                        **text** will be used for corpus processing.
+
+                        :material/priority:
+                        Once you've selected your file, click the **UPLOAD REFERENCE**
+                        button and a processing button will appear in the sidebar.
+
+                        :material/priority: Your reference will be tagged with
+                        **the same model** as your target corpus.
+
+                        :material/priority: Be sure that document IDs are unique
+                        and that they don't share IDs with your target corpus.
+                        """
+                        )
+
+                    corp_df, ready, exceptions = None, False, []
+
+                    with st.form("ref-tabular-form", clear_on_submit=True):
+                        ref_file = st.file_uploader(
+                            "Upload your reference corpus",
+                            type=["parquet", "csv", "tsv"],
+                            accept_multiple_files=False
[…]
