+++ webapp/utilities/processing/corpus_processing.py
@@ -7,6 +7,7 @@ newly uploaded text, and corpus finalization.
 
 import gzip
 import hashlib
+import io
 import os
 import pickle
 from pathlib import Path
@@ -17,7 +18,7 @@ import streamlit as st
 import docuscospacy as ds
 
 # Module-specific imports
-from webapp.utilities.processing.corpus_loading import load_corpus_internal, load_corpus_new
+from webapp.utilities.processing.corpus_loading import load_corpus_internal
 from webapp.utilities.session import (
     build_corpus_metadata_descriptor,
     init_metadata_target,
@@ -29,7 +30,11 @@ from webapp.utilities.session.session_persistence import (
     auto_persist_session,
     mark_session_dirty,
 )
-from webapp.utilities.analysis.data_validation import check_corpus_external, check_corpus_new  # noqa: E501
+from webapp.utilities.analysis.data_validation import (
+    check_corpus_external,
+    check_corpus_new,
+    check_language,
+)
 from webapp.utilities.state import (
     CorpusPersistencePolicy,
     LoadCorpusKeys,
@@ -677,7 +682,10 @@ def process_internal(
                     for stage_name, stage_value in stage_timings.items()
                 )
                 logger.warning(
-                    "LOAD_TEST_PROCESS_TARGET_CALLBACK_STAGE_LOG session={} corpus_type={} {}",
+                    (
+                        "LOAD_TEST_PROCESS_TARGET_CALLBACK_STAGE_LOG session={} "
+                        "corpus_type={} {}"
+                    ),
                     user_session_id,
                     corpus_type,
                     stage_summary,
@@ -1018,6 +1026,195 @@ def handle_uploaded_text(
         df, exceptions = None, []
 
     return df, ready, exceptions
+
+
+def handle_uploaded_tabular(
+        uploaded_file,
+        check_size: bool,
+        max_size: int,
+        check_language_flag=False,
+        check_ref=False,
+        target_docs=None
+        ) -> tuple[pl.DataFrame | None, bool, list]:
+    """
+    Handle a single tabular corpus upload with doc_id and text columns.
+
+    Parameters
+    ----------
+    uploaded_file : UploadedFile or None
+        Uploaded Parquet, CSV, or TSV file containing corpus rows.
+    check_size : bool
+        Whether to check corpus size.
+    max_size : int
+        Maximum allowed corpus size.
+    check_language_flag : bool, optional
+        Whether to check language of documents.
+    check_ref : bool, optional
+        Whether to check for reference documents.
+    target_docs : list, optional
+        Target documents for duplicate checking.
+
+    Returns
+    -------
+    tuple[pl.DataFrame | None, bool, list]
+        Tuple of (dataframe, ready_to_process, exceptions).
+    """
+    if uploaded_file is None:
+        return None, False, []
+
+    file_name = uploaded_file.name
+    file_ext = os.path.splitext(file_name)[1].lower().lstrip(".")
+    file_bytes = uploaded_file.getvalue()
+
+    if file_ext not in {"parquet", "csv", "tsv"}:
+        st.error(
+            "Tabular corpora must be uploaded as Parquet, CSV, or TSV files.",
+            icon=":material/block:"
+        )
+        return None, False, []
+
+    try:
+        file_buffer = io.BytesIO(file_bytes)
+        if file_ext == "parquet":
+            df = pl.read_parquet(file_buffer)
+        elif file_ext == "tsv":
+            df = pl.read_csv(file_buffer, separator="\t", infer_schema=False)
+        else:
+            df = pl.read_csv(file_buffer, infer_schema=False)
+    except Exception as e:
+        st.error(f"Error processing tabular corpus file: {e}")
[…]
