### 2026-08-03 0cd0eda
Update registry.py
