### 2026-08-03 8865fc8
Updated storage path policy
