+++ webapp/pages/12_assisted_analysis.py
@@ -97,7 +97,8 @@ def _build_group_selection(
     group_b = group_b or []
     overlap = sorted(set(group_a).intersection(group_b))
     if overlap:
-        return None, None, f"Categories cannot be in both group A and group B: {', '.join(overlap)}"
+        overlap_labels = ", ".join(overlap)
+        return None, None, f"Categories cannot be in both group A and group B: {overlap_labels}"
     if not group_a or not group_b:
         return None, None, "Select at least one category for group A and group B."
 
