### 2026-08-04 c3c7e4f
linted utilities module
