+++ tests/unit/plotting/test_pca_plots.py
@@ -25,7 +25,12 @@ def test_pca_contributions_matches_reference_values():
         [0.16566422146324572, -0.35308644151685975, 0.1713326186599929, 0.01608960139362119]
     )
     assert pca_df["PC3"].tolist() == pytest.approx(
-        [0.0766327662243805, -0.04934110141569103, -0.19117039844351422, 0.16387873363482452]
+        [
+            0.0766327662243805,
+            -0.04934110141569103,
+            -0.19117039844351422,
+            0.16387873363482452,
+        ]
     )
     assert pca_df["Group"].tolist() == ["A", "B", "A", "B"]
     assert pca_df["doc_id"].tolist() == ["doc1", "doc2", "doc3", "doc4"]

[…]
