### 2026-08-05 f77f9e1
Update qwen-check.sh
