### 2026-08-05 0e363d3
Update qwen-up.sh
