### 2026-08-05 a8d252f
Cleaned up legacy code
