+++ webapp/utilities/common/__init__.py
@@ -5,9 +5,12 @@ This package provides utility functions for working with documents, IDs,
 and basic data structures without complex dependencies.
 """
 
-from webapp.utilities.common.document_utils import get_doc_cats, safe_metadata_get
+from webapp.utilities.common.document_utils import (
+    get_doc_cats, safe_metadata_get, sanitize_doc_id
+)
 
 __all__ = [
     'get_doc_cats',
-    'safe_metadata_get'
+    'safe_metadata_get',
+    'sanitize_doc_id'
 ]
