+++ webapp/utilities/processing/corpus_discovery.py
@@ -7,9 +7,9 @@ files and reference corpora.
 
 import os
 import pathlib
-from webapp.utilities.configuration.config_manager import config_manager
+from webapp.config.unified import config
 
-CORPUS_DIR = config_manager.corpus_dir_path
+CORPUS_DIR = config.corpus_dir_path
 
 
 def find_saved(model_type: str) -> dict:
