+++ webapp/utilities/configuration/__init__.py
@@ -1,28 +1,11 @@
 """
 Configuration management utilities.
 
-This module provides functions for loading and managing configuration
-from TOML files and other configuration sources, including centralized
-logging configuration for the entire application.
+This module provides centralized logging configuration for the entire
+application. General app configuration access lives in
+webapp/config/unified.py (the ConfigManager `config` singleton).
 """
 
-from webapp.utilities.configuration.config_manager import (
-    get_version_from_pyproject,
-    import_options_general,
-    get_ai_configuration,
-    ConfigurationManager,
-    config_manager,
-    # Convenient module-level accessors
-    get_desktop_mode,
-    get_cache_mode,
-    get_llm_model,
-    get_llm_parameters,
-    is_online_mode,
-    should_check_size,
-    should_check_language,
-    get_max_text_size,
-    get_max_polars_size
-)
 from webapp.utilities.configuration.logging_config import (
     LoggingConfig,
     get_logging_config,
@@ -36,22 +19,6 @@ from webapp.utilities.configuration.logging_config import (
 
 
 __all__ = [
-    # Configuration manager
-    'get_version_from_pyproject',
-    'import_options_general',
-    'get_ai_configuration',
-    'ConfigurationManager',
-    'config_manager',
-    # Convenient accessors
-    'get_desktop_mode',
-    'get_cache_mode',
-    'get_llm_model',
-    'get_llm_parameters',
-    'is_online_mode',
-    'should_check_size',
-    'should_check_language',
-    'get_max_text_size',
-    'get_max_polars_size',
     # Logging utilities
     'LoggingConfig',
     'get_logging_config',
