### 2026-08-05 ea01879
Update 7_collocations.py
