### 2026-08-05 c45263e
Update 11_assisted_plotting.py
