### 2026-08-05 33488f8
Update registry.py
