+++ webapp/persistence/registry.py
@@ -2,14 +2,18 @@
 
 from __future__ import annotations
 
+from contextlib import contextmanager
 from dataclasses import dataclass
 from datetime import datetime, timezone
 from functools import lru_cache
 import hashlib
 import json
 import os
+import shutil
+import tempfile
+import uuid
 from pathlib import Path
-from typing import Any
+from typing import Any, Iterator
 
 import polars as pl
 from sqlalchemy import select
@@ -17,6 +21,7 @@ from sqlalchemy.exc import IntegrityError
 
 from webapp.persistence.database import create_session_factory
 from webapp.persistence.models import ArtifactJob, ArtifactRecord
+from webapp.corpus_paths import make_portable_corpus_path
 
 
 ARTIFACT_STORE_ROOT = Path("webapp/_artifacts")
@@ -62,9 +67,13 @@ def _hash_payload(payload: dict[str, Any]) -> str:
 
 
 def _normalize_source_path(raw_path: str) -> str:
-    path = Path(raw_path).resolve()
+    portable_path = make_portable_corpus_path(raw_path)
+    if portable_path != raw_path:
+        return portable_path
+
+    path = Path(raw_path).resolve(strict=False)
     try:
-        return path.relative_to(Path.cwd().resolve()).as_posix()
+        return path.relative_to(Path.cwd().resolve(strict=False)).as_posix()
     except ValueError:
         return path.as_posix()
 
@@ -95,6 +104,47 @@ def _build_artifact_dir(identity: ArtifactIdentity) -> Path:
         identity.selector_hash /
         identity.parameter_hash
     )
+
+
+def _publish_staged_dir(staging_dir: Path, final_dir: Path) -> None:
+    """Atomically swap a fully-written staging directory into its final path.
+
+    Using a same-filesystem rename means readers only ever see either the
+    previous complete artifact or the new complete one, never a partially
+    written directory (e.g. if a worker crashes mid-write).
+    """
+
+    if final_dir.exists():
+        stale_dir = final_dir.with_name(f"{final_dir.name}.stale-{uuid.uuid4().hex}")
+        os.replace(final_dir, stale_dir)
+        try:
+            os.replace(staging_dir, final_dir)
+        finally:
+            shutil.rmtree(stale_dir, ignore_errors=True)
+    else:
+        os.replace(staging_dir, final_dir)
+
+
+@contextmanager
+def _staged_artifact_dir(final_dir: Path) -> Iterator[Path]:
+    """Yield a temporary directory to write artifact files into, then publish it.
+
+    On success, the staging directory is atomically renamed to `final_dir`.
+    On failure, the partially written staging directory is discarded and the
+    previous artifact at `final_dir` (if any) is left untouched.
+    """
+
+    final_dir.parent.mkdir(parents=True, exist_ok=True)
+    staging_dir = Path(tempfile.mkdtemp(
+        prefix=f".{final_dir.name}.staging-", dir=final_dir.parent
+    ))
+    try:
+        yield staging_dir
+    except BaseException:
+        shutil.rmtree(staging_dir, ignore_errors=True)
+        raise
+    else:
+        _publish_staged_dir(staging_dir, final_dir)
 
 
 @lru_cache(maxsize=64)
@@ -563,10 +613,20 @@ class ArtifactRegistryService:
             return session.get(ArtifactJob, job_id)
 
     def get_artifact_by_id(self, artifact_id: int) -> ArtifactRecord | None:
-        """Return an artifact by primary key without changing its state."""
+        """Return an artifact by primary key when its ready storage is usable."""
 
         with self._session_factory() as session:
-            return session.get(ArtifactRecord, artifact_id)
+            artifact = session.get(ArtifactRecord, artifact_id)
[…]
