+++ webapp/queue/config.py
@@ -15,6 +15,8 @@ class RedisQueueConfig:
     queue_name: str
     job_timeout: int
     result_ttl: int
+    max_retries: int
+    retry_interval_seconds: int
 
 
 def get_redis_queue_config() -> RedisQueueConfig:
@@ -26,4 +28,10 @@ def get_redis_queue_config() -> RedisQueueConfig:
         queue_name=os.environ.get("DOCUSCOPE_RQ_QUEUE", "docuscope"),
         job_timeout=int(os.environ.get("DOCUSCOPE_RQ_JOB_TIMEOUT", "600")),
         result_ttl=int(os.environ.get("DOCUSCOPE_RQ_RESULT_TTL", "86400")),
+        # Bounded retries so a transient Redis/filesystem blip doesn't strand
+        # a student on a permanently failed job; 0 disables retries entirely.
+        max_retries=int(os.environ.get("DOCUSCOPE_RQ_MAX_RETRIES", "2")),
+        retry_interval_seconds=int(
+            os.environ.get("DOCUSCOPE_RQ_RETRY_INTERVAL_SECONDS", "15")
+        ),
     )
