### 2026-08-05 f83b657
Create corpus_paths.py
