+++ .github/workflows/test.yml
@@ -75,7 +75,7 @@ jobs:
       run: |
         pip install flake8
         flake8 webapp/ tests/ --count --select=E9,F63,F7,F82 --show-source --statistics
-        flake8 webapp/ tests/ --count --exit-zero --max-complexity=10 --max-line-length=120 --statistics
+        flake8 webapp/ tests/ --count --exit-zero --max-complexity=10 --statistics
     
     - name: Run unit tests
       run: |
