+++ tests/unit/persistence/test_workflows.py
@@ -1,0 +1,234 @@
+"""Unit tests for `SharedArtifactWorkflow` (reserve/load/store control flow).
+
+Uses a mocked registry/logger so these tests stay fast and independent of the
+database — the real `ArtifactRegistryService` behavior is covered separately
+in test_registry.py.
+"""
+
+from __future__ import annotations
+
+from unittest.mock import MagicMock
+
+import pytest
+
+from webapp.persistence.workflows import SharedArtifactWorkflow
+
+
+@pytest.fixture
+def registry():
+    return MagicMock()
+
+
+@pytest.fixture
+def logger():
+    return MagicMock()
+
+
+@pytest.fixture
+def workflow(registry, logger):
+    return SharedArtifactWorkflow(registry, logger)
+
+
+class TestLoadReady:
+    def test_bypasses_when_identity_is_none(self, workflow, registry):
+        result = workflow.load_ready(None, loader=MagicMock(), cache_name="keyness")
+
+        assert result is None
+        registry.find_ready_artifact.assert_not_called()
+
+    def test_returns_none_on_cache_miss(self, workflow, registry):
+        identity = MagicMock(selector_hash="sel", parameter_hash="params")
+        registry.find_ready_artifact.return_value = None
+
+        result = workflow.load_ready(identity, loader=MagicMock(), cache_name="keyness")
+
+        assert result is None
+
+    def test_returns_artifact_and_payload_on_hit(self, workflow, registry):
+        identity = MagicMock(selector_hash="sel", parameter_hash="params")
+        artifact = MagicMock(artifact_id=1)
+        registry.find_ready_artifact.return_value = artifact
+        loader = MagicMock(return_value={"data": 1})
+
+        result = workflow.load_ready(identity, loader=loader, cache_name="keyness")
+
+        assert result == (artifact, {"data": 1})
+        loader.assert_called_once_with(artifact)
+
+    def test_returns_none_when_loader_raises(self, workflow, registry, logger):
+        identity = MagicMock(selector_hash="sel", parameter_hash="params")
+        registry.find_ready_artifact.return_value = MagicMock(artifact_id=1)
+        loader = MagicMock(side_effect=RuntimeError("corrupt"))
+
+        result = workflow.load_ready(identity, loader=loader, cache_name="keyness")
+
+        assert result is None
+        logger.warning.assert_called_once()
+
+
+class TestReserve:
+    def test_bypasses_when_identity_is_none(self, workflow, registry):
+        decision = workflow.reserve(None, cache_name="keyness")
+
+        assert decision.state == "bypass"
+        registry.reserve_artifact.assert_not_called()
+
+    def test_bypasses_when_reservation_raises(self, workflow, registry, logger):
+        identity = MagicMock(selector_hash="sel")
+        registry.reserve_artifact.side_effect = RuntimeError("db down")
+
+        decision = workflow.reserve(identity, cache_name="keyness")
+
+        assert decision.state == "bypass"
+        logger.warning.assert_called_once()
+
+    def test_reserved_marks_job_running(self, workflow, registry):
+        identity = MagicMock(selector_hash="sel")
+        job = MagicMock(job_id=42)
+        registry.reserve_artifact.return_value = MagicMock(state="reserved", job=job)
+
+        decision = workflow.reserve(identity, cache_name="keyness")
+
+        assert decision.state == "reserved"
+        assert decision.job_id == 42
+        registry.mark_job_running.assert_called_once_with(42)
+
+    def test_ready_uses_ready_loader_payload(self, workflow, registry):
+        identity = MagicMock(selector_hash="sel")
+        registry.reserve_artifact.return_value = MagicMock(state="ready", job=None)
+        ready_loader = MagicMock(return_value={"data": 1})
+
+        decision = workflow.reserve(
[…]
