+++ webapp/pages/1_load_corpus.py
@@ -197,7 +197,7 @@ def _render_internal_target_queue_status(user_session_id: str) -> None:
         st.warning("Queued target job state was incomplete. Please try again.")
         return
 
-    job_row = registry_service.get_job_by_id(control_plane_job_id)
+    job_row = registry_service.get_public_job_by_id(control_plane_job_id)
     if job_row is None:
         _clear_internal_target_queue_state(user_session_id)
         st.warning("Queued target job could not be found. Please try again.")
