+++ webapp/pages/5_compare_corpora.py
@@ -176,7 +176,7 @@ def _render_keyness_queue_status(user_session_id: str) -> None:
         st.warning("Queued keyness job state was incomplete. Please try again.")
         return
 
-    job_row = registry_service.get_job_by_id(control_plane_job_id)
+    job_row = registry_service.get_public_job_by_id(control_plane_job_id)
     if job_row is None:
         _clear_keyness_queue_state(user_session_id)
         st.warning("Queued keyness job could not be found. Please try again.")
