### 2026-08-05 08c982f
Update 6_compare_corpus_parts.py
