+++ webapp/pages/6_compare_corpus_parts.py
@@ -178,7 +178,7 @@ def _render_keyness_parts_queue_status(user_session_id: str) -> None:
         st.warning("Queued corpus-parts job state was incomplete. Please try again.")
         return
 
-    job_row = registry_service.get_job_by_id(control_plane_job_id)
+    job_row = registry_service.get_public_job_by_id(control_plane_job_id)
     if job_row is None:
         _clear_keyness_parts_queue_state(user_session_id)
         st.warning("Queued corpus-parts job could not be found. Please try again.")
