### 2026-08-05 a898f72
Update 7_collocations.py
