### 2026-08-05 9071cc7
Update 11_assisted_plotting.py
