+++ webapp/pages/11_assisted_plotting.py
@@ -28,11 +28,9 @@ from webapp.utilities.ai import (
     render_quota_tracker, should_show_api_key_input,
     render_work_preservation_interface, should_show_work_preservation_interface,
     export_conversation_history, clear_plotbot_table,
-    prune_message_thread
+    get_plotbot_builtin_sources, prune_message_thread
 )
 from webapp.utilities.ai.providers import get_openai_compatible_provider_config
-from webapp.utilities.configuration.logging_config import get_logger
-from webapp.persistence import registry_service
 from webapp.queue import (
     enqueue_plotbot_generation,
     get_queue,
@@ -54,8 +52,6 @@ from webapp.menu import (
 
 TITLE = "AI-Assisted Plotting"
 ICON = ":material/smart_toy:"
-
-logger = get_logger()
 
 st.set_page_config(
     page_title=TITLE, page_icon=ICON,
@@ -102,18 +98,8 @@ def _dataframe_to_plotbot_records(df) -> list[dict[str, object]]:
     return json.loads(df.to_json(orient="records", date_format="iso"))
 
 
-def _append_queued_plotbot_result(user_session_id: str, artifact_id: int) -> bool:
-    """Attach a completed queued Plotbot JSON result to the chat state."""
-    artifact = registry_service.get_artifact_by_id(artifact_id)
-    if artifact is None:
-        return False
-
-    try:
-        payload = registry_service.load_json_artifact(artifact)
-    except Exception as exc:
-        logger.warning(f"Queued Plotbot artifact load failed for {artifact_id}: {exc}")
-        return False
-
+def _append_queued_plotbot_result(user_session_id: str, payload: dict) -> bool:
+    """Attach a completed ephemeral Plotbot result to the chat state."""
     result = payload.get("result") if isinstance(payload, dict) else None
     if not isinstance(result, dict):
         return False
@@ -157,6 +143,7 @@ def _submit_plotbot_job(
     llm_params: dict,
     code_chunk: str | None,
     user_email: str,
+    source_refs: list[str],
 ) -> None:
     """Submit Plotbot generation to Redis/RQ and store pending state."""
     if df is None:
@@ -175,27 +162,19 @@ def _submit_plotbot_job(
     schema = df.dtypes.to_string() if hasattr(df, "dtypes") else str(type(df))
     result = enqueue_plotbot_generation(
         dataframe_records=dataframe_records,
+        source_refs=source_refs,
         plot_lib=plot_lib,
         user_input=user_input,
         llm_params=llm_params,
         schema=schema,
         code_chunk=code_chunk,
         api_key=api_key,
-        requester_principal_id=user_email or "anonymous",
+        requester_principal_id=f"{user_email or 'anonymous'}:{user_session_id}",
     )
-
-    if result.state == "ready" and result.artifact_id is not None:
-        _clear_plotbot_queue_state(user_session_id)
-        if _append_queued_plotbot_result(user_session_id, result.artifact_id):
-            st.rerun()
-        st.error("Ready Plotbot result could not be attached. Please retry.")
-        return
 
     st.session_state[PLOTBOT_QUEUE_STATE_KEY] = {
         "session_id": user_session_id,
-        "control_plane_job_id": result.control_plane_job_id,
         "rq_job_id": result.rq_job_id,
-        "artifact_id": result.artifact_id,
     }
     st.rerun()
 
@@ -207,54 +186,38 @@ def _render_plotbot_queue_status(user_session_id: str) -> None:
     if queue_state is None:
         return
 
-    control_plane_job_id = queue_state.get("control_plane_job_id")
     rq_job_id = queue_state.get("rq_job_id")
[…]
