+++ webapp/persistence/cleanup.py
@@ -1,0 +1,251 @@
+"""Scheduled cleanup for control-plane rows and shared artifact payloads."""
+
+from __future__ import annotations
+
+import argparse
+from dataclasses import asdict, dataclass
+from datetime import datetime, timedelta, timezone
+import json
+import os
+from pathlib import Path
+import shutil
+import signal
+from threading import Event
+
+from sqlalchemy import delete, exists, func, or_, select
+
+from webapp.persistence.database import (
+    create_session_factory,
+    initialize_database_schema,
+)
+from webapp.persistence.models import (
+    AccessLog,
+    ArtifactJob,
+    ArtifactRecord,
+    ConfigAuditLog,
+    SessionRecord,
+    UserQuery,
+)
+from webapp.persistence.registry import get_artifact_store_root
+from webapp.utilities.storage.postgres_session_backend import PostgresSessionBackend
+
+
+ACTIVE_JOB_STATUSES = ("pending", "running")
+TERMINAL_JOB_STATUSES = ("completed", "failed")
+
+
+@dataclass(frozen=True)
+class CleanupResult:
+    """Counts removed, or eligible when dry-run mode is enabled."""
+
+    dry_run: bool
+    sessions: int
+    user_queries: int
+    artifacts: int
+    artifact_jobs: int
+    access_logs: int
+    config_audit_logs: int
+    payload_directories: int
+
+
+def _env_days(name: str, default: int) -> int:
+    return int(os.environ.get(name, str(default)))
+
+
+def _safe_payload_path(storage_uri: str) -> Path | None:
+    """Return a root-confined artifact payload path, or None when unsafe."""
+
+    if not storage_uri:
+        return None
+    root = get_artifact_store_root().resolve(strict=False)
+    path = Path(storage_uri).resolve(strict=False)
+    try:
+        path.relative_to(root)
+    except ValueError:
+        return None
+    if path == root:
+        return None
+    if not path.exists():
+        return None
+    return path
+
+
+def run_cleanup(
+    *,
+    dry_run: bool = False,
+    now: datetime | None = None,
+) -> CleanupResult:
+    """Apply one conservative retention pass and return cleanup metrics."""
+
+    initialize_database_schema()
+    session_factory = create_session_factory()
+    current_time = now or datetime.now(timezone.utc)
+    private_cutoff = current_time - timedelta(
+        hours=int(os.environ.get("DOCUSCOPE_PRIVATE_ARTIFACT_TTL_HOURS", "24"))
+    )
+    terminal_job_cutoff = current_time - timedelta(
+        days=_env_days("DOCUSCOPE_TERMINAL_JOB_RETENTION_DAYS", 7)
+    )
+    access_log_cutoff = current_time - timedelta(
+        days=_env_days("DOCUSCOPE_ACCESS_LOG_RETENTION_DAYS", 30)
+    )
+    config_audit_cutoff = current_time - timedelta(
+        days=_env_days("DOCUSCOPE_CONFIG_AUDIT_RETENTION_DAYS", 90)
+    )
+    query_cutoff = current_time - timedelta(days=7)
+
+    active_job_exists = exists().where(
+        ArtifactJob.artifact_id == ArtifactRecord.artifact_id,
+        ArtifactJob.status.in_(ACTIVE_JOB_STATUSES),
+    )
+    artifact_eligible = (
+        or_(
+            ArtifactRecord.expires_at <= current_time,
+            (
+                (ArtifactRecord.scope == "private")
+                & (ArtifactRecord.created_at <= private_cutoff)
+            ),
+        )
+        & ~active_job_exists
+    )
+
+    with session_factory() as session:
+        expired_session_ids = session.scalars(
+            select(SessionRecord.session_id).where(
[…]
