### 2026-08-05 33ad298
Update registry.py
