+++ webapp/persistence/registry.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from contextlib import contextmanager
 from dataclasses import dataclass
-from datetime import datetime, timezone
+from datetime import datetime, timedelta, timezone
 from functools import lru_cache
 import hashlib
 import json
@@ -21,7 +21,10 @@ from sqlalchemy.exc import IntegrityError
 
 from webapp.persistence.database import create_session_factory
 from webapp.persistence.models import ArtifactJob, ArtifactRecord
-from webapp.corpus_paths import make_portable_corpus_path
+from webapp.corpus_paths import (
+    is_builtin_corpus_ref,
+    make_portable_corpus_path,
+)
 
 
 ARTIFACT_STORE_ROOT = Path("webapp/_artifacts")
@@ -32,6 +35,19 @@ FREQUENCY_ARTIFACT_TYPE = "frequency_bundle"
 COLLOCATION_ARTIFACT_TYPE = "collocation_bundle"
 NGRAM_ARTIFACT_TYPE = "ngram_bundle"
 JSON_ARTIFACT_FILENAME = "payload.json"
+MODEL_METADATA_FILENAME = "meta.json"
+PUBLIC_OWNER_PRINCIPAL_ID = "__public__"
+PRIVATE_ARTIFACT_TTL_HOURS_ENV = "DOCUSCOPE_PRIVATE_ARTIFACT_TTL_HOURS"
+
+
+def _artifact_expiry(identity: "ArtifactIdentity", now: datetime) -> datetime | None:
+    """Return expiry for private compatibility artifacts only."""
+
+    if identity.scope != "private":
+        return None
+    return now + timedelta(
+        hours=int(os.getenv(PRIVATE_ARTIFACT_TTL_HOURS_ENV, "24"))
+    )
 
 
 @dataclass(frozen=True)
@@ -47,6 +63,20 @@ class ArtifactIdentity:
     pipeline_version: str
     model_version: str
     owner_principal_id: str | None = None
+
+    def __post_init__(self) -> None:
+        """Normalize ownership so public identities are unique in PostgreSQL."""
+
+        if self.scope == "public":
+            if self.owner_principal_id not in (None, PUBLIC_OWNER_PRINCIPAL_ID):
+                raise ValueError("Public artifacts cannot have a private owner.")
+            object.__setattr__(
+                self,
+                "owner_principal_id",
+                PUBLIC_OWNER_PRINCIPAL_ID,
+            )
+        elif not self.owner_principal_id:
+            raise ValueError("Private artifacts require an owner principal.")
 
 
 @dataclass(frozen=True)
@@ -78,6 +108,87 @@ def _normalize_source_path(raw_path: str) -> str:
         return path.as_posix()
 
 
+def _require_builtin_source(raw_path: str) -> str:
+    """Return a portable built-in source or reject private/session data."""
+
+    portable_path = make_portable_corpus_path(raw_path)
+    if not is_builtin_corpus_ref(portable_path):
+        raise ValueError(
+            "Durable shared artifacts require a built-in corpus source."
+        )
+    return portable_path
+
+
+@lru_cache(maxsize=4)
+def _load_model_fingerprint(model_dir: str) -> str:
+    """Return a stable fingerprint from one bundled spaCy model's metadata."""
+
+    metadata_path = Path(model_dir) / MODEL_METADATA_FILENAME
+    try:
+        metadata = json.loads(metadata_path.read_text(encoding="utf-8"))
+    except (OSError, json.JSONDecodeError) as exc:
+        raise ValueError(
+            f"Unable to resolve model metadata from {metadata_path}."
+        ) from exc
+
+    name = str(metadata.get("name", "")).strip()
+    version = str(metadata.get("version", "")).strip()
[…]
