+++ alembic.ini
@@ -1,6 +1,7 @@
 [alembic]
 script_location = migrations
 prepend_sys_path = .
+path_separator = os
 sqlalchemy.url = postgresql+psycopg://docuscope:docuscope@localhost:5432/docuscope_ca
 
 [loggers]
