### 2026-08-05 b4755a0
Update docker-compose.yml
