+++ migrations/versions/ee00f9d5849e_baseline_control_plane_schema.py
@@ -1,0 +1,173 @@
+# flake8: noqa
+"""baseline control plane schema
+
+Revision ID: 20260608_0001
+Revises: 
+Create Date: 2026-08-05 11:08:39.930199
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+revision: str = '20260608_0001'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('access_log',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('email', sa.String(length=255), nullable=False),
+    sa.Column('page', sa.String(length=255), nullable=True),
+    sa.Column('required_role', sa.String(length=64), nullable=True),
+    sa.Column('required_permission', sa.String(length=128), nullable=True),
+    sa.Column('access_granted', sa.Boolean(), nullable=False),
+    sa.Column('timestamp', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('ip_address', sa.String(length=64), nullable=True),
+    sa.Column('user_agent', sa.Text(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('ix_access_log_email_timestamp', 'access_log', ['email', 'timestamp'], unique=False)
+    op.create_table('artifacts',
+    sa.Column('artifact_id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('artifact_type', sa.String(length=64), nullable=False),
+    sa.Column('scope', sa.String(length=16), nullable=False),
+    sa.Column('owner_principal_id', sa.String(length=255), nullable=False),
+    sa.Column('sharing_principal_id', sa.String(length=255), nullable=True),
+    sa.Column('selector_hash', sa.String(length=128), nullable=False),
+    sa.Column('selector_payload', sa.JSON(), nullable=False),
+    sa.Column('pipeline_version', sa.String(length=64), nullable=False),
+    sa.Column('model_version', sa.String(length=64), nullable=False),
+    sa.Column('parameter_hash', sa.String(length=128), nullable=False),
+    sa.Column('parameter_payload', sa.JSON(), nullable=False),
+    sa.Column('storage_uri', sa.Text(), nullable=False),
+    sa.Column('status', sa.String(length=16), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('last_accessed_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('access_count', sa.Integer(), nullable=False),
+    sa.Column('expires_at', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('artifact_id'),
+    sa.UniqueConstraint('artifact_type', 'scope', 'selector_hash', 'parameter_hash', 'pipeline_version', 'model_version', 'owner_principal_id', name='uq_artifact_identity')
+    )
+    op.create_index('ix_artifacts_scope_status', 'artifacts', ['scope', 'status'], unique=False)
+    op.create_table('config_audit_log',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('config_key', sa.String(length=255), nullable=False),
+    sa.Column('old_value', sa.Text(), nullable=True),
+    sa.Column('new_value', sa.Text(), nullable=True),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
[…]
