+++ webapp/queue/__init__.py
@@ -44,7 +44,6 @@ else:
         QueueNgramEnqueueResult,
         QueuePlotbotEnqueueResult,
         QueueSmokeEnqueueResult,
-        build_plotbot_identity,
         build_internal_target_identity,
         build_queue_smoke_identity,
         enqueue_collocation_preparation,
@@ -68,7 +67,6 @@ else:
         "QueuePlotbotEnqueueResult",
         "QueueSmokeEnqueueResult",
         "RedisQueueConfig",
-        "build_plotbot_identity",
         "build_internal_target_identity",
         "build_queue_smoke_identity",
         "enqueue_collocation_preparation",
