+++ webapp/queue/tasks.py
@@ -36,6 +36,30 @@ from webapp.corpus_paths import (
 logger = get_logger()
 
 
+def mark_control_plane_job_failed(
+    rq_job,
+    _connection,
+    _exception_type,
+    exception_value,
+    _traceback,
+) -> None:
+    """Release a registry reservation after RQ exhausts all retries."""
+
+    if getattr(rq_job, "retries_left", 0) > 0:
+        return
+    control_plane_job_id = rq_job.args[0]
+    registry_service.mark_job_failed(control_plane_job_id, str(exception_value))
+
+
+def _record_retry_if_pending(control_plane_job_id: int, failure_reason: str) -> None:
+    """Record an attempt only when RQ has another retry available."""
+
+    current_rq_job = get_current_job()
+    if current_rq_job is None or current_rq_job.retries_left <= 0:
+        return
+    registry_service.record_job_retry(control_plane_job_id, failure_reason)
+
+
 def _get_shared_artifact_workflow() -> SharedArtifactWorkflow:
     """Build a shared-artifact workflow against the current registry service."""
 
@@ -171,7 +195,7 @@ def _build_keyness_parts_metadata(
 def run_registry_smoke_test(control_plane_job_id: int) -> int:
     """Run a tiny registry-backed background job and persist a JSON artifact."""
 
-    job_row = registry_service.get_job_by_id(control_plane_job_id)
+    job_row = registry_service.get_job_by_id_internal(control_plane_job_id)
     if job_row is None:
         raise ValueError(f"Unknown control-plane job id: {control_plane_job_id}")
 
@@ -193,14 +217,14 @@ def run_registry_smoke_test(control_plane_job_id: int) -> int:
         registry_service.mark_job_completed(control_plane_job_id, artifact.artifact_id)
         return artifact.artifact_id
     except Exception as exc:
-        registry_service.mark_job_failed(control_plane_job_id, str(exc))
+        _record_retry_if_pending(control_plane_job_id, str(exc))
         raise
 
 
 def run_internal_target_preparation(control_plane_job_id: int, corpus_path: str) -> int:
     """Prepare a built-in target corpus off the Streamlit rerun path."""
 
-    job_row = registry_service.get_job_by_id(control_plane_job_id)
+    job_row = registry_service.get_job_by_id_internal(control_plane_job_id)
     if job_row is None:
         raise ValueError(f"Unknown control-plane job id: {control_plane_job_id}")
 
@@ -226,7 +250,7 @@ def run_internal_target_preparation(control_plane_job_id: int, corpus_path: str)
         registry_service.mark_job_completed(control_plane_job_id, artifact.artifact_id)
         return artifact.artifact_id
     except Exception as exc:
-        registry_service.mark_job_failed(control_plane_job_id, str(exc))
+        _record_retry_if_pending(control_plane_job_id, str(exc))
         raise
 
 
@@ -239,7 +263,7 @@ def run_keyness_preparation(
 ) -> int:
     """Generate shared keyness tables for a built-in target/reference pair."""
 
-    job_row = registry_service.get_job_by_id(control_plane_job_id)
+    job_row = registry_service.get_job_by_id_internal(control_plane_job_id)
     if job_row is None:
         raise ValueError(f"Unknown control-plane job id: {control_plane_job_id}")
 
@@ -296,7 +320,7 @@ def run_keyness_preparation(
         registry_service.mark_job_completed(control_plane_job_id, artifact.artifact_id)
         return artifact.artifact_id
     except Exception as exc:
-        registry_service.mark_job_failed(control_plane_job_id, str(exc))
+        _record_retry_if_pending(control_plane_job_id, str(exc))
         raise
 
 
@@ -312,7 +336,7 @@ def run_collocation_preparation(
 ) -> int:
[…]
