+++ webapp/queue/client.py
@@ -9,6 +9,7 @@ import json
 
 from redis import Redis
 from rq import Queue, Retry
+from rq.job import Callback
 
 from webapp.persistence import ArtifactIdentity, registry_service
 from webapp.persistence.registry import (
@@ -18,12 +19,15 @@ from webapp.persistence.registry import (
     build_shared_ngram_identity,
     get_pipeline_version,
 )
+from webapp.corpus_paths import is_builtin_corpus_ref, make_portable_corpus_path
 from webapp.queue.config import RedisQueueConfig, get_redis_queue_config
-from webapp.corpus_paths import make_portable_corpus_path
 
 
 RQ_SMOKE_ARTIFACT_TYPE = "rq_smoke_result"
 INTERNAL_TARGET_ARTIFACT_TYPE = "internal_target_ready"
+CONTROL_PLANE_FAILURE_CALLBACK = Callback(
+    "webapp.queue.tasks.mark_control_plane_job_failed"
+)
 
 
 @dataclass(frozen=True)
@@ -91,9 +95,7 @@ class QueuePlotbotEnqueueResult:
     """Result returned when a Plotbot generation job is requested."""
 
     state: str
-    control_plane_job_id: int | None
-    rq_job_id: str | None = None
-    artifact_id: int | None = None
+    rq_job_id: str
 
 
 def _hash_payload(payload: dict[str, object]) -> str:
@@ -193,46 +195,6 @@ def build_internal_target_identity(
     )
 
 
-def build_plotbot_identity(
-    dataframe_records: list[dict[str, object]],
-    plot_lib: str,
-    user_input: str,
-    llm_params: dict[str, object],
-    code_chunk: str | None = None,
-    cached_code: str | None = None,
-    requester_principal_id: str = "anonymous",
-    model_version: str = "plotbot-v1",
-) -> ArtifactIdentity:
-    """Build a private identity for queued Plotbot generation."""
-
-    table_payload = {
-        "records": dataframe_records,
-    }
-    selector_payload = {
-        "request_type": "plotbot_generate",
-        "table_hash": _hash_payload(table_payload),
-        "plot_lib": plot_lib,
-        "user_input": user_input,
-        "code_chunk_hash": _hash_payload({"code_chunk": code_chunk or ""}),
-        "cached_code_hash": _hash_payload({"cached_code": cached_code or ""}),
-    }
-    parameter_payload = {
-        "job_kind": "plotbot_generate",
-        "llm_params": llm_params,
-    }
-    return ArtifactIdentity(
-        artifact_type="plotbot_result",
-        scope="private",
-        owner_principal_id=requester_principal_id or "anonymous",
-        selector_hash=_hash_payload(selector_payload),
-        selector_payload=selector_payload,
-        parameter_hash=_hash_payload(parameter_payload),
-        parameter_payload=parameter_payload,
-        pipeline_version=get_pipeline_version(),
-        model_version=model_version,
-    )
-
-
 def enqueue_registry_smoke_test(
     request_key: str = "local-smoke",
     requester_principal_id: str | None = None,
@@ -294,6 +256,7 @@ def enqueue_registry_smoke_test(
         job_timeout=config.job_timeout,
         result_ttl=config.result_ttl,
         retry=_build_job_retry(config),
+        on_failure=CONTROL_PLANE_FAILURE_CALLBACK,
     )
     return QueueSmokeEnqueueResult(
         state="queued",
@@ -366,6 +329,7 @@ def enqueue_internal_target_preparation(
         job_timeout=config.job_timeout,
         result_ttl=config.result_ttl,
         retry=_build_job_retry(config),
[…]
