### 2026-08-05 d44e243
Update llm_core.py
