### 2026-08-05 4b4cb81
Update plotbot.py
