### 2026-08-05 497bea6
Update data_manager.py
