+++ tests/unit/queue/test_rq_smoke.py
@@ -65,7 +65,7 @@ def test_enqueue_registry_smoke_test_deduplicates_pending_jobs(
     assert second.control_plane_job_id == first.control_plane_job_id
     assert len(fake_queue.enqueued) == 1
 
-    job_row = registry.get_job_by_id(first.control_plane_job_id)
+    job_row = registry.get_job_by_id_internal(first.control_plane_job_id)
     assert job_row is not None
     assert job_row.status == "pending"
 
@@ -97,17 +97,118 @@ def test_run_registry_smoke_test_stores_json_artifact(control_plane_env, monkeyp
     enqueued = client_module.enqueue_registry_smoke_test(request_key="artifact-demo")
     artifact_id = tasks_module.run_registry_smoke_test(enqueued.control_plane_job_id)
 
-    job_row = registry.get_job_by_id(enqueued.control_plane_job_id)
+    job_row = registry.get_job_by_id_internal(enqueued.control_plane_job_id)
     assert job_row is not None
     assert job_row.status == "completed"
     assert job_row.artifact_id == artifact_id
 
-    artifact = registry.get_artifact_by_id(artifact_id)
+    artifact = registry.get_artifact_by_id_internal(artifact_id)
     assert artifact is not None
     payload = registry.load_json_artifact(artifact)
     assert payload["status"] == "ok"
     assert payload["control_plane_job_id"] == enqueued.control_plane_job_id
     assert payload["worker_id"] == "rq-job-42"
+
+
+def test_registry_smoke_retry_keeps_one_reservation_then_completes(
+    control_plane_env,
+    monkeypatch,
+):
+    registry = ArtifactRegistryService()
+    monkeypatch.setattr(client_module, "registry_service", registry)
+    monkeypatch.setattr(tasks_module, "registry_service", registry)
+
+    from webapp.persistence import registry as registry_module
+
+    artifact_root = Path(control_plane_env) / "_artifacts"
+    monkeypatch.setattr(registry_module, "ARTIFACT_STORE_ROOT", artifact_root)
+    rq_job = SimpleNamespace(
+        id="rq-job-retry",
+        retries_left=1,
+        get_status=lambda: JobStatus.QUEUED,
+    )
+
+    class FakeQueue:
+        def __init__(self):
+            self.job = None
+            self.enqueue_kwargs = None
+
+        def fetch_job(self, _job_id):
+            return self.job
+
+        def enqueue(self, _func, *_args, **kwargs):
+            self.enqueue_kwargs = kwargs
+            self.job = rq_job
+            return rq_job
+
+    fake_queue = FakeQueue()
+    monkeypatch.setattr(client_module, "get_queue", lambda: fake_queue)
+    monkeypatch.setattr(tasks_module, "get_current_job", lambda: rq_job)
+
+    first = client_module.enqueue_registry_smoke_test(request_key="retry-demo")
+    original_store = registry.store_json_artifact
+    monkeypatch.setattr(
+        registry,
+        "store_json_artifact",
+        lambda *_args, **_kwargs: (_ for _ in ()).throw(
+            RuntimeError("temporary failure")
+        ),
+    )
+
+    with pytest.raises(RuntimeError, match="temporary failure"):
+        tasks_module.run_registry_smoke_test(first.control_plane_job_id)
+    tasks_module.mark_control_plane_job_failed(
+        SimpleNamespace(args=(first.control_plane_job_id,), retries_left=1),
+        None,
+        RuntimeError,
+        RuntimeError("temporary failure"),
+        None,
+    )
+
+    pending = client_module.enqueue_registry_smoke_test(request_key="retry-demo")
+    job_row = registry.get_job_by_id_internal(first.control_plane_job_id)
[…]
