### 2026-08-05 5c0ffa5
Update CITATION.cff
