+++ CITATION.cff
@@ -7,8 +7,7 @@ title: "DocuScope Corpus Analysis & Concordancer"
 abstract: >-
   DocuScope CA is an open-source Streamlit application integrating DocuScope rhetorical
   tagging with part-of-speech and statistical corpus analysis for research and teaching.
-version: 0.4.1
-release-date: 2025-10-19
+version: 0.5.0
 doi: 10.5281/zenodo.17392153
 authors:
   - family-names: Brown
@@ -33,8 +32,7 @@ preferred-citation:
     - family-names: Brown
       given-names: David West
       orcid: https://orcid.org/0000-0001-7745-6354
-  version: 0.4.1
-  date-released: 2025-10-19
+  version: 0.5.0
   repository-code: https://github.com/browndw/docuscope-ca-online
   license: Apache-2.0
   doi: 10.5281/zenodo.17392153
