### 2026-08-05 712f7fd
Update interpretation.py
