### 2026-08-05 3dc87f6
Update docker-compose.yml
