+++ docker-compose.yml
@@ -4,23 +4,19 @@ services:
   postgres:
     image: postgres:16-alpine
     environment:
-      POSTGRES_DB: docuscope_ca
-      POSTGRES_USER: docuscope
-      POSTGRES_PASSWORD: docuscope
-    ports:
-      - "5432:5432"
+      POSTGRES_DB: ${POSTGRES_DB:-docuscope_ca}
+      POSTGRES_USER: ${POSTGRES_USER:-docuscope}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-docuscope}
     volumes:
       - postgres_data:/var/lib/postgresql/data
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U docuscope -d docuscope_ca"]
+      test: ["CMD-SHELL", "pg_isready -U $${POSTGRES_USER} -d $${POSTGRES_DB}"]
       interval: 10s
       timeout: 5s
       retries: 5
 
   redis:
     image: redis:7-alpine
-    ports:
-      - "6379:6379"
     healthcheck:
       test: ["CMD", "redis-cli", "ping"]
       interval: 10s
@@ -36,7 +32,7 @@ services:
         condition: service_healthy
     environment:
       PYTHONPATH: /app
-      DATABASE_URL: postgresql+psycopg://docuscope:docuscope@postgres:5432/docuscope_ca
+      DATABASE_URL: postgresql+psycopg://${POSTGRES_USER:-docuscope}:${POSTGRES_PASSWORD:-docuscope}@postgres:5432/${POSTGRES_DB:-docuscope_ca}
     command: alembic upgrade head
 
   streamlit_app:
@@ -52,13 +48,16 @@ services:
         condition: service_healthy
     environment:
       PYTHONPATH: /app
-      DATABASE_URL: postgresql+psycopg://docuscope:docuscope@postgres:5432/docuscope_ca
+      DATABASE_URL: postgresql+psycopg://${POSTGRES_USER:-docuscope}:${POSTGRES_PASSWORD:-docuscope}@postgres:5432/${POSTGRES_DB:-docuscope_ca}
       REDIS_URL: redis://redis:6379/0
       DOCUSCOPE_RQ_ENABLED: "1"
       DOCUSCOPE_RQ_SIMPLE_WORKER: "1"
+      DOCUSCOPE_RQ_PLOTBOT_QUEUE: ${DOCUSCOPE_RQ_PLOTBOT_QUEUE:-docuscope-plotbot}
       DOCUSCOPE_AI_PROVIDER: ${DOCUSCOPE_AI_PROVIDER:-}
       DOCUSCOPE_AI_BASE_URL: ${DOCUSCOPE_AI_BASE_URL:-}
       DOCUSCOPE_AI_MODEL: ${DOCUSCOPE_AI_MODEL:-}
+      DOCUSCOPE_AI_API_KEY: ${DOCUSCOPE_AI_API_KEY:-}
+      DOCUSCOPE_AI_TIMEOUT_SECONDS: ${DOCUSCOPE_AI_TIMEOUT_SECONDS:-60}
     ports:
       - "8501:8501"
     volumes:
@@ -92,15 +91,45 @@ services:
         condition: service_healthy
     environment:
       PYTHONPATH: /app
-      DATABASE_URL: postgresql+psycopg://docuscope:docuscope@postgres:5432/docuscope_ca
+      DATABASE_URL: postgresql+psycopg://${POSTGRES_USER:-docuscope}:${POSTGRES_PASSWORD:-docuscope}@postgres:5432/${POSTGRES_DB:-docuscope_ca}
[…]
