### 2026-08-05 a6d10b6
Create docker-compose.model.yml
