### 2026-08-05 ffed4df
Create MANIFEST.in
