### 2026-08-05 6ac2bdc
Delete qwen-bootstrap.sh
