### 2026-08-05 5d77344
Delete qwen-check.sh
