### 2026-08-05 e48aacb
Delete qwen-up.sh
