### 2026-08-05 91ae82e
Update test_plotbot_provider.py
