### 2026-08-05 38c1069
Update test_rq_smoke.py
