### 2026-08-05 c64d533
Update CITATION.cff
