+++ .github/workflows/release.yml
@@ -50,13 +50,16 @@ jobs:
     - name: Generate changelog
       id: changelog
       run: |
+        RELEASE_VERSION="${{ steps.get_version.outputs.version }}"
+        CHANGELOG_VERSION="${RELEASE_VERSION#v}"
         if [ -f CHANGELOG.md ]; then
           # Extract changelog for this version
-          sed -n "/## \[${GITHUB_REF#refs/tags/v}\]/,/## \[/p" CHANGELOG.md | head -n -1 > release_notes.md
-        else
-          echo "## Changes in ${{ steps.get_version.outputs.version }}" > release_notes.md
+          sed -n "/## \[${CHANGELOG_VERSION}\]/,/## \[/p" CHANGELOG.md | sed '$d' > release_notes.md
+        fi
+        if [ ! -s release_notes.md ]; then
+          echo "## Changes in ${RELEASE_VERSION}" > release_notes.md
           echo "" >> release_notes.md
-          echo "See commit history for detailed changes." >> release_notes.md
+          echo "See CHANGELOG.md and commit history for detailed changes." >> release_notes.md
         fi
     
     - name: Create source distribution
@@ -104,7 +107,7 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
     
     - name: Notify template users (if configured)
-      if: env.TEMPLATE_NOTIFICATION_ENABLED == 'true'
+      if: vars.TEMPLATE_NOTIFICATION_ENABLED == 'true'
       run: |
         echo "🎉 New release ${{ steps.get_version.outputs.version }} available!"
         echo "Template users with automatic sync will receive PRs with updates."
