+++ webapp/pages/6_compare_corpus_parts.py
@@ -13,6 +13,8 @@ import pandas as pd
 import polars as pl
 import streamlit as st
 
+from webapp.config.unified import config
+
 # Core application utilities with standardized patterns
 from webapp.utilities.core import app_core
 
@@ -25,6 +27,7 @@ from webapp.utilities.corpus import (
 )
 from webapp.utilities.ui import (
     keyness_settings_info, keyness_sort_controls,
+    extract_keyness_parts, extract_keyness_parts_settings,
     reference_parts, render_dataframe,
     sidebar_action_button, sidebar_help_link,
     sidebar_keyness_options, tag_filter_multiselect,
@@ -74,6 +77,7 @@ COMPARE_CORPUS_PARTS_PERSISTENT_WIDGETS = [
 app_core.register_page_widgets(COMPARE_CORPUS_PARTS_PERSISTENT_WIDGETS)
 
 KEYNESS_PARTS_QUEUE_STATE_KEY = "keyness_parts_queue_state"
+KEYNESS_PARTS_SWAP_WIDGET_PREFIX = "keyness_parts_swap"
 
 st.set_page_config(
     page_title=TITLE, page_icon=ICON,
@@ -108,7 +112,7 @@ def _clear_keyness_parts_queue_state(user_session_id: str) -> None:
 def _can_queue_keyness_parts(session: dict) -> str | None:
     """Return built-in target path when corpus-parts keyness can be queued."""
 
-    if not get_redis_queue_config().enabled:
+    if config.is_desktop_mode() or not get_redis_queue_config().enabled:
         return None
 
     target_source = safe_session_get(session, SessionKeys.TARGET_DB, "")
@@ -238,7 +242,10 @@ def _render_keyness_parts_queue_status(user_session_id: str) -> None:
         st.error(f"Queued corpus-parts generation failed: {job_row.failure_reason}")
         return
 
-    st.info("Generating corpus-parts keyness tables in the background...")
+    st.sidebar.status(
+        "Generating corpus-parts keyness tables in the background...",
+        state="running",
+    )
 
 
 def render_results_interface(user_session_id: str, session: dict) -> None:
@@ -254,14 +261,16 @@ def render_results_interface(user_session_id: str, session: dict) -> None:
     metadata_target = load_metadata(CorpusKeys.TARGET, user_session_id)
 
     # Display target and reference parts information
+    keyness_parts_data = extract_keyness_parts(metadata_target)
     col1, col2 = st.columns([1, 1])
     with col1:
-        st.info(target_parts(metadata_target.get(SessionKeys.KEYNESS_PARTS)[0]['temp']))
+        st.info(target_parts(keyness_parts_data))
     with col2:
-        st.info(reference_parts(metadata_target.get(SessionKeys.KEYNESS_PARTS)[0]['temp']))
+        st.info(reference_parts(keyness_parts_data))
 
     # Show user selections
-    st.info(keyness_settings_info(user_session_id))
+    pval_threshold, swap_target = extract_keyness_parts_settings(keyness_parts_data)
+    st.info(keyness_settings_info(user_session_id, pval_threshold, swap_target))
 
     # Table type selection in sidebar
     st.sidebar.markdown("### Comparison")
@@ -276,12 +285,12 @@ def render_results_interface(user_session_id: str, session: dict) -> None:
     st.sidebar.markdown("---")
 
     if table_radio == 'Tokens':
-        render_tokens_interface(user_session_id, target_manager)
+        render_tokens_interface(user_session_id, target_manager, session)
     else:
-        render_tags_interface(user_session_id, target_manager)
+        render_tags_interface(user_session_id, target_manager, session)
 
 
-def render_tokens_interface(user_session_id: str, target_manager) -> None:
[…]
