### 2026-08-06 cfccdc7
Update data_manager.py
