### 2026-08-06 45d1575
Update session_core.py
