+++ webapp/utilities/ui/__init__.py
@@ -13,6 +13,8 @@ from webapp.utilities.ui.corpus_display import (
     variance_info,
     contribution_info,
     group_info,
+    extract_keyness_parts,
+    extract_keyness_parts_settings,
     target_parts,
     reference_parts,
     load_and_display_target_corpus,
@@ -85,6 +87,8 @@ __all__ = [
     'variance_info',
     'contribution_info',
     'group_info',
+    'extract_keyness_parts',
+    'extract_keyness_parts_settings',
     'target_parts',
     'reference_parts',
     'load_and_display_target_corpus',
