### 2026-08-06 fc768ee
Update corpus_display.py
