+++ pyproject.toml
@@ -31,6 +31,8 @@ dependencies = [
     "pandas==2.2.3",
     "plotly==6.0.1",
     "polars==1.42.1",
+    "pyarrow==14.0.2; python_version < '3.13'",
+    "pyarrow==18.1.0; python_version >= '3.13'",
     "psutil==7.0.0",
     "SQLAlchemy==2.0.41",
     "alembic==1.16.1",
