+++ requirements.txt
@@ -9,6 +9,8 @@ numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.14"
 pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.14"
 plotly==6.0.1 ; python_version >= "3.11" and python_version < "3.14"
 polars==1.42.1 ; python_version >= "3.11" and python_version < "3.14"
+pyarrow==14.0.2 ; python_version >= "3.11" and python_version < "3.13"
+pyarrow==18.1.0 ; python_version >= "3.13" and python_version < "3.14"
 psutil==7.0.0 ; python_version >= "3.11" and python_version < "3.14"
 SQLAlchemy==2.0.41 ; python_version >= "3.11" and python_version < "3.14"
 alembic==1.16.1 ; python_version >= "3.11" and python_version < "3.14"
