### 2026-08-06 38c2a4c
Added Playwright navigation testing
