### 2026-08-06 b04688f
Update menu.py
