+++ webapp/menu.py
@@ -13,7 +13,6 @@ Both timeouts are configurable via options.toml and include warning systems.
 import base64
 import os
 import time
-from time import perf_counter
 
 import streamlit as st
 
@@ -39,7 +38,7 @@ CORE_PAGE_LINKS = [
     ("pages/3_tag_frequencies.py", "Tag Frequencies", ":material/table_view:"),
     ("pages/4_ngrams.py", "Ngrams & Clusters", ":material/table_view:"),
     ("pages/5_compare_corpora.py", "Compare Corpora", ":material/compare_arrows:"),
-    ("pages/6_compare_corpus_parts.py", "Compare Corpus Parts", ":material/compare_arrows:"),
+    ("pages/6_compare_corpus_parts.py", "Compare Corpus Parts", ":material/compare_arrows:"),  # noqa: E501
     ("pages/7_collocations.py", "Collocations", ":material/network_node:"),
     ("pages/8_kwic.py", "Key Words in Context", ":material/network_node:"),
     ("pages/9_advanced_plotting.py", "Advanced Plotting", ":material/line_axis:"),
@@ -68,19 +67,6 @@ def navigation_experiment_enabled() -> bool:
     """Legacy compatibility shim for removed navigation experiment."""
 
     return False
-
-
-def _log_slow_menu_step(session_id: str, step: str, start_time: float) -> None:
-    """Log slow menu phases during page entry reruns."""
-
-    elapsed_ms = (perf_counter() - start_time) * 1000
-    if elapsed_ms >= SLOW_MENU_STEP_MS:
-        logger.warning(
-            "Slow menu step session={} step={} duration_ms={:.2f}",
-            session_id,
-            step,
-            elapsed_ms,
-        )
 
 
 def _render_page_links(page_links: list[tuple[str, str, str]]) -> None:
@@ -265,37 +251,29 @@ def authenticated_menu(user_session_id: str | None = None):
 
     # Show log out button only if not DESKTOP and user is logged in
     if not DESKTOP and not TEST_MODE and hasattr(st, "user") and getattr(st.user, "is_logged_in", False):  # noqa: E501
-        logout_start = perf_counter()
         st.sidebar.button("Log out of Google", on_click=st.logout, icon=":material/logout:")
-        _log_slow_menu_step("test-or-desktop", "logout_button", logout_start)
 
     if minimal_menu:
-        core_links_start = perf_counter()
         with st.sidebar:
             _render_page_links(MINIMAL_PAGE_LINKS)
-        _log_slow_menu_step("test-or-desktop", "navigation_links", core_links_start)
         return
 
-    nav_expander_start = perf_counter()
     with st.sidebar.expander("**Navigation**",
                              icon=":material/explore:",
                              expanded=False):
-        core_links_start = perf_counter()
         _render_page_links(CORE_PAGE_LINKS)
-        _log_slow_menu_step("test-or-desktop", "navigation_links", core_links_start)
 
         # Admin-only features (only show in online mode with authorization)
-        admin_links_start = perf_counter()
         if _admin_features_enabled(user_session_id):
 
             st.markdown("---")
             st.markdown("**Admin Features**")
             _render_page_links(ADMIN_PAGE_LINKS)
-            _log_slow_menu_step("test-or-desktop", "admin_links", admin_links_start)
-            _log_slow_menu_step("test-or-desktop", "navigation_expander", nav_expander_start)  # noqa: E501
 
 
-def get_navigation_page_sections(user_session_id: str | None = None) -> dict[str, list[tuple[str, str, str]]]:
+def get_navigation_page_sections(
+        user_session_id: str | None = None
+) -> dict[str, list[tuple[str, str, str]]]:
     """Return page-link definitions for the navigation experiment."""
 
     page_sections = {"Navigation": CORE_PAGE_LINKS}
[…]
