### 2026-08-06 7149d03
Update docker-compose.model.yml
