### 2026-08-06 f5e5681
Update test_frequency_cache.py
