+++ requirements.txt
@@ -5,7 +5,8 @@ lingua-language-detector==2.1.1 ; python_version >= "3.11" and python_version < 
 loguru==0.7.3 ; python_version >= "3.11" and python_version < "3.14"
 kaleido==0.2.1 ; python_version >= "3.11" and python_version < "3.14"
 openai==1.66.5 ; python_version >= "3.11" and python_version < "3.14"
-numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.14"
+numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.13"
+numpy==2.1.3 ; python_version >= "3.13" and python_version < "3.14"
 pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.14"
 plotly==6.0.1 ; python_version >= "3.11" and python_version < "3.14"
 polars==1.42.1 ; python_version >= "3.11" and python_version < "3.14"
