+++ pyproject.toml
@@ -26,7 +26,8 @@ dependencies = [
     "lingua-language-detector==2.1.1",
     "loguru==0.7.3",
     "kaleido==0.2.1",
-    "numpy==1.26.4",
+    "numpy==1.26.4; python_version < '3.13'",
+    "numpy==2.1.3; python_version >= '3.13'",
     "openai==1.66.5",
     "pandas==2.2.3",
     "plotly==6.0.1",
