+++ ENTERPRISE_DEPLOYMENT.md
@@ -1,0 +1,366 @@
+# Enterprise and Horizontal Deployment
+
+This guide is for engineers deploying DocuScope CA at institutional or
+classroom scale, beyond a single Docker host. If you just want to try the
+application locally, see the [README](README.md#docker-deployment-recommended)
+instead — none of this is required for local or desktop use.
+
+It covers:
+
+- the **Horizontal Deployment Contract** for running multiple Streamlit
+  replicas behind a load balancer
+- **network exposure** expectations for local and production hosts
+- the optional **authorization bootstrap** for deployments that require login
+- the optional **self-hosted Qwen3 Coder model** for AI-assisted Plotbot
+- measured **concurrency and capacity** data from load testing
+
+## Horizontal Deployment Contract
+
+Use this sequence for a horizontally scaled institutional deployment:
+
+1. **Deploy one self-hosted Qwen service.** Use Docker Model Runner and the
+  pinned model in `docker-compose.model.yml`. Keep its OpenAI-compatible
+  endpoint on the deployment's private network. One model instance is the
+  starting point; it is not replicated with each app container.
+2. **Configure the model connection once.** Set `DOCUSCOPE_AI_PROVIDER=local`,
+  `DOCUSCOPE_AI_BASE_URL`, and `DOCUSCOPE_AI_MODEL` wherever the orchestrator
+  supplies environment values to the Streamlit-app and Plotbot-worker
+  containers (a `.env` file for Compose, or the equivalent secret/config
+  mechanism elsewhere — see [Optional Local Qwen3 Coder
+  Model](#optional-local-qwen3-coder-model) below for the exact variables).
+  Every replica then inherits the same settings.
+3. **Deploy multiple Streamlit replicas behind a load balancer.** The load
+  balancer must support WebSockets and session affinity. Every app replica must
+  use the same PostgreSQL service, Redis service, and shared artifact
+  filesystem.
+4. **Scale the worker pools separately.** Add `rq_worker` replicas for
+  deterministic analysis load and `rq_plotbot_worker` replicas for Plotbot
+  queue load. All Plotbot workers send requests to the same Qwen endpoint.
+
+```mermaid
+flowchart LR
+   U["Students"] --> LB["Load balancer<br/>WebSockets + session affinity"]
+   LB --> A1["Streamlit replica 1"]
+   LB --> A2["Streamlit replica 2+"]
+   A1 --> PG[(PostgreSQL)]
+   A2 --> PG
+   A1 --> R[(Redis)]
+   A2 --> R
+   A1 --> FS[(Shared artifacts)]
+   A2 --> FS
+   R --> DW["Analysis worker pool"]
+   R --> PW["Plotbot worker pool"]
+   DW --> PG
+   DW --> FS
+  PW --> PG
+  PW -->|"Built-in corpus requests"| Q["One self-hosted<br/>Qwen endpoint"]
+  A1 -->|"Uploaded/mixed requests"| Q
+  A2 -->|"Uploaded/mixed requests"| Q
+```
+
+The production orchestrator may translate the Compose services into its native
+service definitions; the checked-in Compose files remain the single-host
+reference. No application-specific discovery mechanism is required. Preserve
+these service invariants:
+
+| Role | Replicas | Required connections or storage |
+| --- | --- | --- |
+| `migrate` | One job per release | PostgreSQL; must finish before app and workers start |
+| `streamlit_app` | Two or more | PostgreSQL, Redis, and the shared artifact filesystem |
+| `rq_worker` | One or more | PostgreSQL, Redis, and the same shared artifact filesystem |
+| `rq_plotbot_worker` | One or more when Plotbot is enabled | PostgreSQL, Redis, and the shared model endpoint |
[…]
