+++ README.md
@@ -1,9 +1,6 @@
 # DocuScope Corpus Analysis & Concordancer
 
-<div class="image" align="center">
-    <img width="150" height="auto" src="webapp/_static/docuscope-logo.png" alt="DocuScope logo">
-    <br>
-</div>
+![DocuScope logo](webapp/_static/docuscope-logo.png)
 
 ---
 
@@ -55,6 +52,59 @@ DocuScope CA uniquely combines:
 - **Flexible Deployment**: Desktop, web, container, and hosted options
 - **Research-Ready**: Built for both exploratory discovery and hypothesis-driven studies
 
+## Quick Start
+
+Pick the path that fits:
+
+- **Try it now, no install**: [DocuScope CA Enterprise](https://docuscope-ca.eberly.cmu.edu/) — hosted, browser only.
+- **Docker (recommended)**:
+
+  ```bash
+  git clone https://github.com/browndw/docuscope-ca-online.git
+  cd docuscope-ca-online
+  docker compose up
+  ```
+
+  Open `http://localhost:8501`. See [Docker Deployment](#docker-deployment-recommended)
+  for the full service list and rebuild instructions.
+- **Local Python (no Docker)**:
+
+  ```bash
+  git clone https://github.com/browndw/docuscope-ca-online.git
+  cd docuscope-ca-online
+  python3.13 -m venv venv && source venv/bin/activate
+  pip install -r requirements.txt
+  streamlit run webapp/index.py
+  ```
+
+  See [Local Installation](#local-installation) for supported Python versions
+  and the automatic Desktop Mode fallback.
+
+Deploying for a classroom or institution beyond one host? See
+[Enterprise and Horizontal Deployment](#enterprise-and-horizontal-deployment).
+
+## Table of Contents
+
+- [Quick Start](#quick-start)
+- [Installation and Usage](#installation-and-usage)
+  - [Live Web Application](#live-web-application-immediate-access)
+  - [Docker Deployment](#docker-deployment-recommended)
+  - [Local Installation](#local-installation)
+  - [Desktop Application](#desktop-application)
+  - [Automated Testing](#automated-testing)
+  - [Reproducible Example Workflow](#reproducible-example-workflow)
+- [Enterprise and Horizontal Deployment](#enterprise-and-horizontal-deployment)
+- [Features](#features)
+- [Configuration](#configuration)
+- [Usage Examples](#usage-examples)
+- [Headless API & CLI](#headless-api--cli)
+- [Citation](#citation)
+- [License](#license)
+- [Using as Template](#using-as-template)
+- [Contributing](#contributing)
+- [Acknowledgments](#acknowledgments)
+- [Support and Documentation](#support-and-documentation)
+
 ## Installation and Usage
 
 DocuScope CA offers multiple deployment options to accommodate different user preferences and technical requirements.
@@ -77,10 +127,80 @@ git clone https://github.com/browndw/docuscope-ca-online.git
 cd docuscope-ca-online
 
 # Launch the application
-docker-compose up
+docker compose up
 ```
 
-The application will be available at `http://localhost:8501`. Docker automatically handles all dependencies, Python environment setup, and package installations.
+This starts the full enterprise-mode stack defined in `docker-compose.yml`:
+
+- `postgres` — the SQLAlchemy-backed control plane (artifact/job registry, authorization, runtime config)
+- `redis` — backs the RQ job queue
+- `migrate` — applies pending Alembic control-plane migrations before application processes start
+- `cleanup` — prunes expired sessions, temporary artifacts, terminal jobs, and old audit rows on an hourly schedule
+- `streamlit_app` — the web application, available at `http://localhost:8501`
[…]
