### 2026-08-18 #10418 @samhforbes
Looks good to me! Congrats @browndw on the new paper!
Many thanks to our reviewers @alfredo-hs, @PetrKorab, @SamHames for taking the time and of course @logological for keeping everything running smoothly - appreciate all you do.
